### 2024-10-24 #7408 @editorialbot
[PRE REVIEW]: XCALibre.jl: an XPU unstructured finite volume Computational Fluid Dynamics library
**Submitting author:** @mberto79 (<a href="http://orcid.org/0000-0001-5691-9292">Humberto Medina</a>)
**Repository:** https://github.com/mberto79/XCALibre.jl
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.3.1
