+++ ext/XCALibre_AMDExt.jl
@@ -34,4 +34,4 @@ end
 _m(A::SparseGPU) = A.dims[1]
 _n(A::SparseGPU) = A.dims[2]
 
-end
+end # end module
