### 2024-10-28 a1eed60
reintroduct -1 on spindex function to correct iteration range
