+++ src/ModelPhysics/Turbulence/RANS_kOmega.jl
@@ -175,29 +175,29 @@ function turbulence!(
     @. mueffk.values = rhof.values * (nuf.values + coeffs.σk*nutf.values)
 
     # Solve omega equation
-    prev .= omega.values
+    # prev .= omega.values
     discretise!(ω_eqn, omega, config)
     apply_boundary_conditions!(ω_eqn, omega.BCs, nothing, time, config)
     constrain_equation!(ω_eqn, omega.BCs, model, config) # active with WFs only
     # implicit_relaxation!(ω_eqn, omega.values, solvers.omega.relax, nothing, config)
-    # implicit_relaxation_diagdom!(ω_eqn, omega.values, solvers.omega.relax, nothing, config)
+    implicit_relaxation_diagdom!(ω_eqn, omega.values, solvers.omega.relax, nothing, config)
     update_preconditioner!(ω_eqn.preconditioner, mesh, config)
     solve_system!(ω_eqn, solvers.omega, omega, nothing, config)
     
     constrain_boundary!(omega, omega.BCs, model, config) # active with WFs only
     bound!(omega, config)
-    explicit_relaxation!(omega, prev, solvers.omega.relax, config)
+    # explicit_relaxation!(omega, prev, solvers.omega.relax, config)
 
     # Solve k equation
-    prev .= k.values
+    # prev .= k.values
     discretise!(k_eqn, k, config)
     apply_boundary_conditions!(k_eqn, k.BCs, nothing, time, config)
     # implicit_relaxation!(k_eqn, k.values, solvers.k.relax, nothing, config)
-    # implicit_relaxation_diagdom!(k_eqn, k.values, solvers.k.relax, nothing, config)
+    implicit_relaxation_diagdom!(k_eqn, k.values, solvers.k.relax, nothing, config)
     update_preconditioner!(k_eqn.preconditioner, mesh, config)
     solve_system!(k_eqn, solvers.k, k, nothing, config)
     bound!(k, config)
-    explicit_relaxation!(k, prev, solvers.k.relax, config)
+    # explicit_relaxation!(k, prev, solvers.k.relax, config)
 
     @. nut.values = k.values/omega.values
 
