### 2024-10-28 f58a686
add nothing as return to update_preconditioner function for nividia precons
