### 2024-10-28 3976d4a
clean solve! function
