### 2024-10-28 8089938
simplify code for _remove_pressure_source! kernel
