### 2024-10-28 c88d55d
comment out gradient calculation - should be moved inside `turbulence!` function
