### 2024-10-28 e37dfc1
add temporary path to 3D BFS mesh
