+++ examples/3D_BFS.jl
@@ -7,11 +7,13 @@ grid = "bfs_unv_tet_4mm.unv"
 grid = "bfs_unv_tet_5mm.unv"
 grid = "bfs_unv_tet_10mm.unv"
 
-mesh_file = joinpath(grids_dir, grid)
+mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
+
+# mesh_file = joinpath(grids_dir, grid)
 
 @time mesh = UNV3D_mesh(mesh_file, scale=0.001)
 
-@time mesh = FOAM3D_mesh(mesh_file, scale=0.001, integer_type=Int64, float_type=Float64)
+# @time mesh = FOAM3D_mesh(mesh_file, scale=0.001, integer_type=Int64, float_type=Float64)
 
 mesh_dev = adapt(CUDABackend(), mesh)
 
@@ -34,6 +36,7 @@ model = Physics(
     # Dirichlet(:sides, [0.0, 0.0, 0.0])
     Dirichlet(:inlet, velocity),
     Wall(:wall, [0.0, 0.0, 0.0]),
+    # Dirichlet(:wall, [0.0, 0.0, 0.0]),
     Neumann(:outlet, 0.0),
     Neumann(:top, 0.0),
     Neumann(:sides, 0.0)
@@ -48,7 +51,7 @@ model = Physics(
 )
 
 schemes = (
-    U = set_schemes(divergence=Upwind, gradient=Midpoint),
+    U = set_schemes(divergence=Upwind, gradient=Orthogonal),
     p = set_schemes(gradient=Midpoint)
     # p = set_schemes()
 )
@@ -58,25 +61,26 @@ solvers = (
     U = set_solver(
         model.momentum.U;
         solver      = BicgstabSolver, #CgSolver, # BicgstabSolver, GmresSolver, #CgSolver
-        preconditioner = Jacobi(),
+        preconditioner = Jacobi(), # Jacobi ILU0GPU
         convergence = 1e-7,
         relax       = 0.8,
-        rtol = 5e-1,
-        atol = 1e-10
+        rtol = 1e-2,
+        # atol = 1e-5
     ),
     p = set_solver(
         model.momentum.p;
         solver      = CgSolver, #GmresSolver, #CgSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
+        preconditioner = Jacobi(), # Jacobi IC0GPU
         convergence = 1e-7,
         relax       = 0.2,
-        rtol = 2.5e-1,
-        atol = 1e-10
+        rtol = 1e-1,
+        # atol = 1e-6
+        itmax = 100
     )
 )
 
 runtime = set_runtime(
-    iterations=500, time_step=1, write_interval=500)
+    iterations=100, time_step=1, write_interval=100)
 
 hardware = set_hardware(backend=CUDABackend(), workgroup=32)
 # hardware = set_hardware(backend=CPU(), workgroup=4)
@@ -89,7 +93,8 @@ GC.gc(true)
 initialise!(model.momentum.U, velocity)
 initialise!(model.momentum.p, 0.0)
 
-residuals = run!(model, config)
+# residuals = run!(model, config, limit_gradient=true)
+residuals = run!(model, config, limit_gradient=false)
 
 plot(; xlims=(0,1000))
 plot!(1:length(Rx), Rx, yscale=:log10, label="Ux")
