### 2024-10-28 9155c41
draft reimplementation of jacobi smoother loop using CSR structure for speed
