+++ src/Solve/Smoothers/Smoothers_jacobi.jl
@@ -29,11 +29,16 @@ end
 
     # for i ∈ 1:loops
         sum = zero(_get_float(mesh))
-        for fi ∈ faces_range
-            nID = cell_neighbours[fi]
-            nIndex = spindex(colptr, rowval, cID, nID)
-            sum += nzval[nIndex]*x[nID]
+        # for fi ∈ faces_range
+        #     nID = cell_neighbours[fi]
+        #     nIndex = spindex(colptr, rowval, cID, nID)
+        #     sum += nzval[nIndex]*x[nID]
+        # end
+        for nzvali ∈ colptr[cID]:(colptr[cID+1] - 1)
+            j = rowval[nzvali]
+            sum += nzval[nzvali]*x[j]
         end
+        sum -= nzval[cIndex]*x[cID] # remove multiplication with diagonal (faster than "if")
         rD = one(_get_int(mesh))/nzval[cIndex]
         x[cID] = rD*(b[cID] - sum)
     # end
