### 2024-10-28 d2184c5
improved default values for solvers and added warning message when convergence not reached in given max number of iters
