### 2024-10-28 41191ee
temp changes to examples to work locally
