### 2024-10-29 #7408 @mberto79
Hi @AnjaliSandip,

Thank you for agreeing to act as the editor for our submission.

I looked at the list and I have a list of potential reviewers (with a short justification):

[justinsgray](https://reviewers.joss.theoj.org/reviewers/1613) - background in aerodynamics could be valuable to comment on application side.

[Vincent Le Chenadec](https://reviewers.joss.theoj.org/reviewers/2897) - background in numerical analysis, an in particular  linear algebra, could be valuable to comment on approaches we used to solve the linear systems resulting from the VFM discretisation

[@mancellin](https://reviewers.joss.theoj.org/reviewers/551) - background in Finite Volume Method could be valuable since this is the underpinning method used in XCALibre.jl

[@simone-silvestri](https://reviewers.joss.theoj.org/reviewers/3283) - background in CUDA could be valuable to comment on the approach used to write for multiple compute backends

Kind regards,
Humberto
