### 2024-10-31 49dce3d
set up for multithreaded CSR
