+++ src/Solve/Solve.jl
@@ -2,10 +2,8 @@ module Solve
 
 using Adapt
 using Krylov
-using LinearAlgebra
 using Statistics
 using SparseArrays
-using SparseMatricesCSR
 using LinearOperators
 using LDLFactorizations
 using KernelAbstractions
@@ -13,9 +11,9 @@ using Atomix
 using ILUZero
 import KrylovPreconditioners as KP
 
-# using ThreadedSparseCSR
-# ThreadedSparseCSR.multithread_matmul(BaseThreads())
-# # ThreadedSparseCSR.multithread_matmul(PolyesterThreads())
+using LinearAlgebra
+using SparseMatricesCSR
+using ThreadedSparseCSR
 
 
 using XCALibre.Mesh
