+++ src/Simulate/Simulate.jl
@@ -6,6 +6,9 @@ using XCALibre.ModelFramework
 using XCALibre.Discretise
 using Adapt
 using KernelAbstractions
+
+using LinearAlgebra
+using SparseMatricesCSR
 using ThreadedSparseCSR
 
 include("Simulate_0_types.jl")
