### 2024-11-04 5bf2cf1
update sparse matrix access functions for CSR format and update AMDExt as well
