+++ src/Solve/Preconditioners/preconditioners_2_functions.jl
@@ -8,8 +8,6 @@ begin
     mesh = phi.mesh
     TF = _get_float(mesh)
     time = zero(TF) # assumes simulation starts at time = 0 (might need generalising)
-    # discretise!(
-        # eqn, ConstantScalar(zero(_get_int(mesh))), config) # should this be float?
 
     if typeof(phi) <: AbstractVectorField
 
@@ -21,7 +19,6 @@ begin
 
     elseif typeof(phi) <: AbstractScalarField
 
-        # discretise!(eqn, ConstantScalar(zero(_get_int(mesh))), config) # should this be float?
         discretise!(eqn, get_phi(eqn), config) # should this be float?
         apply_boundary_conditions!(eqn, BCs, nothing, time, config)
     end
@@ -36,12 +33,12 @@ end
 #     ) where {M<:AbstractSparseArray,PT,S} =
 # begin
 #     A = P.A
-#     # (; colptr, m, n, nzval, rowval) = A
-#     rowval, colptr, nzval, m ,n = sparse_array_deconstructor_preconditioners(A)
+#     # (; rowptr, m, n, nzval, colval) = A
+#     colval, rowptr, nzval, m ,n = sparse_array_deconstructor_preconditioners(A)
 #     storage = P.storage
 #     @inbounds for i ∈ 1:m
-#         idx_start = colptr[i]
-#         idx_next = colptr[i+1]
+#         idx_start = rowptr[i]
+#         idx_next = rowptr[i+1]
 #         column_vals = @view nzval[idx_start:(idx_next-1)] 
 #         storage[i] = 1/norm(column_vals)    
 #     end
@@ -64,10 +61,8 @@ function update_preconditioner!(P::Preconditioner{NormDiagonal,M,PT,S}, mesh, co
     (; backend, workgroup) = hardware
     
     A = P.A
-    # (; colptr, m, n, nzval, rowval) = A
-    # rowval, colptr, nzval, m ,n = sparse_array_deconstructor_preconditioners(A)
     nzval_array = _nzval(A)
-    colptr_array = _colptr(A)
+    colptr_array = _rowptr(A)
     m_array = _m(A)
 
     storage = P.storage
@@ -77,12 +72,12 @@ function update_preconditioner!(P::Preconditioner{NormDiagonal,M,PT,S}, mesh, co
     KernelAbstractions.synchronize(backend)
 end
 
-@kernel function update_NormDiagonal!(colptr, nzval, storage)
+@kernel function update_NormDiagonal!(rowptr, nzval, storage)
     i = @index(Global)
 
     # @inbounds begin
-        idx_start = colptr[i]
-        idx_next = colptr[i+1]
+        idx_start = rowptr[i]
+        idx_next = rowptr[i+1]
         column_vals = @view nzval[idx_start:(idx_next-1)] 
         norm = norm_static(column_vals)
         storage[i] = 1/norm
@@ -92,15 +87,15 @@ end
 # update_preconditioner!(P::Preconditioner{Jacobi,M,PT,S}, mesh) where {M<:AbstractSparseArray,PT,S} =
 # begin
 #     A = P.A
-#     # (; colptr, m, n, nzval, rowval) = A
-#     rowval, colptr, nzval, m, n = sparse_array_deconstructor_preconditioners(A)
+#     # (; rowptr, m, n, nzval, colval) = A
+#     colval, rowptr, nzval, m, n = sparse_array_deconstructor_preconditioners(A)
 #     storage = P.storage
 #     idx_diagonal = zero(eltype(m)) # index to diagonal element
 #     @inbounds for i ∈ 1:m
-#         idx_start = colptr[i]
-#         idx_next = colptr[i+1]
[…]
