+++ ext/XCALibre_CUDAExt.jl
@@ -15,27 +15,19 @@ function XCALibre.Mesh._convert_array!(arr, backend::CUDABackend)
     return adapt(CuArray, arr) # using CuArray
 end
 
-import XCALibre.ModelFramework: _nzval, _colptr, _rowval, get_sparse_fields, 
+import XCALibre.ModelFramework: _nzval, _rowptr, _colval, get_sparse_fields, 
                                 _build_A, _build_opA
 
 _build_A(backend::CUDABackend, i, j, v, n) = begin
-    # idev = adapt(CuArray, i)
-    # jdev = adapt(CuArray, j)
-    # vdev = adapt(CuArray, v)
-    # SparseGPU(idev, jdev, vdev, (n, n))
+
     A = sparse(i, j, v, n, n)
     SparseGPU(A)
 end
-_build_opA(A::SparseGPU) = KP.KrylovOperator(A)
 
+_build_opA(A::SparseGPU) = KP.KrylovOperator(A)
 @inline _nzval(A::SparseGPU) = A.nzVal
-# @inline _colptr(A::SparseGPU) = A.colPtr
-# @inline _rowval(A::SparseGPU) = A.rowVal
-@inline _colptr(A::SparseGPU) = A.rowPtr
-@inline _rowval(A::SparseGPU) = A.colVal
-# @inline get_sparse_fields(A::SparseGPU) = begin
-#     A.nzVal, A.rowVal, A.colPtr
-# end
+@inline _rowptr(A::SparseGPU) = A.rowPtr
+@inline _colval(A::SparseGPU) = A.colVal
 @inline get_sparse_fields(A::SparseGPU) = begin
     A.nzVal, A.colVal, A.rowPtr
 end
