+++ src/ModelFramework/ModelFramework_2_access_functions.jl
@@ -1,7 +1,7 @@
 export get_phi, get_flux
 export get_source, get_source_sign
 export _A, _A0, _b
-export _nzval, _rowval, _colptr
+export _nzval, _colval, _rowptr
 export get_sparse_fields
 export XDir, YDir, ZDir, get_values
 export get_backend
@@ -78,27 +78,21 @@ end
     eqn.equation.bz
 end
 
-# const GPUCSC = Union{
-#     CUDA.CUSPARSE.CuSparseMatrixCSC,
-#     AMDGPU.rocSPARSE.ROCSparseMatrixCSC}
 
-# @inline _nzval(A::GPUCSC) = A.nzVal
 @inline _nzval(A::SparseArrays.SparseMatrixCSC) = A.nzval
 
-# @inline _colptr(A::GPUCSC) = A.colPtr
-@inline _colptr(A::SparseArrays.SparseMatrixCSC) = A.colptr
+@inline _rowptr(A::SparseArrays.SparseMatrixCSC) = A.rowptr
 
-# @inline _rowval(A::GPUCSC) = A.rowVal
-@inline _rowval(A::SparseArrays.SparseMatrixCSC) = A.rowval
+@inline _colval(A::SparseArrays.SparseMatrixCSC) = A.colval
 
 @inline get_sparse_fields(A::SparseArrays.SparseMatrixCSC) = begin
-    A.nzval, A.rowval, A.colptr
+    A.nzval, A.colval, A.rowptr
 end
 
 # Sparse CSR (temporary hack)
 @inline _nzval(A::SparseMatricesCSR.SparseMatrixCSR) = A.nzval
-@inline _colptr(A::SparseMatricesCSR.SparseMatrixCSR) = A.rowptr 
-@inline _rowval(A::SparseMatricesCSR.SparseMatrixCSR) = A.colval
+@inline _rowptr(A::SparseMatricesCSR.SparseMatrixCSR) = A.rowptr 
+@inline _colval(A::SparseMatricesCSR.SparseMatrixCSR) = A.colval
 
 @inline get_sparse_fields(A::SparseMatricesCSR.SparseMatrixCSR) = begin
     A.nzval, A.colval, A.rowptr
