### 2024-11-04 2df1d5e
temp changes to example 3D_BFS
