+++ examples/3D_BFS.jl
@@ -1,4 +1,3 @@
-using Plots
 using XCALibre
 using CUDA
 
@@ -8,19 +7,23 @@ grid = "bfs_unv_tet_5mm.unv"
 grid = "bfs_unv_tet_10mm.unv"
 
 mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
-mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
+# mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
 
 # mesh_file = joinpath(grids_dir, grid)
 
-@time mesh = UNV3D_mesh(mesh_file, scale=0.001)
+# mesh_file = "bfs_unv_tet_5mm.unv"
+mesh = UNV3D_mesh(mesh_file, scale=0.001)
 
-# @time mesh = FOAM3D_mesh(mesh_file, scale=0.001, integer_type=Int64, float_type=Float64)
+workgroup = 32
+backend = CUDABackend()
 
-mesh_dev = adapt(CUDABackend(), mesh)
+mesh_dev = adapt(backend, mesh)
 
+# Inlet conditions
 velocity = [0.5, 0.0, 0.0]
+noSlip = [0.0, 0.0, 0.0]
 nu = 1e-3
-Re = velocity[1]*0.1/nu
+Re = (0.2*velocity[1])/nu
 
 model = Physics(
     time = Steady(),
@@ -29,74 +32,70 @@ model = Physics(
     energy = Energy{Isothermal}(),
     domain = mesh_dev
     )
-    
 
-@assign! model momentum U (
-    # Dirichlet(:inlet, velocity),
-    # Neumann(:outlet, 0.0),0.0]),
-    # Dirichlet(:sides, [0.0, 0.0, 0.0])
+@assign! model momentum U ( 
     Dirichlet(:inlet, velocity),
-    Wall(:wall, [0.0, 0.0, 0.0]),
-    # Dirichlet(:wall, [0.0, 0.0, 0.0]),
     Neumann(:outlet, 0.0),
-    Neumann(:top, 0.0),
-    Neumann(:sides, 0.0)
+    Wall(:wall, noSlip),
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
 )
 
 @assign! model momentum p (
     Neumann(:inlet, 0.0),
     Dirichlet(:outlet, 0.0),
     Neumann(:wall, 0.0),
-    Neumann(:top, 0.0),
-    Neumann(:sides, 0.0)
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
 )
-
-schemes = (
-    U = set_schemes(divergence=Upwind, gradient=Orthogonal),
-    p = set_schemes(gradient=Orthogonal)
-    # p = set_schemes()
-)
-
 
 solvers = (
     U = set_solver(
         model.momentum.U;
-        solver      = BicgstabSolver, #CgSolver, # BicgstabSolver, GmresSolver, #CgSolver
-        preconditioner = Jacobi(), # Jacobi ILU0GPU
+        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+        preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.8,
-        rtol = 0.1,
+        rtol = 0.1
     ),
     p = set_solver(
         model.momentum.p;
-        solver      = CgSolver, #GmresSolver, #CgSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), # Jacobi IC0GPU
[…]
