+++ test/test_mesh_conversion.jl
@@ -6,7 +6,7 @@ mesh = UNV2D_mesh(meshFile, scale=0.001)
 msg = IOBuffer(); println(msg, mesh)
 outputTest = String(take!(msg))
 
-@test outputTest == "2D Mesh\n-> 3484 cells\n-> 5306 faces\n-> 1823 nodes\n\nBoundaries \n-> inlet (faces: 1:40)\n-> outlet (faces: 41:80)\n-> bottom (faces: 81:120)\n-> top (faces: 121:160)\n\n"
+@test outputTest == "2D Mesh\n-> 3484 cells\n-> 5306 faces\n-> 1823 nodes\n\n"
 
 # test quad mesh
 meshFile = joinpath(test_grids_dir, "quad40.unv")
@@ -14,7 +14,7 @@ mesh = UNV2D_mesh(meshFile, scale=0.001)
 msg = IOBuffer(); println(msg, mesh)
 outputTest = String(take!(msg))
 
-@test outputTest == "2D Mesh\n-> 1600 cells\n-> 3280 faces\n-> 1681 nodes\n\nBoundaries \n-> inlet (faces: 1:40)\n-> outlet (faces: 41:80)\n-> bottom (faces: 81:120)\n-> top (faces: 121:160)\n\n"
+@test outputTest == "2D Mesh\n-> 1600 cells\n-> 3280 faces\n-> 1681 nodes\n\n"
 
 # 3D UNV cavity mesh
 meshFile = joinpath(test_grids_dir, "OF_cavity_hex", "cavity_hex.unv")
@@ -22,7 +22,7 @@ mesh = UNV3D_mesh(meshFile, scale=0.001)
 msg = IOBuffer(); println(msg, mesh)
 outputTest_UNV3D = String(take!(msg))
 
-@test outputTest_UNV3D == "3D Mesh\n-> 125 cells\n-> 450 faces\n-> 216 nodes\n\nBoundaries \n-> walls (faces: 1:125)\n-> top (faces: 126:150)\n\n"
+@test outputTest_UNV3D == "3D Mesh\n-> 125 cells\n-> 450 faces\n-> 216 nodes\n\n"
 
 # 3D FOAM cavity mesh
 meshFile = joinpath(test_grids_dir, "OF_cavity_hex", "polyMesh")
@@ -30,7 +30,7 @@ mesh = FOAM3D_mesh(meshFile, scale=0.001)
 msg = IOBuffer(); println(msg, mesh)
 outputTest_FOAM3D = String(take!(msg))
 
-@test outputTest_FOAM3D == "3D Mesh\n-> 125 cells\n-> 450 faces\n-> 216 nodes\n\nBoundaries \n-> walls (faces: 1:125)\n-> top (faces: 126:150)\n\n"
+@test outputTest_FOAM3D == "3D Mesh\n-> 125 cells\n-> 450 faces\n-> 216 nodes\n\n"
 
 # Test 3D UNV and FOAM meshes are equal
 @test outputTest_UNV3D == outputTest_FOAM3D
