### 2024-11-05 3803286
weighted jacobi smoother implemented
