+++ src/Solve/Solve_1_api.jl
@@ -13,7 +13,7 @@ export residual!
         convergence, 
         relax,
         limit=(),
-        itmax::Integer=500, 
+        itmax::Integer=1000, 
         atol=(eps(_get_float(field.mesh)))^0.9,
         rtol=_get_float(field.mesh)(1e-1)
         ) where {S,PT<:PreconditionerType} = begin
@@ -42,7 +42,7 @@ This function is used to provide solver settings that will be used internally in
 - `convergence` sets the stopping criteria of this field
 - `relax` specifies the relaxation factor to be used e.g. set to 1 for no relaxation
 - `limit` used in some solvers to bound the solution within this limits e.g. (min, max). It defaults to `()`
-- `itmax` maximum number of iterations in a single solver pass (defaults to 500) 
+- `itmax` maximum number of iterations in a single solver pass (defaults to 1000) 
 - `atol` absolute tolerance for the solver (default to eps(FloatType)^0.9)
 - `rtol` set relative tolerance for the solver (defaults to 1e-1)
 """
@@ -54,7 +54,7 @@ set_solver(
         convergence, 
         relax,
         limit=(),
-        itmax::Integer=500, 
+        itmax::Integer=1000, 
         atol=(eps(_get_float(field.mesh)))^0.9,
         rtol=_get_float(field.mesh)(1e-1)
     ) where {S,PT<:PreconditionerType} = begin
@@ -183,26 +183,23 @@ function solve_system!(phiEqn::ModelEquation, setup, result, component, config) 
     opA = phiEqn.equation.opA
     b = _b(phiEqn, component)
 
-    # smoother = JacobiSmoother(0, 2)
-    # apply_smoother!(smoother, values, A, b, mesh, config)
-    # x .= values
-    # apply_smoother!(smoother, x, A, b, mesh, config)
-
+    smoother = JacobiSmoother(10,2/3,adapt(backend, zeros(length(values))))
+    apply_smoother!(smoother, values, A, b, mesh, config)
      
     ldiv = typeof(P) <: KP.AbstractKrylovPreconditioner
     solve!(
-        solver, opA, b, values; M=P, itmax=itmax, atol=atol, rtol=rtol, ldiv=ldiv # original
+        solver, opA, b, values; M=P, itmax=itmax, atol=atol, rtol=rtol, ldiv=ldiv
             )
-    # KernelAbstractions.synchronize(backend)
+    KernelAbstractions.synchronize(backend)
 
     statistics(solver).niter == itmax && @warn "Maximum number of iteration reached!"
     
-    kernel! = solve_copy_kernel!(backend, workgroup)
+    kernel! = _copy!(backend, workgroup)
     kernel!(values, x, ndrange = length(values))
     KernelAbstractions.synchronize(backend)
 end
 
-@kernel function solve_copy_kernel!(a, b)
+@kernel function _copy!(a, b)
     i = @index(Global)
 
     @inbounds begin
