+++ src/Solve/Smoothers/Smoothers_jacobi.jl
@@ -1,10 +1,11 @@
-struct JacobiSmoother{D,L}
-    diagonal::D 
+struct JacobiSmoother{L,F,V}
     loops::L
+    omega::F
+    x_temp::V
 end
 Adapt.@adapt_structure JacobiSmoother
 
-function apply_smoother!(smoother::JacobiSmoother, x, A, b, mesh, config)
+function apply_smoother!(smoother, x, A, b, mesh, config)
     (; hardware, runtime) = config
     (; backend, workgroup) = hardware
 
@@ -12,30 +13,69 @@ function apply_smoother!(smoother::JacobiSmoother, x, A, b, mesh, config)
     colval = _colval(A)
     rowptr = _rowptr(A)
 
-    kernel! = _apply_smoother!(backend, workgroup)
+    _smoother_launch(backend, smoother, x, b, nzval, colval, rowptr, mesh, workgroup)
+end
+
+# CPU implementation
+function _smoother_launch(
+    backend::CPU, smoother, x, b, nzval, colval, rowptr, mesh, workgroup
+    )
+    _apply_smoother!(smoother, x, b, nzval, colval, rowptr, mesh)
+end
+
+function _apply_smoother!(
+    smoother::JacobiSmoother, x, b, nzval, colval, rowptr, mesh
+    )
+    ω = smoother.omega
     for _ ∈ 1:smoother.loops
-        kernel!(x, b, nzval, colval, rowptr, mesh; ndrange = length(mesh.cells))
-        KernelAbstractions.synchronize(backend)
+        Threads.@threads for cID ∈ eachindex(mesh.cells)
+            cIndex =  spindex(rowptr, colval, cID, cID)
+            xi = multiply_row(x, b, ω, nzval, rowptr, colval,  cID, cIndex)
+            smoother.x_temp[cID] = xi
+        end
+        Threads.@threads for cID ∈ eachindex(smoother.x_temp)
+            @inbounds x[cID] = smoother.x_temp[cID]
+        end
     end
 end
 
-@kernel function _apply_smoother!(x, b, nzval, colval, rowptr, mesh)
-    cID = @index(Global)
-    (; cells, cell_neighbours) = mesh
-
-    cell = cells[cID]
-    (; faces_range) = cell
-    cIndex = spindex(rowptr, colval, cID, cID)
-
-
-    sum = zero(_get_float(mesh))
-
-    for nzvali ∈ rowptr[cID]:(rowptr[cID+1] - 1)
-        j = colval[nzvali]
-        sum += nzval[nzvali]*x[j]
+# GPU Kernel
+function _smoother_launch(
+    backend::GPU, smoother, x, b, nzval, colval, rowptr, mesh, workgroup
+    )
+    krange = length(mesh.cells)
+    kernel! = _apply_smoother!(backend, workgroup)
+    for _ ∈ 1:smoother.loops
+        kernel!(smoother, x, b, nzval, colval, rowptr, mesh, ndrange = krange)
+        KernelAbstractions.synchronize(backend)
+        x .= smoother.x_temp
     end
-    sum -= nzval[cIndex]*x[cID] # remove multiplication with diagonal (faster than "if")
-    rD = one(_get_int(mesh))/nzval[cIndex]
-    x[cID] = rD*(b[cID] - sum)
+end
 
+@kernel function _apply_smoother!(
+    smoother::JacobiSmoother, x, b, nzval, colval, rowptr, mesh
+    )
+    cID = @index(Global)
+    ω = smoother.omega
+    cIndex =  spindex(rowptr, colval, cID, cID)
+    xi = multiply_row(x, b, ω, nzval, rowptr, colval,  cID, cIndex)
[…]
