### 2024-11-06 c60a35a
add JacobiSmoother constructors and documentation
