### 2024-11-07 9bb135d
remove activate_multithread func from solvers and reimplement for API-user level
