+++ src/Solvers/Solvers_1_SIMPLE.jl
@@ -116,8 +116,7 @@ function SIMPLE(
     # p_model = p_eqn.model
     (; solvers, schemes, runtime, hardware) = config
     (; iterations, write_interval) = runtime
-    (; backend, method) = hardware
-    # activate_multithread(backend, method)
+    (; backend) = hardware
     
     mdotf = get_flux(U_eqn, 2)
     nueff = get_flux(U_eqn, 3)
