+++ src/Simulate/Simulate_0_types.jl
@@ -35,7 +35,7 @@ Adapt.@adapt_structure Configuration
 
 
 """
-    hardware = set_hardware(backend, workgroup, method=BaseThreads())
+    hardware = set_hardware(backend, workgroup)
 
 Function used to configure the backend.
 
@@ -43,13 +43,11 @@ Function used to configure the backend.
 
 - `backend` named tuple used to specify the backend e.g. `CPU()`, `CUDABackend()` or other backends supported by `KernelAbstraction.jl`
 - `workgroup::Int` this is an integer specifying the number of workers that cooperate in a parallel run. For GPUs this could be set to the size of the device's warp e.g. `workgroup = 32`. On CPUs, the default value in `KernelAbstractions.jl` is currently `workgroup = 1024`.
-- `method` to be used by ThreadedSparseCSR for parallel multiplication of sparse matrices. `BaseThreads()` is given as default. Alternatively, it could be set to `PolyesterThreads()`
 
 # Output
 
 This function returns a `NamedTuple` with the fields `backend` and `workgroup` which are accessed by internally in `XCALibre.jl` to execute a given kernel.
 """
-set_hardware(;backend, workgroup, method=BaseThreads()) = begin
-    ThreadedSparseCSR.multithread_matmul(method)
-    (backend=backend, workgroup=workgroup, method=method)
+set_hardware(;backend, workgroup) = begin
+    (backend=backend, workgroup=workgroup)
 end
