### 2024-11-07 5d066a7
temporarily nuke support for ILU0, LDL and DILU preconditioners
