+++ test/0_TEST_CASES/2d_compressible_laminar_flatplate_fixedT.jl
@@ -86,7 +86,7 @@ solvers = (
 
 runtime = set_runtime(iterations=500, write_interval=500, time_step=1)
 
-hardware = set_hardware(backend=CPU(), workgroup=cld(length(mesh.faces), 4))
+hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
 # hardware = set_hardware(backend=ROCBackend(), workgroup=32)
 
