+++ test/0_TEST_CASES/2d_incompressible_flatplate_KOmega_HighRe.jl
@@ -75,28 +75,28 @@ solvers = (
     U = set_solver(
         model.momentum.U;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), #ILU0(),
+        preconditioner = Jacobi(), 
         convergence = 1e-7,
         relax       = 0.7,
     ),
     p = set_solver(
         model.momentum.p;
         solver      = CgSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), #ILU0(),
+        preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.3,
     ),
     k = set_solver(
         model.turbulence.k;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), #ILU0(),
+        preconditioner = Jacobi(), 
         convergence = 1e-7,
         relax       = 0.3,
     ),
     omega = set_solver(
         model.turbulence.omega;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), #ILU0(),
+        preconditioner = Jacobi(), 
         convergence = 1e-7,
         relax       = 0.3,
     )
@@ -104,7 +104,7 @@ solvers = (
 
 runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
 
-hardware = set_hardware(backend=CPU(), workgroup=32)
+hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
 # hardware = set_hardware(backend=ROCBackend(), workgroup=32)
 
@@ -123,4 +123,4 @@ residuals = run!(model, config)
 
 outlet = boundary_average(:outlet, model.momentum.U, config)
 
-@test Umag ≈ outlet[1] atol = 0.035
+@test Umag ≈ outlet[1] atol = 0.1*Umag
