### 2024-11-07 25e6e5a
remove references to noncompatible precons and add wraning to documentation
