### 2024-11-07 0c3e1a3
add tests for smoother
