+++ src/Solve/Smoothers/Smoothers_jacobi.jl
@@ -17,7 +17,7 @@ Structure to hold information for using the weighted Jacobi smoother.
 - `x_temp` is a vector used internally to store intermediate solutions.
 
 """
-struct JacobiSmoother{L,F,V} <: AbstractSmoother
+struct JacobiSmoother{L<:Integer,F<:Number,V<:AbstractArray} <: AbstractSmoother
     loops::L
     omega::F
     x_temp::V
@@ -47,30 +47,30 @@ JacobiSmoother(; domain, loops, omega=2/3) = begin
     JacobiSmoother(loops, F(omega), adapt(backend, x))
 end
 
-function apply_smoother!(smoother, x, A, b, mesh, config)
-    (; hardware, runtime) = config
+function apply_smoother!(smoother, x, A, b, config)
+    (; hardware) = config
     (; backend, workgroup) = hardware
 
     nzval = _nzval(A)
     colval = _colval(A)
     rowptr = _rowptr(A)
 
-    _smoother_launch(backend, smoother, x, b, nzval, colval, rowptr, mesh, workgroup)
+    _smoother_launch(backend, smoother, x, b, nzval, colval, rowptr, workgroup)
 end
 
 # CPU implementation
 function _smoother_launch(
-    backend::CPU, smoother, x, b, nzval, colval, rowptr, mesh, workgroup
+    backend::CPU, smoother, x, b, nzval, colval, rowptr, workgroup
     )
-    _apply_smoother!(smoother, x, b, nzval, colval, rowptr, mesh)
+    _apply_smoother!(smoother, x, b, nzval, colval, rowptr)
 end
 
 function _apply_smoother!(
-    smoother::JacobiSmoother, x, b, nzval, colval, rowptr, mesh
+    smoother::JacobiSmoother, x, b, nzval, colval, rowptr
     )
     ω = smoother.omega
     for _ ∈ 1:smoother.loops
-        Threads.@threads for cID ∈ eachindex(mesh.cells)
+        Threads.@threads for cID ∈ eachindex(x)
             cIndex =  spindex(rowptr, colval, cID, cID)
             xi = multiply_row(x, b, ω, nzval, rowptr, colval,  cID, cIndex)
             smoother.x_temp[cID] = xi
@@ -83,19 +83,19 @@ end
 
 # GPU Kernel
 function _smoother_launch(
-    backend::GPU, smoother, x, b, nzval, colval, rowptr, mesh, workgroup
+    backend::GPU, smoother, x, b, nzval, colval, rowptr, workgroup
     )
-    krange = length(mesh.cells)
+    krange = length(x)
     kernel! = _apply_smoother!(backend, workgroup)
     for _ ∈ 1:smoother.loops
-        kernel!(smoother, x, b, nzval, colval, rowptr, mesh, ndrange = krange)
+        kernel!(smoother, x, b, nzval, colval, rowptr, ndrange = krange)
         KernelAbstractions.synchronize(backend)
         x .= smoother.x_temp
     end
 end
 
 @kernel function _apply_smoother!(
-    smoother::JacobiSmoother, x, b, nzval, colval, rowptr, mesh
+    smoother::JacobiSmoother, x, b, nzval, colval, rowptr
     )
     cID = @index(Global)
     ω = smoother.omega
