### 2024-11-08 61a102b
fix tests using percentage of inlet velocity for tolerance
