+++ test/0_TEST_CASES/2d_incompressible_flatplate_KOmega_HighRe.jl
@@ -123,4 +123,4 @@ residuals = run!(model, config)
 
 outlet = boundary_average(:outlet, model.momentum.U, config)
 
-@test Umag ≈ outlet[1] atol = 0.035
+@test Umag ≈ outlet[1] atol =0.1*Umag
