### 2024-11-08 cc397ef
add compat entries for new packages
