+++ Project.toml
@@ -53,3 +53,5 @@ Reexport = "1.2"
 Setfield = "1.1"
 StaticArrays = "1.9"
 julia = "1.8"
+SparseMatricesCSR = "0.6"
+ThreadedSparseCSR = "0.1"
