### 2024-11-08 2711130
remove unused function `interger_type`
