### 2024-11-08 d0aaf2f
update release notes
