+++ README.md
@@ -1,3 +1,5 @@
+<meta name="google-site-verification" content="UZSnZbbvZqRUM_1_d5d9ox1IeO5z9iE8Oynt7mBjJaM" />
+
 [![][docs-stable-img]][docs-stable-url] [![][docs-dev-img]][docs-dev-url]
 
 [docs-stable-img]: https://img.shields.io/badge/docs-stable-blue.svg
