### 2024-11-08 9dec8bb
add CALibre to docs deps
