+++ versions.js
@@ -2,5 +2,5 @@ var DOC_VERSIONS = [
   "stable",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.3.1";
+var DOCUMENTER_NEWEST = "v0.3.2";
 var DOCUMENTER_STABLE = "stable";
