### 2024-11-09 3feccfa
CompatHelper: add new compat entry for CUDA in [weakdeps] at version 5, (keep existing compat)
