+++ Project.toml
@@ -36,6 +36,7 @@ XCALibre_CUDAExt = "CUDA"
 Accessors = "0.1.30"
 Adapt = "4.0"
 Atomix = "0.1"
+CUDA = "5"
 GPUArrays = "10.2"
 ILUZero = "0.2"
 IncompleteLU = "0.2"
