### 2024-11-09 8bdd17c
CompatHelper: bump compat for GPUArrays to 11, (keep existing compat)
