+++ Project.toml
@@ -36,7 +36,7 @@ XCALibre_CUDAExt = "CUDA"
 Accessors = "0.1.30"
 Adapt = "4.0"
 Atomix = "0.1"
-GPUArrays = "10.2"
+GPUArrays = "10.2, 11"
 ILUZero = "0.2"
 IncompleteLU = "0.2"
 KernelAbstractions = "0.9.25"
