### 2024-11-09 d9778ad
CompatHelper: add new compat entry for Statistics at version 1, (keep existing compat)
