+++ Project.toml
@@ -48,4 +48,5 @@ ProgressMeter = "1.10"
 Reexport = "1.2"
 Setfield = "1.1"
 StaticArrays = "1.9"
+Statistics = "1"
 julia = "1.8"
