### 2024-11-10 a37c6ce
simplify extracting sparse diagonal indices
