+++ src/Solve/Preconditioners/preconditioners_1_DILU.jl
@@ -28,15 +28,7 @@ function diagonal_indices!(Di, A::AbstractSparseArray{Tf,Ti}) where {Tf,Ti}
     (; rowptr, n, colval) = A
     idx_diagonal = zero(eltype(n)) # index to diagonal element
     @inbounds for i ∈ 1:n
-        idx_start = rowptr[i]
-        idx_next = rowptr[i+1]
-        @inbounds for p ∈ idx_start:(idx_next-1)
-            row = colval[p]
-            if row == i
-                idx_diagonal = p
-                break
-            end
-        end
+        idx_diagonal = spindex(rowptr, colval, i, i)
         Di[i] = idx_diagonal
     end
 end
