### 2024-11-12 dbf28be
draft implementation of smoothing step for scalar and vector equation models
