+++ src/Solve/Solve_1_api.jl
@@ -12,6 +12,7 @@ export residual!
         preconditioner::PT, 
         convergence, 
         relax,
+        smoother=nothing,
         limit=(),
         itmax::Integer=1000, 
         atol=(eps(_get_float(field.mesh)))^0.9,
@@ -25,7 +26,7 @@ export residual!
             preconditioner=preconditioner, 
             convergence=convergence |> TF, 
             relax=relax |> TF, 
-            smoother=nothing,
+            smoother=smoother,
             limit=limit,
             itmax=itmax, 
             atol=atol |> TF, 
@@ -69,6 +70,7 @@ set_solver(
         preconditioner=preconditioner, 
         convergence=convergence |> TF, 
         relax=relax |> TF, 
+        smoother=smoother,
         limit=limit,
         itmax=itmax, 
         atol=atol |> TF, 
@@ -85,7 +87,7 @@ end
         time_step::N
         ) where {I<:Integer,N<:Number} = begin
         
-        # returned `NamedTuple``
+        # returned `NamedTuple`
             (
             iterations=iterations, 
             dt=time_step, 
@@ -111,10 +113,36 @@ set_runtime(; iterations::I, write_interval::I, time_step::N) where {I<:Integer,
     (iterations=iterations, dt=time_step, write_interval=write_interval)
 end
 
+function smoother_solve!(
+    smoother::Nothing, eqn::ModelEquation{T,M,E,S,P}, phi, config;
+    time, ref, irelax) where {T<:ScalarModel,M,E,S,P}
+    nothing
+end
+
+function smoother_solve!(
+    smoother, eqn::ModelEquation{T,M,E,S,P}, phi, config; 
+    time=nothing, ref=nothing, irelax=nothing
+    ) where {T<:ScalarModel,M,E,S,P}
+
+    discretise!(eqn, phi, config)       
+    apply_boundary_conditions!(eqn, phi.BCs, nothing, time, config)
+    setReference!(eqn, ref, 1, config)
+    # if !isnothing(irelax)
+    #     implicit_relaxation!(eqn, phi.values, irelax, nothing, config)
+    #     # implicit_relaxation_diagdom!(eqn, phi.values, irelax, nothing, config)
+    # end
+    A = _A(eqn)
+    b = _b(eqn, nothing)
+    apply_smoother!(smoother, phi.values, A, b, config.hardware)
+end
+
 function solve_equation!(
     eqn::ModelEquation{T,M,E,S,P}, phi, solversetup, config; time=nothing, ref=nothing, irelax=nothing
     ) where {T<:ScalarModel,M,E,S,P}
 
+    smoother_solve!(
+        solversetup.smoother, eqn, phi, config; time=time, ref=ref, irelax=irelax
+    )
     discretise!(eqn, phi, config)       
     apply_boundary_conditions!(eqn, phi.BCs, nothing, time, config)
     setReference!(eqn, ref, 1, config)
@@ -126,11 +154,51 @@ function solve_equation!(
     solve_system!(eqn, solversetup, phi, nothing, config)
 end
 
+function smoother_solve!(
+    smoother::Nothing, psiEqn::ModelEquation{T,M,E,S,P}, psi, solversetup, xdir, ydir, zdir, config; time
+    ) where {T<:VectorModel,M,E,S,P}
+    nothing
+end
+
+function smoother_solve!(
+    smoother, psiEqn::ModelEquation{T,M,E,S,P}, psi, solversetup, xdir, ydir, zdir, config; time=nothing
+    ) where {T<:VectorModel,M,E,S,P}
+
+    mesh = psi.mesh
[…]
