+++ examples/3D_BFS.jl
@@ -1,21 +1,27 @@
+using Plots 
+using ThreadPinning
 using XCALibre
-using CUDA
+# using CUDA
+
+pinthreads(:cores)
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
-grid = "bfs_unv_tet_4mm.unv"
-grid = "bfs_unv_tet_5mm.unv"
+# grid = "bfs_unv_tet_4mm.unv"
+# grid = "bfs_unv_tet_5mm.unv"
 grid = "bfs_unv_tet_10mm.unv"
+mesh_file = joinpath(grids_dir, grid)
 
-mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
+# mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
 # mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
 
-# mesh_file = joinpath(grids_dir, grid)
 
 # mesh_file = "bfs_unv_tet_5mm.unv"
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
 
-workgroup = 32
-backend = CUDABackend()
+workgroup = cld(length(mesh.cells), Threads.nthreads())
+backend = CPU(); activate_multithread(backend)
+# workgroup = 32
+# backend = CUDABackend()
 
 mesh_dev = adapt(backend, mesh)
 
@@ -54,6 +60,7 @@ solvers = (
         model.momentum.U;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
         preconditioner = Jacobi(),
+        smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=0.5),
         convergence = 1e-7,
         relax       = 0.8,
         rtol = 0.1
@@ -62,6 +69,7 @@ solvers = (
         model.momentum.p;
         solver      = CgSolver, # BicgstabSolver, GmresSolver
         preconditioner = Jacobi(), #NormDiagonal(),
+        smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         convergence = 1e-7,
         relax       = 0.2,
         rtol = 0.1,
@@ -90,7 +98,7 @@ residuals = run!(model, config)
 
 # Now get timing information
 
-runtime = set_runtime(iterations=500, write_interval=500, time_step=1)
+runtime = set_runtime(iterations=2000, write_interval=500, time_step=1)
 config = Configuration(solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
 
 GC.gc(true)
@@ -98,4 +106,11 @@ GC.gc(true)
 initialise!(model.momentum.U, velocity)
 initialise!(model.momentum.p, 0.0)
 
-residuals = run!(model, config)
+@time residuals = run!(model, config)
+
+iterations = runtime.iterations
+plot(yscale=:log10, ylims=(1e-5,1e-1))
+plot!(1:iterations, residuals.Ux, label="Ux")
+plot!(1:iterations, residuals.Uy, label="Uy")
+plot!(1:iterations, residuals.Uz, label="Uz")
+plot!(1:iterations, residuals.p, label="p")

[…]
