### 2024-11-12 30c4f6a
working set up for smoother (needs cleaning code)
