+++ src/Solve/Smoothers/Smoothers_jacobi.jl
@@ -47,9 +47,9 @@ JacobiSmoother(; domain, loops, omega=2/3) = begin
     JacobiSmoother(loops, F(omega), adapt(backend, x))
 end
 
-apply_smoother(smoother::Nothing, x, A, b, hardware) = nothing
+apply_smoother!(smoother::Nothing, x, A, b, hardware) = nothing
 
-function apply_smoother!(smoother, x, A, b, hardware)
+function apply_smoother!(smoother::AbstractSmoother, x, A, b, hardware)
     (; backend, workgroup) = hardware
 
     nzval = _nzval(A)
