### 2024-11-12 a9ffd40
clean preconditioner and smoother implementations
