+++ src/Solve/Preconditioners/preconditioners_0_types.jl
@@ -1,32 +1,33 @@
 
-export PreconditionerType, Preconditioner
+export PreconditionerType
 export Jacobi, NormDiagonal #, ILU0 # , LDL
 export DILU, DILUprecon
 export IC0GPU, ILU0GPU
 
 abstract type PreconditionerType end
+abstract type LDIVPreconditioner <: PreconditionerType end
+abstract type MULPreconditioner <: PreconditionerType end
 
-struct NormDiagonal <: PreconditionerType end
+struct NormDiagonal <: MULPreconditioner end
 Adapt.@adapt_structure NormDiagonal
 
-struct Jacobi <: PreconditionerType end
+struct Jacobi <: MULPreconditioner end
 Adapt.@adapt_structure Jacobi
 
-# struct LDL <: PreconditionerType end
+# struct LDL <: MULPreconditioner end
 # Adapt.@adapt_structure LDL
 
-# struct ILU0 <: PreconditionerType end
+# struct ILU0 <: MULPreconditioner end
 # Adapt.@adapt_structure ILU0
 
-struct DILU <: PreconditionerType end
+struct DILU <: MULPreconditioner end
 Adapt.@adapt_structure DILU
 
-struct IC0GPU <: PreconditionerType end
+struct IC0GPU <: LDIVPreconditioner end
 Adapt.@adapt_structure IC0GPU
 
-struct ILU0GPU <: PreconditionerType end
+struct ILU0GPU <: LDIVPreconditioner end
 Adapt.@adapt_structure ILU0GPU
-
 
 struct Preconditioner{T,M,P,S}
     A::M
@@ -34,14 +35,14 @@ struct Preconditioner{T,M,P,S}
     storage::S
 end
 function Adapt.adapt_structure(to, itp::Preconditioner{T,M,Pr,S}) where {T,M,Pr,S}
-    # A = Adapt.adapt_structure(to, itp.A)
-    # P = Adapt.adapt_structure(to, itp.P)
-    # storage = Adapt.adapt_structure(to, itp.storage) 
     A = Adapt.adapt(to, itp.A)
     P = Adapt.adapt(to, itp.P)
     storage = Adapt.adapt(to, itp.storage) 
     Preconditioner{T,typeof(A),typeof(P),typeof(storage)}(A,P,storage)
 end
+
+is_ldiv(precon::Preconditioner{T,M,P,S}) where {T,M,P,S} = T <: LDIVPreconditioner
+
 Preconditioner{NormDiagonal}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     backend = get_backend(A)
     m, n = size(A)
