### 2024-11-12 e0928b2
update docs for preconditioner status
