### 2024-11-13 0a17344
update docs compat entry for documenter
