### 2024-11-13 e54309f
fix version number on release notes
