### 2024-11-13 756f816
add code to remove results from doctests
