+++ docs/makeLocal.jl
@@ -48,5 +48,5 @@ makedocs(
     ]
 )
 
-# foreach(rm, filter(endswith(".vtk"), readdir("docs", join=true)))
-# foreach(rm, filter(endswith(".vtu"), readdir("docs", join=true)))
+foreach(rm, filter(endswith(".vtk"), readdir("docs", join=true)))
+foreach(rm, filter(endswith(".vtu"), readdir("docs", join=true)))
