+++ src/Discretise/Discretise.jl
@@ -11,7 +11,7 @@ using GPUArrays
 
 using LinearAlgebra
 using SparseMatricesCSR
-using ThreadedSparseCSR
+# using ThreadedSparseCSR
 
 using XCALibre.Mesh
 using XCALibre.Fields
