+++ src/Solve/Solve.jl
@@ -13,7 +13,7 @@ import KrylovPreconditioners as KP
 
 using LinearAlgebra
 using SparseMatricesCSR
-using ThreadedSparseCSR
+# using ThreadedSparseCSR
 
 
 using XCALibre.Mesh
