+++ Project.toml
@@ -25,7 +25,6 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-ThreadedSparseCSR = "39591389-4884-462a-89e7-9edfce40b6eb"
 
 [weakdeps]
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
@@ -40,8 +39,8 @@ AMDGPU = "1"
 Accessors = "0.1.30"
 Adapt = "4.0"
 Atomix = "0.1"
-GPUArrays = "10.2, 11"
 CUDA = "5"
+GPUArrays = "10.2, 11"
 ILUZero = "0.2"
 IncompleteLU = "0.2"
 KernelAbstractions = "0.9.25"
@@ -53,8 +52,7 @@ PrecompileTools = "1.2"
 ProgressMeter = "1.10"
 Reexport = "1.2"
 Setfield = "1.1"
+SparseMatricesCSR = "0.6"
 StaticArrays = "1.9"
 Statistics = "1"
 julia = "1.8"
-SparseMatricesCSR = "0.6"
-ThreadedSparseCSR = "0.1"
