### 2024-11-17 c6d7139
fix `activate_multithred` function
