### 2024-11-17 5eefc7a
add note on origin of sparse matrix vector multiply code (`ThreadedSpareCSR`)
