### 2024-11-19 0069629
bring multithreaded spvm into a module
