+++ src/Solvers/Solvers.jl
@@ -15,8 +15,8 @@ using Adapt
 
 using LinearAlgebra
 using SparseMatricesCSR
-# using ThreadedSparseCSR
 
+using XCALibre.Multithread
 using XCALibre.Mesh
 using XCALibre.Fields
 using XCALibre.ModelFramework
