### 2024-11-21 da3939a
only run with Flux example and stop evaluation of optimisation example
