+++ docs/src/examples/05_2d-aerofoil-inflow-optimisation.md
@@ -20,7 +20,7 @@ add Plots, Distributions, LinearAlgebra, GaussianProcesses, BayesianOptimization
 
 This will download and install the required packages. Note that "CUDA" must also be added for GPU acceleration. Once installed, the packages can be loaded as follows:
 
-```@example optimisation
+```julia
 using Pkg; # hide
 installed = "Flux" ∈ keys(Pkg.project().dependencies) # hide
 installed && Pkg.rm("Flux", io=devnull) #hide
@@ -37,7 +37,7 @@ nothing # hide
 
 In order to calculate the lift-to-drag ratio, the pressure and viscous forces over the aerofoil must be calculated and summed. The lift and drag components can then be calculated (noting the rotated inflow vector to allow for different aerofoil angles of attack with the same mesh). A function returning the coefficients of lift and drag (not specifically needed for this example) is also presented for convenience.
 
-```@example optimisation
+```julia
 # Lift to drag ratio calculation
 lift_to_drag(patch::Symbol, model, ρ, nu, α) = begin
     Fp = pressure_force(patch, model.momentum.p, ρ)
@@ -67,7 +67,7 @@ end
 
 Next, the aerofoil mesh .unv file must be imported. It must also be adapted to work with the GPU, if desired.
 
-```@example optimisation
+```julia
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
 grid = "NACAMesh.unv"
 mesh_file = joinpath(grids_dir, grid)
@@ -81,7 +81,7 @@ mesh_dev = mesh # running on CPU
 
 The `BayesianOptimization.jl` package can optimise a Julia function that is passed to it. To interface with XCALibre.jl, the entire CFD simulation setup must simply be wrapped within a function, which can then be passed to the optimiser. The function must take the variable to be changed (angle of attack, in this case) as its input, and must return the desired output (lift-to-drag ratio). Therefore, the post-processing step to calculate lift-to-drag ratio is also wrapped in the same function.
 
-```@example optimisation
+```julia
 function foil_optim(α::Vector{Float64})
     println("\nSelected α value: $(α[1])")
 
@@ -237,7 +237,7 @@ Note that this code saves a single .vtk file of the last CFD iteration each time
 
 Finally, the Bayesian optimiser must be configured before the optimisation can be performed. This example follows the BayesianOptimization.jl default configuration for the Gaussian process surrogate model, limiting input dimensions to 1 (the angle of attack). The surrogate model is then set to be optimised every 10 iterations. The inputs are limited to between 0 and 15 degrees. The problem is configured as a maximisation problem, with an initial sample period of 10 iterations and 50 maximum allowed iterations. The final line of the following code block is then run to perform the optimisation.
 
-```@example optimisation
+```julia
 # Bayesian Optimisation (using BayesianOptimization.jl)
 isdir("vtk_results") || mkdir("vtk_results")
 
