+++ docs/Project.toml
@@ -4,6 +4,7 @@ Changelog = "5217a498-cd5d-4ec6-b8c2-9b85a09b6e3e"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 GaussianProcesses = "891a1506-143c-57d2-908e-e1f8e92e6de9"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
