+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.3",
   "dev",
 ];
 var DOCUMENTER_NEWEST = "v0.3.2";
