### 2024-11-22 896ce4c
add benchmarking documentation (framework)
