+++ docs/src/reference.md
@@ -8,6 +8,7 @@ Modules = [
     XCALibre.UNV3, 
     XCALibre.FoamMesh, 
     XCALibre.Multithread,
+    XCALibre.Multithread,
     XCALibre.Fields,
     XCALibre.ModelFramework,
     XCALibre.Discretise, 
