+++ src/Solve/Preconditioners/preconditioners_1_DILU.jl
@@ -6,11 +6,11 @@ export ldiv!
 # NOTE ADDED ON 2024/11/07 - THIS IS LIKELY BROKEN DUE TO CHANGE TO CSR FORMAT
 
 
-function extract_diagonal!(D, Di, A::AbstractSparseArray{Tf,Ti}, config) where {Tf,Ti}
+function extract_diagonal!(D, Di, A::SparseMatrixCSC{Tf,Ti}, config) where {Tf,Ti}
     (; hardware) = config
     (; backend, workgroup) = hardware
     
-    colval, rowptr, nzval, m ,n = sparse_array_deconstructor_preconditioners(A)
+    (; nzval, n) = A
 
     kernel! = extract_diagonal_kernel!(backend, workgroup)
     kernel!(D, Di, nzval, ndrange = n)
@@ -24,11 +24,11 @@ end
     end
 end
 
-function diagonal_indices!(Di, A::AbstractSparseArray{Tf,Ti}) where {Tf,Ti} 
-    (; rowptr, n, colval) = A
-    idx_diagonal = zero(eltype(n)) # index to diagonal element
+function diagonal_indices!(Di, A::SparseMatrixCSC{Tf,Ti}) where {Tf,Ti} 
+    (; colptr, n, rowval) = A
+    idx_diagonal = zero(Ti) # index to diagonal element
     @inbounds for i ∈ 1:n
-        idx_diagonal = spindex(rowptr, colval, i, i)
+        idx_diagonal = spindex_csc(colptr, rowval, i, i)
         Di[i] = idx_diagonal
     end
 end
@@ -37,8 +37,8 @@ end
 # integer_type(A::CUDA.CUSPARSE.CuSparseMatrixCSC{Tf,Ti}) where {Tf,Ti} = Ti
 
 function upper_row_indices(A, Di) # upper triangular row column indices
-    (; rowptr, n, colval) = A
-    TI = eltype(rowptr)
+    (; colptr, n, rowval) = A
+    TI = eltype(colptr)
     Ri = TI[] # column pointers on i-th row
     J = TI[] # column indices on i-th row
     upper_indices_IDs = UnitRange{TI}[]
@@ -48,10 +48,10 @@ function upper_row_indices(A, Di) # upper triangular row column indices
         upper_indices_start = length(Ri)
         offset = 0
         for j = (i+1):n
-            idx_start = rowptr[j]
+            idx_start = colptr[j]
             idx_next = Di[j] # access column down to diagonal only
             @inbounds for p ∈ idx_start:idx_next
-                row = colval[p]
+                row = rowval[p]
                 if row == i
                     push!(R_temp, p) # array of pointers
                     push!(J_temp, j) # column indices
@@ -73,17 +73,19 @@ function upper_row_indices(A, Di) # upper triangular row column indices
 end
 
 function update_dilu_diagonal!(P, mesh, config) # must rename
-    # (; A, storage) = P
-    # (; rowptr, m, n, nzval, colval) = A
-    # (; Di, D) = storage
-    # extract_diagonal!(D, Di, A) 
-    # for i ∈ 1:m
-    #     # for j ∈ 1:(i-1)
-    #     for j ∈ (i+1):m
-    #         # D[i] -= A[i,j]*A[j,i]/D[j]
-    #         D[j] -= A[i,j]*A[j,i]/D[i]
-    #     end
-    # end
+    (; A, storage) = P
+    (; m) = A
+    (; Di, D) = storage
+    extract_diagonal!(D, Di, A, config)
+    for i ∈ 1:m 
[…]
