+++ src/Solve/Preconditioners/preconditioners_1_DILU.jl
@@ -6,17 +6,17 @@ export ldiv!
 # NOTE ADDED ON 2024/11/07 - THIS IS LIKELY BROKEN DUE TO CHANGE TO CSR FORMAT
 
 
-function extract_diagonal!(D, Di, A::SparseMatrixCSC{Tf,Ti}, config) where {Tf,Ti}
+function extract_diagonal!(D, Di, A::AbstractSparseArray, config)
     (; hardware) = config
     (; backend, workgroup) = hardware
     
     (; nzval, n) = A
 
-    kernel! = extract_diagonal_kernel!(backend, workgroup)
+    kernel! = _extract_diagonal!(backend, workgroup)
     kernel!(D, Di, nzval, ndrange = n)
 end
 
-@kernel function extract_diagonal_kernel!(D, Di, nzval)
+@kernel function _extract_diagonal!(D, Di, nzval)
     i = @index(Global)
     
     @inbounds begin
@@ -29,6 +29,15 @@ function diagonal_indices!(Di, A::SparseMatrixCSC{Tf,Ti}) where {Tf,Ti}
     idx_diagonal = zero(Ti) # index to diagonal element
     @inbounds for i ∈ 1:n
         idx_diagonal = spindex_csc(colptr, rowval, i, i)
+        Di[i] = idx_diagonal
+    end
+end
+
+function diagonal_indices!(Di, A::SparseMatrixCSR{N, Tf,Ti}) where {N, Tf,Ti} 
+    (; rowptr, n, colval) = A
+    idx_diagonal = zero(Ti) # index to diagonal element
+    @inbounds for i ∈ 1:n
+        idx_diagonal = spindex(rowptr, colval, i, i)
         Di[i] = idx_diagonal
     end
 end
@@ -73,58 +82,34 @@ function upper_row_indices(A, Di) # upper triangular row column indices
 end
 
 function update_dilu_diagonal!(P, mesh, config) # must rename
-    (; A, storage) = P
-    (; m) = A
-    (; Di, D) = storage
-    extract_diagonal!(D, Di, A, config)
-    for i ∈ 1:m 
-        D[i] = A[i,i]
-    end
+    # # Algo 1
+    # (; A, storage) = P
+    # (; m) = A
+    # (; D) = storage
+    # for i ∈ 1:m 
+    #     D[i] = A[i,i]
+    # end
 
-    for i ∈ 1:m
-        for j ∈ (i+1):m
-            D[j] = D[j] - A[i,j]*A[j,i]/D[i]
-        end
-    end
+    # for i ∈ 1:m
+    #     for j ∈ (i+1):m
+    #         D[j] = D[j] - A[i,j]*A[j,i]/D[i]
+    #     end
+    # end
 
     # Algo 2
-    # (; A, storage) = P
-    # (; rowptr, m, n, nzval, colval) = A
-    # (; Di, D) = storage
-    # extract_diagonal!(D, Di, A) 
-    # sum = 0.0
-    # for i ∈ 2:m
-    #     for j ∈ 1:(i-1)
-    #         sum += A[i,j]*A[j,i]/D[j]
-    #     end
-    #     D[i] -= sum
-    #     sum = 0.0
-    # end
-    
-    # # Algo 3
-    # # backend = _get_backend(mesh)
-
-    # (; A, storage) = P
-    # (; colptr, n, nzval, rowval) = A
-    # (; Di, Ri, D, upper_indices_IDs) = storage
-    
-    # extract_diagonal!(D, Di, A, config)
-
-    # for i ∈ 1:n
-    #     # D[i] = nzval[Di[i]]
-    #     upper_index_ID = upper_indices_IDs[i] 
-    #     c_start = Di[i] + 1
-    #     c_end = colptr[i+1] - 1
[…]
