+++ src/Solve/Preconditioners/preconditioners_0_types.jl
@@ -20,7 +20,7 @@ Adapt.@adapt_structure Jacobi
 # struct ILU0 <: MULPreconditioner end
 # Adapt.@adapt_structure ILU0
 
-struct DILU <: MULPreconditioner end
+struct DILU <: LDIVPreconditioner end
 Adapt.@adapt_structure DILU
 
 struct IC0GPU <: LDIVPreconditioner end
@@ -112,6 +112,7 @@ struct DILUprecon{M,V,VI,VUR}
     upper_indices_IDs::VUR
 end
 Adapt.@adapt_structure DILUprecon
+
 Preconditioner{DILU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     m, n = size(A)
     m == n || throw("Matrix not square")
@@ -123,5 +124,19 @@ Preconditioner{DILU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     P  = LinearOperator(
         F, m, n, false, false, (y, v) -> ldiv!(y, S, v)
         )
+    Preconditioner{DILU,typeof(A),typeof(P),typeof(S)}(A,P,S)
+end
+
+Preconditioner{DILU}(A::SparseMatrixCSR{N, F,I}) where {N, F,I} = begin
+    m, n = size(A)
+    m == n || throw("Matrix not square")
+    D = zeros(F, m)
+    Di = zeros(I, m)
+    diagonal_indices!(Di, A)
+    S = DILUprecon(A, D, Di, Di, Di, nothing) # temp constructor
+    # P  = LinearOperator(
+    #     F, m, n, false, false, (y, v) -> ldiv!(y, S, v)
+    #     )
+    P = S
     Preconditioner{DILU,typeof(A),typeof(P),typeof(S)}(A,P,S)
 end
