+++ src/Solve/Preconditioners/preconditioners_1_DILU.jl
@@ -2,10 +2,6 @@ import LinearAlgebra.ldiv!, LinearAlgebra.\
 
 export ldiv!
 
-# THIS WHOLE IMPLEMENTATION NEEDS TO BE CLEANED UP AND CHECKED FOR CORRECTNESS
-# NOTE ADDED ON 2024/11/07 - THIS IS LIKELY BROKEN DUE TO CHANGE TO CSR FORMAT
-
-
 function extract_diagonal!(D, Di, A::AbstractSparseArray, config)
     (; hardware) = config
     (; backend, workgroup) = hardware
@@ -42,50 +38,7 @@ function diagonal_indices!(Di, A::SparseMatrixCSR{N, Tf,Ti}) where {N, Tf,Ti}
     end
 end
 
-# integer_type(A::SparseMatrixCSC{Tf,Ti}) where {Tf,Ti} = Ti
-# integer_type(A::CUDA.CUSPARSE.CuSparseMatrixCSC{Tf,Ti}) where {Tf,Ti} = Ti
-
-function upper_row_indices(A, Di) # upper triangular row column indices
-    (; colptr, n, rowval) = A
-    TI = eltype(colptr)
-    Ri = TI[] # column pointers on i-th row
-    J = TI[] # column indices on i-th row
-    upper_indices_IDs = UnitRange{TI}[]
-    @inbounds for i ∈ 1:n
-        R_temp = TI[]
-        J_temp = TI[]
-        upper_indices_start = length(Ri)
-        offset = 0
-        for j = (i+1):n
-            idx_start = colptr[j]
-            idx_next = Di[j] # access column down to diagonal only
-            @inbounds for p ∈ idx_start:idx_next
-                row = rowval[p]
-                if row == i
-                    push!(R_temp, p) # array of pointers
-                    push!(J_temp, j) # column indices
-                    offset += 1
-                end
-            end
-        end
-        upper_indices_end = upper_indices_start + offset
-        push!(Ri, R_temp...)
-        push!(J, J_temp...)
-        if upper_indices_end == upper_indices_start
-            IDs_range = UnitRange{TI}(upper_indices_start:upper_indices_end)
-        else
-            IDs_range = UnitRange{TI}(upper_indices_start+1:upper_indices_end)
-        end
-        push!(upper_indices_IDs, IDs_range)
-    end
-    return Ri, J, upper_indices_IDs
-end
-
 function update_dilu_diagonal!(P, mesh, config) # must rename
-    # # Algo 1
-    # (; A, storage) = P
-    # (; m) = A
-    # (; D) = storage
     # for i ∈ 1:m 
     #     D[i] = A[i,i]
     # end
@@ -96,157 +49,68 @@ function update_dilu_diagonal!(P, mesh, config) # must rename
     #     end
     # end
 
-    # Algo 2
     (; A, storage) = P 
     (; rowptr, colval, nzval, m, n) = A
     (; D, Di) = storage
     extract_diagonal!(D, Di, A, config)
     @inbounds for i ∈ 1:m
-        @inbounds for j ∈ (i+1):m
-            nzIndex1 = spindex(rowptr, colval, i, j)
-            nzIndex2 = spindex(rowptr, colval, j, i)
-            if nzIndex1 !== 0 && nzIndex2 !== 0
-                D[j] = D[j] - nzval[nzIndex1]*nzval[nzIndex2]/D[i]
+        start_index = rowptr[i]
+        end_index = rowptr[i+1] - 1
+        @inbounds for nzi ∈ start_index:end_index # j ∈ (i+1):m
+            j = colval[nzi]
[…]
