### 2024-11-25 3af48fd
docs: update warning to remove reference to DILU
