### 2024-11-26 ebb2371
make lacation operator "negative"
