### 2024-11-29 76c79a0
reactivate gradient calclation
