### 2024-11-29 d75c0ed
moving `constrain_equation` in KOmega model to correct location (must be just before updating precon)
