### 2024-11-30 6d1b5ae
use simpler approach to see cell value in linear system for `constrain_equation`
