+++ examples/2D_flatplate_lowRe.jl
@@ -25,8 +25,6 @@ model = Physics(
     domain = mesh_dev
     )
 
-println(typeof(model.fluid))
-
 @assign! model momentum U (
     Dirichlet(:inlet, velocity),
     Neumann(:outlet, 0.0),
@@ -108,7 +106,7 @@ solvers = (
 runtime = set_runtime(iterations=1000, write_interval=100, time_step=1)
 
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
-hardware = set_hardware(backend=CPU(), workgroup=4)
+hardware = set_hardware(backend=CPU(), workgroup=1024)
 
 config = Configuration(
     solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
