### 2024-11-30 0d5d8df
reactivate full form of time discretisation for KOmega model
