### 2024-11-30 8b7420c
update HighRe case to match OpenFOAM's case
