### 2024-12-01 dcc5de5
fix `_constrain` implementation
