+++ src/ModelPhysics/Turbulence/RANS_functions.jl
@@ -151,12 +151,12 @@ end
         # Atomix.@atomic b[cID] += nzval[nzIndex]*ωc
         # Atomix.@atomic nzval[nzIndex] += nzval[nzIndex] 
 
-        z = zero(eltype[nzval])
+        z = zero(eltype(nzval))
         for nzi ∈ rowptr[cID]:(rowptr[cID+1] - 1)
             nzval[nzi] = z
         end
         cIndex = spindex(rowptr, colval, cID, cID)
-        nzval[cIndex] = one(eltype[nzval])
+        nzval[cIndex] = one(eltype(nzval))
         b[cID] = ωc
     end
 end
