### 2024-12-02 54feddb
A hacky implementation of hybrid `DILU` that works on CUDA
