+++ examples/2D_BFS.jl
@@ -1,5 +1,5 @@
 using XCALibre
-# using CUDA
+using CUDA
 
 # backwardFacingStep_2mm, backwardFacingStep_10mm
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
@@ -8,8 +8,10 @@ mesh_file = joinpath(grids_dir, grid)
 
 mesh = UNV2D_mesh(mesh_file, scale=0.001)
 
-# mesh_dev = adapt(CUDABackend(), mesh)
-mesh_dev = mesh
+# backend = CPU(); activate_multithread(backend)
+# mesh_dev = mesh
+backend = CUDABackend()
+mesh_dev = adapt(backend, mesh)
 
 velocity = [0.5, 0.0, 0.0]
 nu = 1e-3
@@ -52,7 +54,7 @@ solvers = (
     U = set_solver(
         model.momentum.U;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = DILU(), # ILU0GPU, Jacobi, DILU
+        preconditioner = Jacobi(), # ILU0GPU, Jacobi, DILU
         # smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         convergence = 1e-7,
         relax       = 0.6,
@@ -61,7 +63,7 @@ solvers = (
     p = set_solver(
         model.momentum.p;
         solver      = CgSolver, # BicgstabSolver, GmresSolver, CgSolver
-        preconditioner = Jacobi(), # IC0GPU, Jacobi, DILU
+        preconditioner = DILU(), # IC0GPU, Jacobi, DILU
         # smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         convergence = 1e-7,
         relax       = 0.4,
@@ -73,8 +75,8 @@ runtime = set_runtime(
     iterations=2000, time_step=1, write_interval=100)
     # iterations=1, time_step=1, write_interval=1)
 
-# hardware = set_hardware(backend=CUDABackend(), workgroup=32)
-hardware = set_hardware(backend=CPU(), workgroup=32)
+hardware = set_hardware(backend=CUDABackend(), workgroup=32)
+# hardware = set_hardware(backend=backend, workgroup=1024)
 
 config = Configuration(
     solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
