### 2024-12-02 361f981
update changelog to include PR references
