### 2024-12-02 899c03b
draft implementation
