### 2024-12-03 c313ff8
improve KOmega model implementation
