+++ examples/2D_flatplate_highRe.jl
@@ -1,4 +1,4 @@
-using Plots
+# using Plots
 using XCALibre
 # using CUDA
 
@@ -8,10 +8,11 @@ grid = "flatplate_2D_highRe.unv"
 mesh_file = joinpath(grids_dir, grid)
 mesh = UNV2D_mesh(mesh_file, scale=0.001)
 
+backend = CPU(); activate_multithread(backend)
 mesh_dev = mesh
 # mesh_dev = adapt(CUDABackend(), mesh)
 
-velocity = [15, 0.0, 0.0]
+velocity = [10, 0.0, 0.0]
 nu = 1e-5
 Re = velocity[1]*1/nu
 k_inlet = 0.375
@@ -29,6 +30,7 @@ model = Physics(
     Dirichlet(:inlet, velocity),
     Neumann(:outlet, 0.0),
     Wall(:wall, [0.0, 0.0, 0.0]),
+    # Dirichlet(:wall, [0.0, 0.0, 0.0]),
     Neumann(:top, 0.0)
 )
 
@@ -43,6 +45,7 @@ model = Physics(
     Dirichlet(:inlet, k_inlet),
     Neumann(:outlet, 0.0),
     KWallFunction(:wall),
+    # Neumann(:wall, 0.0),
     Neumann(:top, 0.0)
 )
 
@@ -99,10 +102,10 @@ solvers = (
     )
 )
 
-runtime = set_runtime(iterations=1000, write_interval=100, time_step=1)
+runtime = set_runtime(iterations=2000, write_interval=100, time_step=1)
 
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
-hardware = set_hardware(backend=CPU(), workgroup=1024)
+hardware = set_hardware(backend=backend, workgroup=1024)
 
 config = Configuration(
     solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
@@ -117,27 +120,27 @@ initialise!(model.turbulence.nut, k_inlet/ω_inlet)
 
 residuals = run!(model, config) # 9.39k allocs
 
-using DelimitedFiles
-using LinearAlgebra
+# using DelimitedFiles
+# using LinearAlgebra
 
-OF_data = readdlm("flatplate_OF_wall_kOmega_highRe.csv", ',', Float64, skipstart=1)
-oRex = OF_data[:,7].*velocity[1]./nu[1]
-oCf = sqrt.(OF_data[:,12].^2 + OF_data[:,13].^2)/(0.5*velocity[1]^2)
+# OF_data = readdlm("flatplate_OF_wall_kOmega_highRe.csv", ',', Float64, skipstart=1)
+# oRex = OF_data[:,7].*velocity[1]./nu[1]
+# oCf = sqrt.(OF_data[:,12].^2 + OF_data[:,13].^2)/(0.5*velocity[1]^2)
 
-tauw, pos = wall_shear_stress(:wall, model)
-tauMag = [norm(tauw[i]) for i ∈ eachindex(tauw)]
-x = [pos[i][1] for i ∈ eachindex(pos)]
-Rex = velocity[1].*x./nu
+# tauw, pos = wall_shear_stress(:wall, model)
+# tauMag = [norm(tauw[i]) for i ∈ eachindex(tauw)]
+# x = [pos[i][1] for i ∈ eachindex(pos)]
+# Rex = velocity[1].*x./nu
 
-x_corr = [0:0.0002:1;]
-Rex_corr = velocity[1].*x_corr/nu
-Cf_corr = 0.074.*(Rex_corr).^(-1/5)
-plot(; xaxis="Rex", yaxis="Cf")
[…]
