### 2024-12-04 24bf300
complete benchmark entry in documentation
