### 2024-12-04 c773871
add entry to changelog
