+++ src/Calculate/Calculate_0_gradient.jl
@@ -21,7 +21,6 @@ function Adapt.adapt_structure(to, itp::Grad{S}) where {S}
 end
 
 # Grad outer constructor for scalar field definition
-
 Grad{S}(phi::ScalarField) where S= begin
     # Retrieve mesh and define grad as vector field
     mesh = phi.mesh
@@ -38,7 +37,6 @@ Grad{S}(phi::ScalarField) where S= begin
 end
 
 # Grad outer constructor for vector field definition
-
 Grad{S}(psi::VectorField) where S = begin
     # Retrieve mesh and define grad as tensor field
     mesh = psi.mesh
@@ -55,13 +53,6 @@ Grad{S}(psi::VectorField) where S = begin
 end
 
 Base.getindex(grad::Grad{S,F,R,I,M}, i::Integer) where {S,F,R<:VectorField,I,M} = begin
-    # Tf = eltype(grad.result.x.values)
-    # SVector{3,Tf}(
-    #     grad.result.x[i], 
-    #     grad.result.y[i], 
-    #     grad.result.z[i]
-    #     )
-    # Tf = eltype(grad.result.x.values)
     @inbounds SVector{3}(
         grad.result.x[i], 
         grad.result.y[i], 
