+++ src/ModelPhysics/Turbulence/RANS_laminar.jl
@@ -54,7 +54,7 @@ end
 
 # Model solver call (implementation)
 """
-    turbulence!(rans::LaminarModel, model::Physics{T,F,M,Tu,E,D,BI}, S, S2, prev, time, config
+    turbulence!(rans::LaminarModel, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, config
     ) where {T,F,M,Tu<:Laminar,E,D,BI}
 
 Run turbulence model transport equations.
@@ -63,14 +63,13 @@ Run turbulence model transport equations.
 - `rans::LaminarModel` -- Laminar turbulence model.
 - `model`  -- Physics model defined by user.
 - `S`   -- Strain rate tensor.
-- `S2`  -- Square of the strain rate magnitude.
 - `prev`  -- Previous field.
 - `time`   -- 
 - `config` -- Configuration structure defined by user with solvers, schemes, runtime and 
               hardware structures set.
 
 """
-function turbulence!(rans::LaminarModel, model::Physics{T,F,M,Tu,E,D,BI}, S, S2, prev, time, config
+function turbulence!(rans::LaminarModel, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, config
     ) where {T,F,M,Tu<:Laminar,E,D,BI}
     nothing
 end
