+++ src/Solvers/Solvers_2_PISO.jl
@@ -64,11 +64,10 @@ function PISO(
     # Define aux fields 
     gradU = Grad{schemes.U.gradient}(U)
     gradUT = T(gradU)
-    S = StrainRate(gradU, gradUT)
-    S2 = ScalarField(mesh)
+    Uf = FaceVectorField(mesh)
+    S = StrainRate(gradU, gradUT, U, Uf)
 
     n_cells = length(mesh.cells)
-    Uf = FaceVectorField(mesh)
     pf = FaceScalarField(mesh)
     Hv = VectorField(mesh)
     rD = ScalarField(mesh)
@@ -171,11 +170,7 @@ function PISO(
         
         # correct_mass_flux(mdotf, p, rDf, config) # new approach
 
-
-    # grad!(gradU, Uf, U, U.BCs, time, config)
-    # limit_gradient && limit_gradient!(gradU, U, config)
-
-    turbulence!(turbulenceModel, model, S, S2, prev, time, config) 
+    turbulence!(turbulenceModel, model, S, prev, time, config) 
     update_nueff!(nueff, nu, model.turbulence, config)
 
     residual!(R_ux, U_eqn, U.x, iteration, xdir, config)

[…]
