+++ examples/2D_BFS.jl
@@ -8,7 +8,7 @@ mesh_file = joinpath(grids_dir, grid)
 
 mesh = UNV2D_mesh(mesh_file, scale=0.001)
 
-backend = CPU(); # activate_multithread(backend)
+backend = CPU(); activate_multithread(backend)
 mesh_dev = mesh
 # backend = CUDABackend()
 # mesh_dev = adapt(backend, mesh)
