### 2024-12-09 4a13012
Add benchmark to release notes
