### 2024-12-09 e64ad0c
remove old implementation of gradient limiter
