+++ src/Calculate/Calculate_0_gradient.jl
@@ -331,8 +331,9 @@ end
 end
 
 
-# TEMP LOCATION
-function limit_gradient!(xv, yv, zv, F, config)
+### GRADIENT LIMITER - EXPERIMENTAL
+
+function limit_gradient!(∇F, F, config)
     (; hardware) = config
     (; backend, workgroup) = hardware
 
@@ -341,76 +342,90 @@ function limit_gradient!(xv, yv, zv, F, config)
 
     minPhi0 = maximum(F.values) # use min value so all values compared are larger
     maxPhi0 = minimum(F.values)
+    (; x, y, z) = ∇F.result
 
     kernel! = _limit_gradient!(backend, workgroup)
-    kernel!(xv, yv, zv, F, cells, cell_neighbours, cell_faces, cell_nsign, faces, minPhi0, maxPhi0, ndrange=length(cells))
+    kernel!(x, y, z, F, cells, cell_neighbours, cell_faces, cell_nsign, faces, minPhi0, maxPhi0, ndrange=length(cells))
     KernelAbstractions.synchronize(backend)
 end
 
-@kernel function _limit_gradient!(xv, yv, zv, F, cells, cell_neighbours, cell_faces, cell_nsign, faces, minPhi, maxPhi)
+function limit_gradient!(∇F::Grad{S,FF,R,I,M}, F, config) where {S,FF,R<:TensorField,I,M}
+    (; hardware) = config
+    (; backend, workgroup) = hardware
+
+    mesh = F.mesh
+    (; cells, cell_neighbours, cell_faces, cell_nsign, faces) = mesh
+
+    (; xx, yx, zx) = ∇F.result
+    (; xy, yy, zy) = ∇F.result
+    (; xz, yz, zz) = ∇F.result
+
+    kernel! = _limit_gradient!(backend, workgroup)
+    kernel!(xx, yx, zx, F.x, cells, cell_neighbours, cell_faces, cell_nsign, faces, ndrange=length(cells))
+    KernelAbstractions.synchronize(backend)
+
+    kernel! = _limit_gradient!(backend, workgroup)
+    kernel!(xy, yy, zy, F.y, cells, cell_neighbours, cell_faces, cell_nsign, faces, ndrange=length(cells))
+    KernelAbstractions.synchronize(backend)
+
+    kernel! = _limit_gradient!(backend, workgroup)
+    kernel!(xz, yz, zz, F.z, cells, cell_neighbours, cell_faces, cell_nsign, faces, ndrange=length(cells))
+    KernelAbstractions.synchronize(backend)
+end
+
+@kernel function _limit_gradient!(x, y, z, F, cells, cell_neighbours, cell_faces, cell_nsign, faces)
     cID = @index(Global)
-    # mesh = F.mesh
-    # (; cells, cell_neighbours, cell_faces, cell_nsign, faces) = mesh
 
-    # minPhi0 = maximum(F.values) # use min value so all values compared are larger
-    # maxPhi0 = minimum(F.values)
+    cell = cells[cID]
+    faces_range = cell.faces_range
+    phiP = F[cID]
+    phiMax = phiMin = phiP
+ 
+    for fi ∈ faces_range
+        nID = cell_neighbours[fi]
+        phiN = F[nID]
+        phiMax = max(phiN, phiMax)
+        phiMin = min(phiN, phiMin)
+    end
 
-    # for (cID, cell) ∈ enumerate(cells)
-        cell = cells[cID]
-        # minPhi = minPhi0 # reset for next cell
-        # maxPhi = maxPhi0
+    # g0 = ∇F[cID]
+    grad0 = SVector{3}(x[cID] , y[cID] , z[cID])
 
-        # find min and max values around cell
-        faces_range = cell.faces_range
-        
[…]
