### 2024-12-09 628268a
draft implementation of wrapper type for sparse matrix on CPU
