+++ src/ModelFramework/ModelFramework_2_access_functions.jl
@@ -79,14 +79,21 @@ end
 end
 
 
-@inline _nzval(A::SparseArrays.SparseMatrixCSC) = A.nzval
+# @inline _nzval(A::SparseArrays.SparseMatrixCSC) = A.nzval
+# @inline _rowptr(A::SparseArrays.SparseMatrixCSC) = A.rowptr
+# @inline _colval(A::SparseArrays.SparseMatrixCSC) = A.colval
 
-@inline _rowptr(A::SparseArrays.SparseMatrixCSC) = A.rowptr
+# @inline get_sparse_fields(A::SparseArrays.SparseMatrixCSC) = begin
+#     A.nzval, A.colval, A.rowptr
+# end
 
-@inline _colval(A::SparseArrays.SparseMatrixCSC) = A.colval
+@inline _nzval(A::SparseXCSR) = parent(A).nzval
+@inline _rowptr(A::SparseXCSR) = parent(A).rowptr
+@inline _colval(A::SparseXCSR) = parent(A).colval
 
-@inline get_sparse_fields(A::SparseArrays.SparseMatrixCSC) = begin
-    A.nzval, A.colval, A.rowptr
+@inline get_sparse_fields(A::SparseXCSR) = begin
+    Acsr = parent(A)
+    Acsr.nzval, Acsr.colval, Acsr.rowptr
 end
 
 # Sparse CSR (temporary hack)
