### 2024-12-09 9389fd8
dummy implementation for `activate_multithread`
