### 2024-12-10 c790609
fix call to set number of BLAS threads
