+++ src/Multithread/spmvm.jl
@@ -75,44 +75,10 @@ end
 # """
     function activate_multithread(backend::CPU)
 
-        LinearAlgebra.BLAS.set_num_threads(1)
-
-        # @eval function  mul!(y::AbstractVector, A::SparseMatrixCSR, x::AbstractVector, alpha::Number, beta::Number)
-        #     return xmul!(y, A, x, alpha, beta)
-        # end
-
-        # @eval function  mul!(y::AbstractVector, A::SparseMatrixCSR, x::AbstractVector)
-        #     return xmul!(y, A, x, true, false)
-        # end
-
-        # @eval function  *(A::SparseMatrixCSR, x::AbstractVector)
-        #     return xmul(A, x)
-        # end
+        BLAS.set_num_threads(1)
 
         nothing
     end
-
-# Function to activate multithreading for CSR sparse matrices. The only input required is the backend (which must be `CPU()`).
-
-# """
-# function activate_multithread(backend::CPU)
-
-#     LinearAlgebra.BLAS.set_num_threads(1)
-
-#     @eval function  LinearAlgebra.mul!(y::AbstractVector, A::SparseMatrixCSR, x::AbstractVector, alpha::Number, beta::Number)
-#         return xmul!(y, A, x, alpha, beta)
-#     end
-    
-#     @eval function  LinearAlgebra.mul!(y::AbstractVector, A::SparseMatrixCSR, x::AbstractVector)
-#         return xmul!(y, A, x, true, false)
-#     end
-    
-#     @eval function  Base.:*(A::SparseMatrixCSR, x::AbstractVector)
-#         return xmul(A, x)
-#     end
-
-#     nothing
-# end
 
 function  LinearAlgebra.mul!(y::AbstractVector, A::SparseXCSR, x::AbstractVector, alpha::Number, beta::Number)
     return xmul!(y, A, x, alpha, beta)
@@ -125,5 +91,3 @@ end
 function  Base.:*(A::SparseMatrixCSR, x::SparseXCSR)
     return xmul(A, x)
 end
-
-
