### 2024-12-10 efcf7bf
clean multithreaded code
