+++ src/ModelPhysics/Turbulence/LES_Smagorinsky.jl
@@ -94,7 +94,7 @@ Run turbulence model transport equations.
 
 """
 function turbulence!(
-    les::SmagorinskyModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, config
+    les::SmagorinskyModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, limit_gradient, config
     ) where {T,F,M,Tu<:Smagorinsky,E,D,BI,E1,E2}
 
     mesh = model.domain
@@ -104,6 +104,7 @@ function turbulence!(
     (; Δ, magS) = les
 
     grad!(gradU, Uf, U, U.BCs, time, config) # update gradient (internal structure of S)
+    limit_gradient && limit_gradient!(gradU, U, config)
     magnitude!(magS, S, config)
     @. magS.values *= sqrt(2) # should fuse into definition of magnitude function!
 
