### 2024-12-10 ef08e35
reactivate all tests
