+++ test/runtests.jl
@@ -8,35 +8,35 @@ TEST_CASES_DIR = pkgdir(XCALibre, "test/0_TEST_CASES")
 
 @testset verbose = true "Functionality tests" begin
 
-    # @testset "Mesh conversion" begin
-    #     include("test_mesh_conversion.jl")
-    # end
+    @testset "Mesh conversion" begin
+        include("test_mesh_conversion.jl")
+    end
 
-    # @testset "Smoothers" begin
-    #     include("test_smoothers.jl")
-    # end
+    @testset "Smoothers" begin
+        include("test_smoothers.jl")
+    end
 
-    # @testset "DILU" begin
-    #     include("test_DILU.jl")
-    # end
+    @testset "DILU" begin
+        include("test_DILU.jl")
+    end
 
-    # @testset "Incompressible" begin
+    @testset "Incompressible" begin
 
-    #     test_files = [
-    #         "2d_incompressible_laminar_BFS.jl",
-    #         "2d_incompressible_flatplate_KOmega_lowRe.jl",
-    #         "2d_incompressible_flatplate_KOmega_HighRe.jl",
-    #         "3d_incompressible_laminar_BFS.jl",
-    #         "3d_incompressible_laminar_cascade_periodic.jl",
-    #         "2d_incompressible_transient_laminar_BFS.jl",
-    #         "2d_incompressible_transient_KOmega_BFS_lowRe.jl"
-    #     ]
+        test_files = [
+            "2d_incompressible_laminar_BFS.jl",
+            "2d_incompressible_flatplate_KOmega_lowRe.jl",
+            "2d_incompressible_flatplate_KOmega_HighRe.jl",
+            "3d_incompressible_laminar_BFS.jl",
+            "3d_incompressible_laminar_cascade_periodic.jl",
+            "2d_incompressible_transient_laminar_BFS.jl",
+            "2d_incompressible_transient_KOmega_BFS_lowRe.jl"
+        ]
 
-    #     for test ∈ test_files
-    #         test_path = joinpath(TEST_CASES_DIR, test)
-    #         include(test_path)
-    #     end
-    # end
+        for test ∈ test_files
+            test_path = joinpath(TEST_CASES_DIR, test)
+            include(test_path)
+        end
+    end
 
     @testset "Compressible" begin
         test_files = [
