### 2024-12-10 7b4a1e9
Add entry to changelog and up package version for release
