### 2024-12-10 #7441 @mberto79
@AnjaliSandip thank you for following this up. We have just released v0.3.3 with some good performance improvements and housekeeping, the version we initially submitted was v0.3.1, should I merge the latest changes into the git branch that has contains the paper? (@vlc1 and @simone-silvertry FYI). Thanks!
