### 2024-12-11 0540403
ensure `limit_gradient` is exported
