### 2024-12-12 30cf8e1
update project version to DEV
