### 2024-12-12 #7441 @mberto79
@vlc1 and @simone-silvestri just a quick note to let you know that I have now merged the code for version v0.3.3 into the branch that contains the paper (`joss-submission`).
