### 2024-12-12 c6c9d85
start playing with gradient implementation
