### 2024-12-13 d87e781
draft implementation of new method for vector gradient calculation
