+++ src/Calculate/Calculate_1_green_gauss.jl
@@ -1,9 +1,8 @@
 export green_gauss!
 
-# Green gauss function definition
+# Green gauss for FaceScalarField
 
 function green_gauss!(dx, dy, dz, phif, config)
-    # Retrieve required varaibles for function
     (; hardware) = config
     (; backend, workgroup) = hardware
     
@@ -12,10 +11,9 @@ function green_gauss!(dx, dy, dz, phif, config)
     kernel!(dx, dy, dz, phif, ndrange=length(dx))
     KernelAbstractions.synchronize(backend)
 
-    # Retrieve number of boundary faces
+    # number of boundary faces
     nbfaces = length(phif.mesh.boundary_cellsID)
     
-    # Launch boundary faces contribution kernel
     kernel! = boundary_faces_contribution!(backend, workgroup)
     kernel!(dx, dy, dz, phif, ndrange=nbfaces)
     KernelAbstractions.synchronize(backend)
@@ -31,26 +29,19 @@ end
     end
      
     @inbounds begin
-        # Extract required fields from work item cell
         (; volume, faces_range) = cells[i]
 
-        # Allocate static vector for results
         res = SVector{3}(0.0,0.0,0.0)
 
-        # Loop over cell faces to calculate result
         for fi ∈ faces_range
-            # Define reqquired variables
             fID = cell_faces[fi]
             nsign = cell_nsign[fi]
             (; area, normal) = faces[fID]
             
-            # Result calcaultion
             res += values[fID]*(area*normal*nsign)
         end
-        # Normalise results with cell volume
         res /= volume
 
-        # Store results in x, y, z value array
         dx.values[i] = res[1]
         dy.values[i] = res[2]
         dz.values[i] = res[3]
@@ -68,23 +59,108 @@ end
     end
 
     @inbounds begin
-        # Extract required variables from work item face and cell
         (; ownerCells, area, normal) = faces[i]
         cID = ownerCells[1]
         (; volume) = cells[cID]
 
-        # Allocate static vector for results
-        # res = SVector{3}(0.0,0.0,0.0)
-
-        # Results calculation
         res = values[i]*(area*normal)
-
-        # Normalise results with cell volume
         res /= volume 
         
-        # Increment x, y, z values vectors with results
         Atomix.@atomic dx.values[cID] += res[1]
         Atomix.@atomic dy.values[cID] += res[2]
         Atomix.@atomic dz.values[cID] += res[3]
+    end
+end
+
+# Green gauss for FaceVectorField
+
+function green_gauss!(
+    grad::Grad{Orthogonal,F,R,I,M}, psif, config) where {F,R<:TensorField,I,M}
+
+    (; hardware) = config
+    (; backend, workgroup) = hardware
+
+    (; xx, xy, xz, yx, yy, yz, zx, zy, zz) = grad.result
+    
+    # Launch result calculation kernel
+    kernel! = _green_gauss_vector!(backend, workgroup)
+    kernel!(xx, xy, xz, yx, yy, yz, zx, zy, zz, psif, ndrange=length(xx))
+    KernelAbstractions.synchronize(backend)
+
+    # number of boundary faces
+    nbfaces = length(psif.mesh.boundary_cellsID)
+    
+    kernel! = boundary_faces_contribution_vector!(backend, workgroup)
+    kernel!(xx, xy, xz, yx, yy, yz, zx, zy, zz, psif, ndrange=nbfaces)
+    KernelAbstractions.synchronize(backend)
+end
+
+# Green Gauss kernel definition
[…]
