### 2024-12-14 aa0087e
simplify access to transpose and return to neighbour cell based limiter
