### 2024-12-16 24f778c
New call logic for Midpoint gradient calculation with nice gains for vector gradients
