### 2024-12-16 50cf8c5
fix logic to set vectorfield
