+++ src/Fields/Fields_0_types.jl
@@ -141,12 +141,12 @@ end
 
 # Base.getindex(v::AbstractVectorField, i::Integer) = @inbounds SVector{3, eltype(v.x)}(v.x[i], v.y[i], v.z[i])
 Base.getindex(v::AbstractVectorField, i::Integer) = @inbounds SVector{3}(v.x[i], v.y[i], v.z[i])
-Base.setindex!(v::AbstractVectorField, x::SVector{3, T}, i::Integer) where T= begin
+Base.setindex!(v::AbstractVectorField, vec::SVector{3, T}, i::Integer) where T= begin
     # length(x) == 3 || throw("Vectors must have 3 components")
     (; x, y, z) = v
-    x[i] = x[1]
-    y[i] = y[2]
-    z[i] = z[3]
+    x[i] = vec[1]
+    y[i] = vec[2]
+    z[i] = vec[3]
 end
 Base.length(v::AbstractVectorField) = length(v.x)
 Base.eachindex(v::AbstractVectorField) = eachindex(v.x)
