### 2024-12-16 be0cf50
clean variables
