+++ examples/2D_cylinder_heated.jl
@@ -3,10 +3,12 @@ using XCALibre
 # using CUDA # Run this if using NVIDIA GPU
 # using AMDGPU # Run this if using AMD GPU
 
-# quad, backwardFacingStep_2mm, backwardFacingStep_10mm, trig40
-mesh_file = "unv_sample_meshes/cylinder_d10mm_5mm.unv"
-# mesh_file = "unv_sample_meshes/cylinder_d10mm_2mm.unv"
-# mesh_file = "unv_sample_meshes/cylinder_d10mm_10-7.5-2mm.unv"
+grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
+grid = "cylinder_d10mm_5mm.unv"
+# grid = "cylinder_d10mm_2mm.unv"
+# grid = "cylinder_d10mm_10-7.5-2mm.unv"
+mesh_file = joinpath(grids_dir, grid)
+
 mesh = UNV2D_mesh(mesh_file, scale=0.001)
 
 # mesh_dev = adapt(CUDABackend(), mesh)
@@ -25,11 +27,11 @@ Pr = 0.7
 
 model = Physics(
     time = Steady(),
-    fluid = WeaklyCompressible(
-        mu = nu,
-        cp = ConstantScalar(cp),
-        gamma = ConstantScalar(gamma),
-        Pr = ConstantScalar(Pr)
+    fluid = Fluid{WeaklyCompressible}(
+        nu = nu,
+        cp = cp,
+        gamma = gamma,
+        Pr = Pr
         ),
     turbulence = RANS{Laminar}(),
     energy = Energy{SensibleEnthalpy}(Tref=288.15),
@@ -78,8 +80,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.2,
-        rtol = 1e-4,
-        atol = 1e-3
+        rtol = 1e-4
     ),
     h = set_solver(
         model.energy.h;
@@ -87,8 +88,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.8,
-        rtol = 1e-4,
-        atol = 1e-2
+        rtol = 1e-4
     )
 )
 
@@ -100,7 +100,7 @@ schemes = (
 
 runtime = set_runtime(iterations=1000, write_interval=100, time_step=1)
 
-hardware = set_hardware(backend=CPU(), workgroup=4)
+hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
 # hardware = set_hardware(backend=ROCBackend(), workgroup=32)
 
@@ -115,7 +115,7 @@ initialise!(model.energy.T, temp)
 
 println("Maxh ", maximum(model.energy.T.values), " minh ", minimum(model.energy.T.values))
 
-residuals = run!(model, config); #, pref=0.0)
+residuals = run!(model, config, ncorrectors=2); #, pref=0.0)
 
 plot(; xlims=(0,runtime.iterations), ylims=(1e-8,0))
 plot!(1:length(Rx), Rx, yscale=:log10, label="Ux")
