+++ prototype/gradient_calculation_vector.jl
@@ -14,7 +14,7 @@ backend = CPU()
 psi = VectorField(mesh)
 psif = FaceVectorField(mesh)
 gradScheme = Orthogonal
-# gradScheme = Midpoint
+gradScheme = Midpoint
 ∇psi = Grad{gradScheme}(psi)
 
 psi = assign(
@@ -54,15 +54,3 @@ Tu = 0.025
 k_inlet = 3/2*(Tu*Ux)^2
 ω_inlet = k_inlet/(νR*nu)
 k_inlet/ω_inlet
-
-function test(psif::Field) where Field 
-    a = ifelse(
-        <:(Field, AbstractVectorField), 
-        "Yep, this works", 
-        "Nope, you passed soemthing else")
-    
-    println(a)
-    nothing
-end
-
-test(psif)
