+++ src/Calculate/Calculate_0_gradient.jl
@@ -109,12 +109,6 @@ end
 function grad!(grad::Grad{Orthogonal,F,R,I,M}, psif, psi, BCs, time, config) where {F,R<:TensorField,I,M}
     interpolate!(psif, psi, config)
     correct_boundaries!(psif, psi, BCs, time, config)
-
-    # Launch green-gauss for all tensor field dimensions
-    # green_gauss!(grad.result.xx, grad.result.yx, grad.result.zx, psif.x, config)
-    # green_gauss!(grad.result.xy, grad.result.yy, grad.result.zy, psif.y, config)
-    # green_gauss!(grad.result.xz, grad.result.yz, grad.result.zz, psif.z, config)
-
     green_gauss!(grad, psif, config)
 end
 
@@ -126,7 +120,6 @@ function interpolate_midpoint!(phif::FaceScalarField, phi::ScalarField, config)
     # Extract required variables for function
     (; mesh) = phi
     (; faces) = mesh
-    # backend = _get_backend(mesh)
     (; hardware) = config
     (; backend, workgroup) = hardware
 
@@ -153,15 +146,12 @@ end
     end
 end
 
-# Interpolate function for vector field definition (NEEDS KERNEL IMPLEMENTATION!!!!!!)
-
 # Scalar field calculation definition
 
 function interpolate_midpoint!(phif::FaceVectorField, phi::VectorField, config)
     # Extract required variables for function
     (; mesh) = phi
     (; faces) = mesh
-    # backend = _get_backend(mesh)
     (; hardware) = config
     (; backend, workgroup) = hardware
 
@@ -184,7 +174,6 @@ end
     @inbounds begin
         # Retrieve face, weight and ownerCells for loop iteration
         face = faces[fID]
-        # weight = face.weight
         ownerCells = face.ownerCells
         c1 = ownerCells[1]; c2 = ownerCells[2]
         
@@ -200,39 +189,6 @@ end
     end
 end
 
-# interpolate_midpoint!(psif::FaceVectorField, psi::VectorField, config) = begin
-#     # Extract individual value vectors from vector field
-#     (; x, y, z) = psif
-
-#     # Retrieve mesh and faces
-#     mesh = psi.mesh
-#     faces = mesh.faces
-    
-#     # Set initial weight
-#     weight = 0.5
-    
-#     @inbounds begin
-#         # Loop over all faces to calculate interpolation
-#         for fID ∈ eachindex(faces)
-#             # Retrieve face, weight and ownerCells for loop iteration
-#             face = faces[fID]
-#             weight = face.weight
-#             ownerCells = face.ownerCells
-#             c1 = ownerCells[1]; c2 = ownerCells[2]
-            
-#             # Set values to interpolate between
-#             x1 = psi.x[c1]; x2 = psi.x[c2]
-#             y1 = psi.y[c1]; y2 = psi.y[c2]
-#             z1 = psi.z[c1]; z2 = psi.z[c2]
-
-#             # Interpolate calculation
-#             x[fID] = weight*(x1 + x2)
-#             y[fID] = weight*(y1 + y2)
-#             z[fID] = weight*(z1 + z2)
-#         end
-#     end
-# end
-
 # Correct interpolation function definition
 
 function correct_interpolation!(grad, phif, phi, config)
@@ -240,7 +196,6 @@ function correct_interpolation!(grad, phif, phi, config)
     (; mesh) = phif
     (; faces, cells) = mesh
     nbfaces = length(mesh.boundary_cellsID)
-    # backend = _get_backend(mesh)
     (; hardware) = config
[…]
