### 2024-12-17 60db8ba
removed old function implementation
