### 2024-12-17 7dc0dd7
minor updates for some examples
