+++ examples/2D_cylinder_RANS.jl
@@ -1,6 +1,6 @@
 using Plots
 using XCALibre
-using CUDA
+# using CUDA
 
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
@@ -136,7 +136,7 @@ initialise!(model.turbulence.k, k_inlet)
 initialise!(model.turbulence.omega, ω_inlet)
 initialise!(model.turbulence.nut, k_inlet/ω_inlet)
 
-residuals = run!(model, config); #, pref=0.0)
+residuals = run!(model, config, ncorrectors=2); #, pref=0.0)
 
 Reff = stress_tensor(model.momentum.U, nu, model.turbulence.nut)
 Fp = pressure_force(:cylinder, model.momentum.p, 1.25)
