### 2024-12-17 ab10641
add diffusion solver to test nonorthoginality effects more easily
