### 2024-12-17 bfb78d5
fix/replace all weight functions
