+++ src/UNV2/UNV2_2_geometry.jl
@@ -68,7 +68,8 @@ function internal_face_properties!(mesh::Mesh2{I,F}) where {I,F}
         # Calculate delta and interpolation weight
         delta = norm(d_12) 
         e = d_12/delta
-        weight = abs((d_1f⋅normal)/(d_1f⋅normal + d_f2⋅normal)) 
+        # weight = abs((d_1f⋅normal)/(d_1f⋅normal + d_f2⋅normal)) 
+        weight = norm(d_f2)/norm(d_12)
 
         # Assign values to face
         face = @set face.area = area
