### 2024-12-17 0a375a0
add `ncorrectors` to heated cylinder case
