### 2024-12-17 da90fdb
relax variable right after solving
