### 2024-12-18 d16d921
add `eltype` method for both `Scalar` and `VectorField`
