### 2024-12-19 af92a29
temporarily active new `limit_gradient!` function
