### 2024-12-19 e2d08ee
change default for `limit_gradient` to `nothing
