+++ src/Solvers/Solvers_3_solver_dispatch.jl
@@ -3,7 +3,7 @@ export run!
 """
     function run!(
         model::Physics, config; 
-        limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+        limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
         )
 
         # here an internal function is used for solver dispatch
@@ -48,7 +48,7 @@ run!() = nothing # dummy function for providing general documentation
 """
     run!(
         model::Physics{T,F,M,Tu,E,D,BI}, config;
-        limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+        limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
         ) where{T<:Steady,F<:Incompressible,M,Tu,E,D,BI} = 
     begin
         residuals = simple!(model, config, pref=pref)
@@ -73,7 +73,7 @@ This function returns a `NamedTuple` for accessing the residuals (e.g. `residual
 """
 run!(
     model::Physics{T,F,M,Tu,E,D,BI}, config; 
-    limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
     ) where{T<:Steady,F<:Incompressible,M,Tu,E,D,BI} = 
 begin
     residuals = simple!(
@@ -90,7 +90,7 @@ end
 """
     run!(
         model::Physics{T,F,M,Tu,E,D,BI}, config; 
-        limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+        limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
         ) where{T<:Transient,F<:Incompressible,M,Tu,E,D,BI} = 
     begin
         residuals = piso!(model, config, pref=pref); #, pref=0.0)
@@ -117,7 +117,7 @@ This function returns a `NamedTuple` for accessing the residuals (e.g. `residual
 """
 run!(
     model::Physics{T,F,M,Tu,E,D,BI}, config; 
-    limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=2
+    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=2
     ) where{T<:Transient,F<:Incompressible,M,Tu,E,D,BI} = 
 begin
     residuals = piso!(
@@ -134,7 +134,7 @@ end
 """
     run!(
         model::Physics{T,F,M,Tu,E,D,BI}, config; 
-        limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+        limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
         ) where{T<:Steady,F<:WeaklyCompressible,M,Tu,E,D,BI} = 
     begin
         residuals = csimple!(model, config, pref=pref); #, pref=0.0)
@@ -162,7 +162,7 @@ This function returns a `NamedTuple` for accessing the residuals (e.g. `residual
 """
 run!(
     model::Physics{T,F,M,Tu,E,D,BI}, config; 
-    limit_gradient=false, pref=nothing, ncorrectors=0, inner_loops=0
+    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0
     ) where{T<:Steady,F<:WeaklyCompressible,M,Tu,E,D,BI} = 
 begin
     residuals = csimple!(
@@ -178,7 +178,7 @@ end
 # Compressible solver (steady)
 run!(
     model::Physics{T,F,M,Tu,E,D,BI}, config; 
[…]
