### 2024-12-19 a924300
implementation of `symmetry` for `ScalarField` dispatch
