### 2024-12-19 28b143f
add `Symmetry` boundary condition for `Si` operator
