### 2024-12-20 4ba3c48
move synchonize out of the loop
