+++ src/Discretise/Discretise_5_apply_bcs.jl
@@ -50,9 +50,9 @@ function _apply_boundary_conditions!(
         kernel!(
             model, BC, model.terms, faces, cells, start_ID, boundary_cellsID, colval, rowptr, nzval, b, ione, component, time, ndrange=kernel_range
             )
-        KernelAbstractions.synchronize(backend)
+        # KernelAbstractions.synchronize(backend)
     end
-    # KernelAbstractions.synchronize(backend)
+    KernelAbstractions.synchronize(backend)
     nothing
 end
 
