### 2024-12-23 dc5a0a9
update all solvers to use new call signature for `limit_gradient`
