### 2024-12-23 f2afa54
remove synchonisation in gradient and green gauss kernels for testing
