### 2025-01-06 #7441 @vlc1
@mberto79 and coauthors, nice work, thanks for your contribution. A couple of questions:

1. The tests took quite a bit of time to run on my laptop (30 minutes or so). Are they all required, or would it be possible leave some of them in the examples (or reduce the mesh sizes in the grid files)?
2. Gerris has been rewritten a few years back, maybe reference its successor instead? ([Basilisk](basilisk.fr))
