### 2025-01-09 d278673
revert symmetry boundary condition and update test case
