### 2025-01-09 b248e69
revert number of iterations in test case
