+++ test/0_TEST_CASES/2d_incompressible_laminar_BFS.jl
@@ -66,7 +66,7 @@ solvers = (
 )
 
 runtime = set_runtime(
-    iterations=2000, time_step=1, write_interval=2000)
+    iterations=1000, time_step=1, write_interval=1000)
 
 hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
