### 2025-01-10 99a4a96
fix BFS incompressible test case
