+++ examples/2D_BFS.jl
@@ -41,7 +41,7 @@ model = Physics(
     Dirichlet(:outlet, 0.0),
     Neumann(:wall, 0.0),
     # Neumann(:top, 0.0)
-    Symmetry(:top, 0.0)
+    Symmetry(:top)
 )
 
 schemes = (
@@ -56,7 +56,7 @@ solvers = (
         model.momentum.U;
         solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
         preconditioner = Jacobi(), # ILU0GPU, Jacobi, DILU
-        # smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
+        smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         convergence = 1e-7,
         relax       = 0.8,
         rtol = 1e-1
@@ -65,7 +65,7 @@ solvers = (
         model.momentum.p;
         solver      = CgSolver, # BicgstabSolver, GmresSolver, CgSolver
         preconditioner = Jacobi(), # IC0GPU, Jacobi, DILU
-        # smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
+        smoother=JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         convergence = 1e-7,
         relax       = 0.2,
         rtol = 1e-2
