+++ examples/2D_flatplate_fixedT.jl
@@ -85,7 +85,7 @@ solvers = (
 runtime = set_runtime(iterations=1000, write_interval=100, time_step=1)
 
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
-hardware = set_hardware(backend=CPU(), workgroup=cld(length(mesh.faces),4))
+hardware = set_hardware(backend=CPU(), workgroup=cld(length(mesh.cells),4))
 
 config = Configuration(
     solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
