### 2025-01-10 050ec38
add smoother to BFS test to test top level API
