+++ test/0_TEST_CASES/2d_incompressible_laminar_BFS.jl
@@ -49,18 +49,20 @@ schemes = (
 solvers = (
     U = set_solver(
         model.momentum.U;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+        solver = BicgstabSolver, # BicgstabSolver, GmresSolver
+        smoother = JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         preconditioner = Jacobi(),
         convergence = 1e-7,
-        relax       = 0.8,
+        relax = 0.8,
         rtol = 1e-2,
     ),
     p = set_solver(
         model.momentum.p;
-        solver      = CgSolver, # BicgstabSolver, GmresSolver
+        solver = CgSolver, # BicgstabSolver, GmresSolver
+        smoother = JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
         preconditioner = Jacobi(),
         convergence = 1e-7,
-        relax       = 0.2,
+        relax = 0.2,
         rtol = 1e-3,
     )
 )
