### 2025-01-10 f50c017
calculate d.d only once
