### 2025-01-13 c7b4360
fix call to smoothers on GPU
