+++ src/Solve/Smoothers/Smoothers_jacobi.jl
@@ -91,7 +91,7 @@ function _smoother_launch(
     backend::GPU, smoother, x, b, nzval, colval, rowptr, workgroup
     )
     krange = length(x)
-    kernel! = _apply_smoother!(backend, workgroup)
+    kernel! = _apply_smoother_gpu!(backend, workgroup)
     for _ ∈ 1:smoother.loops
         kernel!(smoother, x, b, nzval, colval, rowptr, ndrange = krange)
         KernelAbstractions.synchronize(backend)
@@ -99,7 +99,7 @@ function _smoother_launch(
     end
 end
 
-@kernel function _apply_smoother!(
+@kernel function _apply_smoother_gpu!(
     smoother::JacobiSmoother, x, b, nzval, colval, rowptr
     )
     cID = @index(Global)
