### 2025-01-15 8f49bd6
update KOmega to new limiter
