### 2025-01-19 6a30c57
set up cases to run on GPU
