### 2025-01-20 a1049c0
change case to run with GPU
