+++ examples/3D_BFS.jl
@@ -1,9 +1,9 @@
 using Plots 
-using ThreadPinning
+# using ThreadPinning
 using XCALibre
-# using CUDA
+using CUDA
 
-pinthreads(:cores)
+# pinthreads(:cores)
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
 # grid = "bfs_unv_tet_4mm.unv"
@@ -13,17 +13,17 @@ mesh_file = joinpath(grids_dir, grid)
 
 # mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
 # mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
-
-
 # mesh_file = "bfs_unv_tet_5mm.unv"
+
+mesh_file = "/Users/hmedi/Desktop/BFS_GRIDS/bfs_unv_tet_4mm.unv"
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
 
-workgroup = cld(length(mesh.cells), Threads.nthreads())
+# workgroup = cld(length(mesh.cells), Threads.nthreads())
 # backend = CPU(); activate_multithread(backend)
-backend = CPU()
-activate_multithread1()
-# workgroup = 32
-# backend = CUDABackend()
+# backend = CPU()
+# activate_multithread1()
+workgroup = 32
+backend = CUDABackend()
 
 mesh_dev = adapt(backend, mesh)
 
@@ -110,9 +110,9 @@ initialise!(model.momentum.p, 0.0)
 
 @time residuals = run!(model, config)
 
-iterations = runtime.iterations
-plot(yscale=:log10, ylims=(1e-7,1e-1))
-plot!(1:iterations, residuals.Ux, label="Ux")
-plot!(1:iterations, residuals.Uy, label="Uy")
-plot!(1:iterations, residuals.Uz, label="Uz")
-plot!(1:iterations, residuals.p, label="p")
+# iterations = runtime.iterations
+# plot(yscale=:log10, ylims=(1e-7,1e-1))
+# plot!(1:iterations, residuals.Ux, label="Ux")
+# plot!(1:iterations, residuals.Uy, label="Uy")
+# plot!(1:iterations, residuals.Uz, label="Uz")
+# plot!(1:iterations, residuals.p, label="p")
