### 2025-01-22 9c198d7
add BFS 5mm case for testing
