### 2025-01-24 846e4dd
set up changelog for current dev version and fix date of previous release
