### 2025-01-24 efbfe65
document new limiters
