+++ src/Discretise/Discretise_4_assign_boundaries.jl
@@ -167,15 +167,17 @@ end
         time=SteadyState,
         divergence=Linear, 
         laplacian=Linear, 
-        gradient=Orthogonal) = begin
+        gradient=Orthogonal,
+        limiter=nothing) = begin
         
         # Returns NamedTuple definition for scheme 
         (
             time=time,
             divergence=divergence,
             laplacian=laplacian,
-            gradient=gradient
-        )
+            gradient=gradient,
+            limiter=limiter
+        )   
     end
 
 The `set_schemes` function is used at the top-level API to help users define discretisation schemes for every field solved. It offers default values, thus users can pick and choose which entry they wish to modify.
@@ -186,6 +188,7 @@ The `set_schemes` function is used at the top-level API to help users define dis
 - `divergence` is used to set the divergence scheme (default is `Linear`) 
 - `laplacian` is used to set the laplacian scheme (default is `Linear`)
 - `gradient`  is used to set the gradient scheme (default is `Orthogonal`)
+- `limiter` is used to specify if gradient limiters should be used, currently supported limiters include `FaceBased` and `MFaceBased` (default is `nothing`)
 
 """
 set_schemes(;
