### 2025-01-24 850e6e7
add breaking change - removed `limit_gradient` keyword in solver API
