+++ src/ModelPhysics/Turbulence/RANS_kOmega.jl
@@ -144,7 +144,7 @@ Run turbulence model transport equations.
 
 """
 function turbulence!(
-    rans::KOmegaModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, limit_gradient, config
+    rans::KOmegaModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, config
     ) where {T,F,M,Tu<:KOmega,E,D,BI,E1,E2}
 
     mesh = model.domain
