+++ src/Solvers/Solvers_2_CPISO.jl
@@ -2,7 +2,7 @@ export cpiso!
 
 """
     cpiso!(model, config; 
-        limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=0)
+        pref=nothing, ncorrectors=0, inner_loops=0)
 
 Compressible and transient variant of the PISO algorithm with a sensible enthalpy transport equation for the energy. 
 
@@ -10,7 +10,6 @@ Compressible and transient variant of the PISO algorithm with a sensible enthalp
 
 - `model` reference to a `Physics` model defined by the user.
 - `config` Configuration structure defined by the user with solvers, schemes, runtime and hardware structures configuration details.
-- `limit_gradient` flag use to activate gradient limiters in the solver (default = `false`)
 - `pref` Reference pressure value for cases that do not have a pressure defining BC. Incompressible solvers only (default = `nothing`)
 - `ncorrectors` number of non-orthogonality correction loops (default = `0`)
 - `inner_loops` number to inner loops used in transient solver based on PISO algorithm (default = `0`)
@@ -24,11 +23,10 @@ Compressible and transient variant of the PISO algorithm with a sensible enthalp
 """
 function cpiso!(
     model, config; 
-    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=2) 
+    pref=nothing, ncorrectors=0, inner_loops=2) 
 
     residuals = setup_unsteady_compressible_solvers(
         CPISO, model, config; 
-        limit_gradient=limit_gradient, 
         pref=pref,
         ncorrectors=ncorrectors, 
         inner_loops=inner_loops
@@ -40,7 +38,7 @@ end
 # Setup for all compressible algorithms
 function setup_unsteady_compressible_solvers(
     solver_variant, model, config; 
-    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=2
+    pref=nothing, ncorrectors=0, inner_loops=2
     ) 
 
     (; solvers, schemes, runtime, hardware) = config
@@ -128,7 +126,7 @@ function setup_unsteady_compressible_solvers(
     turbulenceModel = initialise(model.turbulence, model, mdotf, p_eqn, config)
 
     residuals  = solver_variant(
-        model, turbulenceModel, energyModel, ∇p, U_eqn, p_eqn, config; limit_gradient=limit_gradient, 
+        model, turbulenceModel, energyModel, ∇p, U_eqn, p_eqn, config;
         pref=pref, 
         ncorrectors=ncorrectors, 
         inner_loops=inner_loops)
@@ -138,7 +136,7 @@ end # end function
 
 function CPISO(
     model, turbulenceModel, energyModel, ∇p, U_eqn, p_eqn, config; 
-    limit_gradient=nothing, pref=nothing, ncorrectors=0, inner_loops=2)
+    pref=nothing, ncorrectors=0, inner_loops=2)
     
     # Extract model variables and configuration
     (; U, p) = model.momentum
@@ -215,7 +213,7 @@ function CPISO(
     @. rhof.values = Psif.values * pf.values
     flux!(mdotf, Uf, rhof, config)
 
-    limit_gradient!(limit_gradient, ∇p, p, config)
+    limit_gradient!(schemes.p.limiter, ∇p, p, config)
 
     update_nueff!(nueff, nu, model.turbulence, config)
     @. mueff.values = rhof.values*nueff.values
@@ -293,7 +291,7 @@ function CPISO(
 
             # Gradient
             grad!(∇p, pf, p, p.BCs, time, config) 
-            limit_gradient!(limit_gradient, ∇p, p, config)
+            limit_gradient!(schemes.p.limiter, ∇p, p, config)
 
             # non-orthogonal correction
[…]
