### 2025-01-24 4600dc5
update dev version to 0.4 - reflecting small breaking change to solver API
