+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format used for this `changelog` is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Notice that until the package reaches version `v1.0.0` minor releases are likely to be `breaking`. Starting from version `v0.3.1` breaking changes will be recorded here. 
 
-## Version [v0.3.4] - 2025-xx-xx
+## Version [v0.4.0] - 2025-xx-xx
 
 ### Added
 * Implementation of `Symmetry` boundary condition for `ScalarField` types
