+++ test/0_TEST_CASES/2d_compressible_KOmega_flatplate_fixedT.jl
@@ -81,80 +81,82 @@ model = Physics(
     Neumann(:top, 0.0)
 )
 
-schemes = (
-    U = set_schemes(divergence=Linear),
-    p = set_schemes(divergence=Linear),
-    h = set_schemes(divergence=Linear),
-    k = set_schemes(divergence=Upwind),
-    omega = set_schemes(divergence=Upwind)
-)
-
-solvers = (
-    U = set_solver(
-        model.momentum.U;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
-        convergence = 1e-7,
-        relax       = 0.7,
-    ),
-    p = set_solver(
-        model.momentum.p;
-        solver      = GmresSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
-        convergence = 1e-7,
-        relax       = 0.3,
-    ),
-    h = set_solver(
-        model.energy.h;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
-        convergence = 1e-7,
-        relax       = 0.7,
-        rtol = 1e-2,
-        atol = 1e-4
-    ),
-    k = set_solver(
-        model.turbulence.k;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), 
-        convergence = 1e-7,
-        relax       = 0.8,
-        atol = 1e-6
-    ),
-    omega = set_solver(
-        model.turbulence.omega;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
-        convergence = 1e-7,
-        relax       = 0.8,
-        atol = 1e-6
+for grad_limiter ∈ [nothing, FaceBased(model.domain), MFaceBased(model.domain)]
+    schemes = (
+        U = set_schemes(divergence=Linear, limiter=grad_limiter),
+        p = set_schemes(divergence=Linear, limiter=grad_limiter),
+        h = set_schemes(divergence=Linear),
+        k = set_schemes(divergence=Upwind),
+        omega = set_schemes(divergence=Upwind)
     )
-)
 
-runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
+    solvers = (
+        U = set_solver(
+            model.momentum.U;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(),
+            convergence = 1e-7,
+            relax       = 0.7,
+        ),
+        p = set_solver(
+            model.momentum.p;
+            solver      = GmresSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(),
+            convergence = 1e-7,
+            relax       = 0.3,
+        ),
+        h = set_solver(
+            model.energy.h;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(),
+            convergence = 1e-7,
+            relax       = 0.7,
+            rtol = 1e-2,
+            atol = 1e-4
+        ),
+        k = set_solver(
+            model.turbulence.k;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(), 
+            convergence = 1e-7,
+            relax       = 0.8,
+            atol = 1e-6
+        ),
+        omega = set_solver(
+            model.turbulence.omega;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
[…]
