+++ test/0_TEST_CASES/2d_incompressible_flatplate_KOmega_HighRe.jl
@@ -63,64 +63,66 @@ model = Physics(
     Neumann(:top, 0.0)
 )
 
-schemes = (
-    U = set_schemes(divergence=Upwind),
-    p = set_schemes(divergence=Upwind),
-    k = set_schemes(divergence=Upwind),
-    omega = set_schemes(divergence=Upwind)
-)
-
-
-solvers = (
-    U = set_solver(
-        model.momentum.U;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), 
-        convergence = 1e-7,
-        relax       = 0.7,
-    ),
-    p = set_solver(
-        model.momentum.p;
-        solver      = CgSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(),
-        convergence = 1e-7,
-        relax       = 0.3,
-    ),
-    k = set_solver(
-        model.turbulence.k;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), 
-        convergence = 1e-7,
-        relax       = 0.3,
-    ),
-    omega = set_solver(
-        model.turbulence.omega;
-        solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
-        preconditioner = Jacobi(), 
-        convergence = 1e-7,
-        relax       = 0.3,
+for grad_limiter ∈ [nothing, FaceBased(model.domain), MFaceBased(model.domain)]
+    schemes = (
+        U = set_schemes(divergence=Upwind, limiter=grad_limiter),
+        p = set_schemes(divergence=Upwind, limiter=grad_limiter),
+        k = set_schemes(divergence=Upwind),
+        omega = set_schemes(divergence=Upwind)
     )
-)
 
-runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
 
-hardware = set_hardware(backend=CPU(), workgroup=1024)
-# hardware = set_hardware(backend=CUDABackend(), workgroup=32)
-# hardware = set_hardware(backend=ROCBackend(), workgroup=32)
+    solvers = (
+        U = set_solver(
+            model.momentum.U;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(), 
+            convergence = 1e-7,
+            relax       = 0.7,
+        ),
+        p = set_solver(
+            model.momentum.p;
+            solver      = CgSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(),
+            convergence = 1e-7,
+            relax       = 0.3,
+        ),
+        k = set_solver(
+            model.turbulence.k;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(), 
+            convergence = 1e-7,
+            relax       = 0.3,
+        ),
+        omega = set_solver(
+            model.turbulence.omega;
+            solver      = BicgstabSolver, # BicgstabSolver, GmresSolver
+            preconditioner = Jacobi(), 
+            convergence = 1e-7,
+            relax       = 0.3,
+        )
+    )
 
-config = Configuration(
-    solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
+    runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
 
-GC.gc()
+    hardware = set_hardware(backend=CPU(), workgroup=1024)
+    # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
+    # hardware = set_hardware(backend=ROCBackend(), workgroup=32)
 
[…]
