### 2025-01-24 7c3edd3
update compressible tests
