+++ src/Calculate/Calculate_0_gradient_face_limiter.jl
@@ -15,7 +15,7 @@ end
 
 ### GRADIENT LIMITER - EXPERIMENTAL
 
-function limit_gradient!(method::FaceBased, ∇F, F::ScalarField, config)
+function limit_gradient!(method::FaceBased, ∇F, F::AbstractField, config)
     (; hardware) = config
     (; backend, workgroup) = hardware
 
@@ -48,7 +48,8 @@ end
     end
 end
 
-@kernel function _limit_gradient!(::FaceBased, limiter, ∇F, F, cells, faces, nbfaces)
+@kernel function _limit_gradient!(::FaceBased, limiter, ∇F, F::Field, cells, faces, nbfaces
+    ) where {Field<:AbstractScalarField}
     i = @index(Global)
     fID = i + nbfaces
 
@@ -76,10 +77,60 @@ end
     deltaF = (maxF - minF)
     minF -= deltaF
     maxF += deltaF
-    F1_ext = d1⋅grad1
-    F2_ext = d2⋅grad2
+
+    F1_ext = grad1⋅d1
+    F2_ext = grad2⋅d2
 
     set_limiter(limiter, owner1, maxF, minF, F1_ext, F1, F2)
+    set_limiter(limiter, owner2, maxF, minF, F2_ext, F2, F1)
+end
+
+@kernel function _limit_gradient!(::FaceBased, limiter, ∇F, F::Field, cells, faces, nbfaces
+    ) where {Field<:AbstractVectorField}
+    i = @index(Global)
+    fID = i + nbfaces
+
+    face = faces[fID]
+    ownerCells = face.ownerCells
+    owner1 = ownerCells[1]
+    owner2 = ownerCells[2]
+    cell1 = cells[owner1]
+    cell2 = cells[owner2]
+
+    n = face.normal
+    cf = face.centre 
+    c1 = cell1.centre
+    c2 = cell2.centre
+    d1 = (cf - c1)
+    d2 = (cf - c2)
+    grad1 = ∇F[owner1]
+    grad2 = ∇F[owner2]
+
+    gradf = grad1*d1
+    F1 = gradf⋅F[owner1]
+    F2 = gradf⋅F[owner2]
+
+    minF = min(F1, F2)
+    maxF = max(F1, F2)
+    deltaF = (maxF - minF)
+    minF -= deltaF
+    maxF += deltaF
+
+    F1_ext = gradf⋅gradf
+    set_limiter(limiter, owner1, maxF, minF, F1_ext, F1, F2)
+    
+    gradf = grad2*d2
+    F1 = gradf⋅F[owner1]
+    F2 = gradf⋅F[owner2]
+
+    minF = min(F1, F2)
+    maxF = max(F1, F2)
+    deltaF = (maxF - minF)
+    minF -= deltaF
+    maxF += deltaF
+    
+    F2_ext = gradf⋅gradf
+
     set_limiter(limiter, owner2, maxF, minF, F2_ext, F2, F1)
 end
 
