### 2025-01-24 496c288
modify tests to reduce their runtime (reduce tolerances and iteration count where possible also reduce mesh size)
