+++ test/0_TEST_CASES/2d_incompressible_flatplate_KOmega_lowRe.jl
@@ -77,7 +77,7 @@ solvers = (
         preconditioner = Jacobi(), 
         convergence = 1e-7,
         relax       = 0.7,
-        rtol = 1e-6
+        rtol = 1e-1
     ),
     p = set_solver(
         model.momentum.p;
@@ -85,7 +85,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.3,
-        rtol = 1e-6
+        rtol = 1e-2
     ),
     k = set_solver(
         model.turbulence.k;
@@ -93,7 +93,7 @@ solvers = (
         preconditioner = Jacobi(), 
         convergence = 1e-7,
         relax       = 0.3,
-        rtol = 1e-6
+        rtol = 1e-1
     ),
     omega = set_solver(
         model.turbulence.omega;
@@ -101,7 +101,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.3,
-        rtol = 1e-6
+        rtol = 1e-1
     )
 )
 
