+++ test/0_TEST_CASES/3d_incompressible_laminar_BFS.jl
@@ -3,7 +3,8 @@ using XCALibre
 # using AMDGPU # Run this if using AMD GPU
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
-grid = "bfs_unv_tet_10mm.unv"
+# grid = "bfs_unv_tet_10mm.unv"
+grid = "bfs_unv_tet_15mm.unv"
 
 mesh_file = joinpath(grids_dir, grid)
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
@@ -52,8 +53,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.8,
-        rtol = 1e-3,
-        atol = 1e-20
+        rtol = 1e-1,
     ),
     p = set_solver(
         model.momentum.p;
@@ -61,8 +61,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.2,
-        rtol = 1e-4,
-        atol = 1e-20
+        rtol = 1e-2,
     )
 )
 
