+++ test/0_TEST_CASES/2d_compressible_laminar_flatplate_fixedT.jl
@@ -65,6 +65,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.7,
+        rtol = 1e-1
     ),
     p = set_solver(
         model.momentum.p;
@@ -72,6 +73,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.3,
+        rtol = 1e-2
     ),
     h = set_solver(
         model.energy.h;
@@ -79,12 +81,11 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.7,
-        rtol = 1e-2,
-        atol = 1e-4
+        rtol = 1e-1
     )
 )
 
-runtime = set_runtime(iterations=500, write_interval=500, time_step=1)
+runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
 
 hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
