### 2025-01-24 947f4f4
add precompile workload for incompressible solvers - about 15% reduction in test time
