+++ src/XCALibre.jl
@@ -40,13 +40,202 @@ using Reexport
 @reexport using XCALibre.UNV3
 @reexport using XCALibre.UNV2
 
-# using PrecompileTools: @setup_workload, @compile_workload
+
+using PrecompileTools: @setup_workload, @compile_workload
 
 
-# @setup_workload begin
-#         @compile_workload begin
-#                 include("../CASE_UNV_BFS.jl")
-#         end
-# end
+@setup_workload begin
+    grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
+    grid = "backwardFacingStep_10mm.unv"
+    mesh_file = joinpath(grids_dir, grid)
+
+    # Inlet conditions
+    Umag = 0.5
+    velocity = [Umag, 0.0, 0.0]
+    nu = 1e-3
+    
+    k_inlet = 1
+    ω_inlet = 1000
+    ω_wall = ω_inlet
+
+    println(mesh_file)
+        @compile_workload begin
+
+            # Incompressible: Laminar and steady 
+
+            mesh = UNV2D_mesh(mesh_file, scale=0.001)
+            mesh_dev = mesh
+
+            model = Physics(
+                time = Steady(),
+                fluid = Fluid{Incompressible}(nu = nu),
+                turbulence = RANS{Laminar}(),
+                energy = Energy{Isothermal}(),
+                domain = mesh # mesh_dev  # use mesh_dev for GPU backend
+                )
+            
+            @assign! model momentum U (
+                Dirichlet(:inlet, velocity),
+                Neumann(:outlet, 0.0),
+                Wall(:wall, [0.0, 0.0, 0.0]),
+                Symmetry(:top)
+            )
+            
+            @assign! model momentum p (
+                Neumann(:inlet, 0.0),
+                Dirichlet(:outlet, 0.0),
+                Neumann(:wall, 0.0),
+                Symmetry(:top)
+            )
+            
+            schemes = (
+                U = set_schemes(divergence = Linear),
+                p = set_schemes()
+            )
+            
+            solvers = (
+                U = set_solver(
+                    model.momentum.U;
+                    solver = BicgstabSolver, # BicgstabSolver, GmresSolver
+                    smoother = JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
+                    preconditioner = Jacobi(),
+                    convergence = 1e-7,
+                    relax = 0.8,
+                    rtol = 1e-1,
+                ),
+                p = set_solver(
+                    model.momentum.p;
+                    solver = CgSolver, # BicgstabSolver, GmresSolver
+                    smoother = JacobiSmoother(domain=mesh_dev, loops=5, omega=2/3),
+                    preconditioner = Jacobi(),
+                    convergence = 1e-7,
+                    relax = 0.2,
+                    rtol = 1e-2,
+                )
+            )
+            
+            runtime = set_runtime(
+                iterations=10, time_step=1, write_interval=10)
+            
+            hardware = set_hardware(backend=CPU(), workgroup=1024)
+            
+            config = Configuration(
+                solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)
+            
+            initialise!(model.momentum.U, velocity) === nothing
+            initialise!(model.momentum.p, 0.0) === nothing
+            
+            residuals = run!(model, config)
+            
+            inlet = boundary_average(:inlet, model.momentum.U, config)
+            outlet = boundary_average(:outlet, model.momentum.U, config)
+
+            # Incompressible: Laminar and unsteady
+
+            model = Physics(
+            time = Transient(),
+            fluid = Fluid{Incompressible}(nu = nu),
+            turbulence = RANS{KOmega}(),
+            energy = Energy{Isothermal}(),
+            domain = mesh_dev
+            )
+
[…]
