### 2025-01-24 ef24a74
make variable inside loop local to suppress warning messages
