### 2025-01-24 163ab31
add precompile statements to reduce test runtimes via precompilation
