+++ test/runtests.jl
@@ -38,18 +38,18 @@ TEST_CASES_DIR = pkgdir(XCALibre, "test/0_TEST_CASES")
         end
     end
 
-    # @testset "Compressible" begin
-    #     test_files = [
-    #         "2d_compressible_KOmega_flatplate_fixedT.jl",
-    #         "2d_compressible_laminar_flatplate_fixedT.jl",
-    #         "2d_compressible_transient_laminar_heated_cylinder.jl"
-    #     ]
+    @testset "Compressible" begin
+        test_files = [
+            "2d_compressible_KOmega_flatplate_fixedT.jl",
+            "2d_compressible_laminar_flatplate_fixedT.jl",
+            "2d_compressible_transient_laminar_heated_cylinder.jl"
+        ]
 
-    #     for test ∈ test_files
-    #         test_path = joinpath(TEST_CASES_DIR, test)
-    #         include(test_path)
-    #     end
-    # end
+        for test ∈ test_files
+            test_path = joinpath(TEST_CASES_DIR, test)
+            include(test_path)
+        end
+    end
 
     foreach(rm, filter(endswith(".vtk"), readdir(pwd(), join=true)))
     foreach(rm, filter(endswith(".vtu"), readdir(pwd(), join=true)))
