+++ src/XCALibre.jl
@@ -40,202 +40,206 @@ using Reexport
 @reexport using XCALibre.UNV3
 @reexport using XCALibre.UNV2
 
+using StaticArrays, LinearAlgebra, SparseMatricesCSR, SparseArrays, LinearOperators
+using ProgressMeter, Printf, Adapt
 
-using PrecompileTools: @setup_workload, @compile_workload
+include("precompile.jl")
+
+# using PrecompileTools: @setup_workload, @compile_workload
 
 
-@setup_workload begin
-    grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
-    grid = "backwardFacingStep_10mm.unv"
-    mesh_file = joinpath(grids_dir, grid)
+# @setup_workload begin
+#     grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
+#     grid = "backwardFacingStep_10mm.unv"
+#     mesh_file = joinpath(grids_dir, grid)
 
-    # Inlet conditions
-    Umag = 0.5
-    velocity = [Umag, 0.0, 0.0]
-    nu = 1e-3
+#     # Inlet conditions
+#     Umag = 0.5
+#     velocity = [Umag, 0.0, 0.0]
+#     nu = 1e-3
     
-    k_inlet = 1
-    ω_inlet = 1000
-    ω_wall = ω_inlet
+#     k_inlet = 1
+#     ω_inlet = 1000
+#     ω_wall = ω_inlet
 
-    println(mesh_file)
-        @compile_workload begin
+#     println(mesh_file)
+#         @compile_workload begin
 
-            # Incompressible: Laminar and steady 
+#             # Incompressible: Laminar and steady 
 
-            mesh = UNV2D_mesh(mesh_file, scale=0.001)
-            mesh_dev = mesh
+#             mesh = UNV2D_mesh(mesh_file, scale=0.001)
+#             mesh_dev = mesh
 
-            model = Physics(
-                time = Steady(),
-                fluid = Fluid{Incompressible}(nu = nu),
-                turbulence = RANS{Laminar}(),
-                energy = Energy{Isothermal}(),
-                domain = mesh # mesh_dev  # use mesh_dev for GPU backend
-                )
+#             model = Physics(
+#                 time = Steady(),
+#                 fluid = Fluid{Incompressible}(nu = nu),
+#                 turbulence = RANS{Laminar}(),
+#                 energy = Energy{Isothermal}(),
+#                 domain = mesh # mesh_dev  # use mesh_dev for GPU backend
+#                 )
             
-            @assign! model momentum U (
-                Dirichlet(:inlet, velocity),
-                Neumann(:outlet, 0.0),
-                Wall(:wall, [0.0, 0.0, 0.0]),
-                Symmetry(:top)
-            )
+#             @assign! model momentum U (
+#                 Dirichlet(:inlet, velocity),
+#                 Neumann(:outlet, 0.0),
+#                 Wall(:wall, [0.0, 0.0, 0.0]),
+#                 Symmetry(:top)
+#             )
             
-            @assign! model momentum p (
-                Neumann(:inlet, 0.0),
-                Dirichlet(:outlet, 0.0),
-                Neumann(:wall, 0.0),
-                Symmetry(:top)
-            )
+#             @assign! model momentum p (
+#                 Neumann(:inlet, 0.0),
+#                 Dirichlet(:outlet, 0.0),
+#                 Neumann(:wall, 0.0),
+#                 Symmetry(:top)
+#             )
             
-            schemes = (
-                U = set_schemes(divergence = Linear),
-                p = set_schemes()
-            )
+#             schemes = (
+#                 U = set_schemes(divergence = Linear),
+#                 p = set_schemes()
+#             )
             
[…]
