### 2025-01-24 dd3a519
Further reduce number of iterations that would allow tests passing
