+++ test/0_TEST_CASES/2d_compressible_KOmega_flatplate_fixedT.jl
@@ -81,7 +81,7 @@ model = Physics(
     Neumann(:top, 0.0)
 )
 
-for grad_limiter ∈ [nothing, FaceBased(model.domain), MFaceBased(model.domain)]
+for grad_limiter ∈ [nothing] #, FaceBased(model.domain), MFaceBased(model.domain)]
     local schemes = (
         U = set_schemes(divergence=Linear, limiter=grad_limiter),
         p = set_schemes(divergence=Linear, limiter=grad_limiter),
@@ -133,7 +133,7 @@ for grad_limiter ∈ [nothing, FaceBased(model.domain), MFaceBased(model.domain)
         )
     )
 
-    local runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
+    local runtime = set_runtime(iterations=100, write_interval=100, time_step=1)
 
     local hardware = set_hardware(backend=CPU(), workgroup=1024)
     # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
