+++ test/0_TEST_CASES/2d_compressible_transient_laminar_heated_cylinder.jl
@@ -97,7 +97,7 @@ schemes = (
     h = set_schemes(divergence=Upwind, gradient=Midpoint, time=Euler)
 )
 
-runtime = set_runtime(iterations=200, write_interval=200, time_step=0.01)
+runtime = set_runtime(iterations=100, write_interval=100, time_step=0.01)
 
 hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
