+++ test/0_TEST_CASES/2d_compressible_laminar_flatplate_fixedT.jl
@@ -85,7 +85,7 @@ solvers = (
     )
 )
 
-runtime = set_runtime(iterations=200, write_interval=200, time_step=1)
+runtime = set_runtime(iterations=100, write_interval=100, time_step=1)
 
 hardware = set_hardware(backend=CPU(), workgroup=1024)
 # hardware = set_hardware(backend=CUDABackend(), workgroup=32)
