+++ joss-paper/paper.bib
@@ -25,13 +25,12 @@
     url = {https://github.com/code-saturne/code_saturne}
 }
 
-@article{Gerris,
-    author = {Stephane Popinet},
-    title = {Gerris},
-    year = {2018},
-    publisher = {Institut Jean le Rond d'Alembert},
-    journal = {sourceforge repository},
-    url = {http://gerris.dalembert.upmc.fr/}
+@article{Basilisk,
+    author = {Stephane Popinet et al.},
+    title = {Basilisk},
+    year = {2024},
+    journal = {Basilisk},
+    url = {http://basilisk.fr/}
 }
 
 @article{rackauckas2017differentialequations,
