+++ joss-paper/paper.md
@@ -51,7 +51,7 @@ Understanding the behaviour of fluid flow, such as air over a wing, oil lubricat
 
 # Statement of need
 
-Given the importance of fluid flow simulation in engineering applications, it is not surprising that there is a wealth of CFD solvers available, both open-source and commercially available. Well established open-source codes include: OpenFOAM [@OpenFOAM], SU2 [@SU2], CODE_SATURN [@CODE_SATURN], Gerris [@Gerris], etc. It is a testament to the open-source philosophy, and their developers, that some of these codes offer almost feature parity with commercial codes. However, established open-source and commercial codes have large codebases and, for performance reasons, have been implemented in statically compiled languages which makes it difficult to adapt and incorporate recent trends in scientific computing, for example, GPU computing and interfacing with machine learning frameworks to support the development of new models [@Ellis1; @Ellis2]. As a result, the research community has been actively developing new CFD codes, which is evident within the Julia ecosystem. 
+Given the importance of fluid flow simulation in engineering applications, it is not surprising that there is a wealth of CFD solvers available, both open-source and commercially available. Well established open-source codes include: OpenFOAM [@OpenFOAM], SU2 [@SU2], CODE_SATURN [@CODE_SATURN], Basilisk [@Basilisk], etc. It is a testament to the open-source philosophy, and their developers, that some of these codes offer almost feature parity with commercial codes. However, established open-source and commercial codes have large codebases and, for performance reasons, have been implemented in statically compiled languages which makes it difficult to adapt and incorporate recent trends in scientific computing, for example, GPU computing and interfacing with machine learning frameworks to support the development of new models [@Ellis1; @Ellis2]. As a result, the research community has been actively developing new CFD codes, which is evident within the Julia ecosystem. 
 
 The Julia programming language offers a fresh approach to scientific computing, with the benefits of dynamism whilst retaining the performance of statically typed languages thanks to its just-in-time compilation approach (using LLVM compiler technology). Thus, Julia makes it easy to prototype and test new ideas whilst producing performant machine code. This simplicity-performance dualism has resulted in a remarkable growth in its ecosystem offering for scientific computing, which includes state-of-the-art packages for solving differential equations e.g. `DifferentialEquations.jl` [@rackauckas2017differentialequations] , building machine learning models such as `Flux.jl` [@innes:2018], `Knet.jl` [@KNET] and `Lux.jl` [@pal2023lux], optimisation frameworks e.g. `JUMP.jl` [@Lubin2023] , automatic differentiation, such as `Enzyme.jl` [@NEURIPS2020_9332c513] , etc. Likewise, excellent CFD packages have also been developed, e.g. `Oceananigans.jl` [@OceananigansJOSS], which provides tools for ocean modelling, `Trixi.jl` [@schlottkelakemper2021purely] which provides high-order solvers using the Discontinuous Garlekin method, and `Waterlilly.jl` [@WeymouthFont2024], which implements the immersed boundary method on structured grids using a staggered finite volume method. To complement and extend the CFD offering of the Julia ecosystem, `XCALibre.jl` provides a composable framework which enables the aforementioned strengths of the Julia programming language to be realised in an unstructured finite volume solver with XPU capabilities. The package is intended primarily for researchers and students, as well as engineers, who are interested in CFD applications using the built-in solvers. Additionally, `XCALibre.jl` offers developers a user-friendly and computationally efficient framework for prototyping cutting-edge CFD solvers or methodologies, which can be tested in complex geometries relevant to a wide range of engineering applications. 
 
