### 2025-01-25 #7441 @mberto79
> [@mberto79](https://github.com/mberto79) and coauthors, nice work, thanks for your contribution. A couple of questions:
>
> 1. The tests took quite a bit of time to run on my laptop (30 minutes or so). Are they all required, or would it be possible leave some of them in the examples (or reduce the mesh sizes in the grid files)?
> 2. Gerris has been rewritten a few years back, maybe reference its successor instead? ([Basilisk](basilisk.fr))

Hi @vlc1! Thank you for the comments and apologies for the delay getting back.

Regading 2: Good point. The paper has been updated.

Regarding 1: Reducing test time has been on our mind for a while (we have been eating up Github action time pretty fast). We managed to bring down the test time by about 50% by adding precompile statements, reducing iterations in solvers to a minimum that would allow tests to run/check correctly and, as you suggesting, reducing mesh sizes (this was done for 3 of the 5 grids used in the tests). This is despite adding a few more tests to cover new functionality (gradient limiters and improved gradient calculation overall). Changes are found in this PR https://github.com/mberto79/XCALibre.jl/pull/31

FYI - I have temporarily labelled the latest version as 0.4.0-dev since the new limiters introduced a small breaking change in the top-level API. I will merge this into the `joss-submission` branch for your convenience. You may need to sync up and resolve the new environment since a new dependancy has been added (`PrecompileTools.jl`).
