### 2025-01-25 b95c3cf
fix DOI entry
