+++ joss-paper/paper.bib
@@ -143,7 +143,7 @@
     pages = {150-163},
     year = {2018},
     issn = {0142-727X},
-    doi = {https://doi.org/10.1016/j.ijheatfluidflow.2017.12.008},
+    doi = {10.1016/j.ijheatfluidflow.2017.12.008},
     url = {https://www.sciencedirect.com/science/article/pii/S0142727X17306343},
     author = {H. Medina and A. Beechook and H. Fadhila and S. Aleksandrova and S. Benjamin},
     keywords = {Laminar kinetic energy, Boundary layer, Transition, OpenFOAM, Plate, Separation, Bubble}
