### 2025-01-25 dac01bc
remove all syncrhonise statements since kernels are executed sequentially on GPU
