+++ src/Discretise/boundary_conditions/symmetry_interpolation.jl
@@ -6,7 +6,7 @@ function adjust_boundary!(b_cpu, BC::Symmetry, phif::FaceScalarField, phi, bound
 
     kernel! = adjust_boundary_symmetry_scalar!(backend, workgroup)
     kernel!(BC, phif, phi, boundaries, boundary_cellsID, time, phif_values, phi_values, ndrange = kernel_range)
-    # KernelAbstractions.synchronize(backend)
+    # # KernelAbstractions.synchronize(backend)
 end
 
 @kernel function adjust_boundary_symmetry_scalar!(BC, phif, phi, boundaries, boundary_cellsID, time, phif_values, phi_values)
@@ -27,7 +27,7 @@ function adjust_boundary!(b_cpu, BC::Symmetry, psif::FaceVectorField, psi::Vecto
 
     kernel! = adjust_boundary_symmetry_vector!(backend, workgroup)
     kernel!(BC, psif, psi, boundaries, boundary_cellsID, time, x, y, z, ndrange = kernel_range)
-    # KernelAbstractions.synchronize(backend)
+    # # KernelAbstractions.synchronize(backend)
 end
 
 @kernel function adjust_boundary_symmetry_vector!(BC, psif, psi, boundaries, boundary_cellsID, time, x, y, z)
