### 2025-01-26 d5d2ee7
improved implementation of CUDA preconditioners (draft)
