+++ src/Solve/Preconditioners/preconditioners_2_functions.jl
@@ -44,15 +44,15 @@ end
 #     end
 # end
 
-function update_preconditioner!(P::Preconditioner{IC0GPU,M,PT,S}, mesh, config) where {M<:AbstractSparseArray,PT,S}
-    KP.update!(P.P, P.A)
-    nothing
-end
+# function update_preconditioner!(P::Preconditioner{IC0GPU,M,PT,S}, mesh, config) where {M<:AbstractSparseArray,PT,S}
+#     KP.update!(P.P, P.A)
+#     nothing
+# end
 
-function update_preconditioner!(P::Preconditioner{ILU0GPU,M,PT,S}, mesh, config) where {M<:AbstractSparseArray,PT,S}
-    KP.update!(P.P, P.A)
-    nothing
-end
+# function update_preconditioner!(P::Preconditioner{ILU0GPU,M,PT,S}, mesh, config) where {M<:AbstractSparseArray,PT,S}
+#     KP.update!(P.P, P.A)
+#     nothing
+# end
 
 function update_preconditioner!(P::Preconditioner{NormDiagonal,M,PT,S}, mesh, config) where {M<:AbstractSparseArray,PT,S}
     # backend = _get_backend(mesh)
