+++ src/Solve/Preconditioners/preconditioners_0_types.jl
@@ -23,7 +23,10 @@ Adapt.@adapt_structure Jacobi
 struct DILU <: LDIVPreconditioner end
 Adapt.@adapt_structure DILU
 
-struct IC0GPU <: LDIVPreconditioner end
+# struct IC0GPU <: LDIVPreconditioner end
+# Adapt.@adapt_structure IC0GPU
+
+struct IC0GPU <: MULPreconditioner end
 Adapt.@adapt_structure IC0GPU
 
 struct ILU0GPU <: LDIVPreconditioner end
@@ -52,15 +55,15 @@ Preconditioner{NormDiagonal}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     Preconditioner{NormDiagonal,typeof(A),typeof(P),typeof(S)}(A,P,S)
 end
 
-Preconditioner{IC0GPU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
-    backend = get_backend(A)
-    m, n = size(A)
-    m == n || throw("Matrix not square")
-    # S = _convert_array!(zeros(m), backend)
-    S = zero(I)
-    P = KP.kp_ic0(A)
-    Preconditioner{IC0GPU,typeof(A),typeof(P),typeof(S)}(A,P,S)
-end
+# Preconditioner{IC0GPU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
+#     backend = get_backend(A)
+#     m, n = size(A)
+#     m == n || throw("Matrix not square")
+#     # S = _convert_array!(zeros(m), backend)
+#     S = zero(I)
+#     P = KP.kp_ic0(A)
+#     Preconditioner{IC0GPU,typeof(A),typeof(P),typeof(S)}(A,P,S)
+# end
 
 Preconditioner{ILU0GPU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     backend = get_backend(A)
