### 2025-01-27 7d9e8cd
draft improved implementation of `ILU0GPU`
