+++ src/Solve/Preconditioners/preconditioners_0_types.jl
@@ -26,10 +26,13 @@ Adapt.@adapt_structure DILU
 # struct IC0GPU <: LDIVPreconditioner end
 # Adapt.@adapt_structure IC0GPU
 
+# struct ILU0GPU <: LDIVPreconditioner end
+# Adapt.@adapt_structure ILU0GPU
+
 struct IC0GPU <: MULPreconditioner end
 Adapt.@adapt_structure IC0GPU
 
-struct ILU0GPU <: LDIVPreconditioner end
+struct ILU0GPU <: MULPreconditioner end
 Adapt.@adapt_structure ILU0GPU
 
 struct Preconditioner{T,M,P,S}
@@ -65,15 +68,15 @@ end
 #     Preconditioner{IC0GPU,typeof(A),typeof(P),typeof(S)}(A,P,S)
 # end
 
-Preconditioner{ILU0GPU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
-    backend = get_backend(A)
-    m, n = size(A)
-    m == n || throw("Matrix not square")
-    # S = _convert_array!(zeros(m), backend)
-    S = zero(I)
-    P = KP.kp_ic0(A)
-    Preconditioner{ILU0GPU,typeof(A),typeof(P),typeof(S)}(A,P,S)
-end
+# Preconditioner{ILU0GPU}(A::AbstractSparseArray{F,I}) where {F,I} = begin
+#     backend = get_backend(A)
+#     m, n = size(A)
+#     m == n || throw("Matrix not square")
+#     # S = _convert_array!(zeros(m), backend)
+#     S = zero(I)
+#     P = KP.kp_ic0(A)
+#     Preconditioner{ILU0GPU,typeof(A),typeof(P),typeof(S)}(A,P,S)
+# end
 
 Preconditioner{Jacobi}(A::AbstractSparseArray{F,I}) where {F,I} = begin
     backend = get_backend(A)
