### 2025-01-27 778c85b
add mesh path for windows machine for testing
