### 2025-01-31 #7441 @simone-silvestri
Hello @mberto79 and @AnjaliSandip
I have gone once again thoroughly through the paper and the examples and I would say that all the checklist is respected and the review is complete. I am quite impressed with the user interface and the usability of the software. The code is well documented, very accessible and understandable, congratulations, I am sure this package will find a lot of users.

Maybe just a final comment, please test the code on Metal GPU before claiming usability on Apple GPUs since Metal does not accept double precision anywhere in the kernels so if there is some float64 that slipped the code might crash. (or add a comment in the readme file)
