### 2025-01-31 7c50249
update extensions using constants for easier maintainability
