+++ prototype/parallelKernelLaunch.jl
@@ -12,9 +12,9 @@ grid = "bfs_unv_tet_10mm.unv"
 mesh_file = joinpath(grids_dir, grid)
 
 # mesh_file = "/Users/hmedi/Desktop/BFS_GRIDS/bfs_unv_tet_4mm.unv"
-mesh_file = "/Users/hmedi/Desktop/BFS_GRIDS/bfs_unv_tet_5mm.unv"
+# mesh_file = "/Users/hmedi/Desktop/BFS_GRIDS/bfs_unv_tet_5mm.unv"
 
-# mesh_file = "/home/humberto/Desktop/BFS_GRIDS/bfs_unv_tet_5mm.unv"
+mesh_file = "/home/humberto/Desktop/BFS_GRIDS/bfs_unv_tet_5mm.unv"
 
 # mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
 # mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
@@ -66,7 +66,7 @@ solvers = (
         # preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.7,
-        rtol = 1e-3,
+        rtol = 1e-1,
     ),
     p = set_solver(
         model.momentum.p;
@@ -75,7 +75,7 @@ solvers = (
         # preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.3,
-        rtol = 1e-3,
+        rtol = 1e-2,
     )
 )
 
@@ -91,7 +91,7 @@ config = Configuration(
 
 GC.gc(true)
 
- initialise!(model.momentum.U, velocity) === nothing
- initialise!(model.momentum.p, 0.0) === nothing
+initialise!(model.momentum.U, velocity) === nothing
+initialise!(model.momentum.p, 0.0) === nothing
 
 residuals = run!(model, config)
