+++ Project.toml
@@ -29,12 +29,15 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 [weakdeps]
 AMDGPU = "21141c5a-9bdb-4563-92ae-f87d6854732e"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+oneAPI = "8f75cd03-7ff8-4ecb-9b8f-daf728133b1b"
 
 [extensions]
 XCALibre_AMDExt = "AMDGPU"
 XCALibre_CUDAExt = "CUDA"
+XCALibre_oneAPIExt = "oneAPI"
 
 [compat]
+oneAPI = "2"
 AMDGPU = "1"
 Accessors = "0.1.30"
 Adapt = "4.0"
