+++ Project.toml
@@ -37,7 +37,6 @@ XCALibre_CUDAExt = "CUDA"
 XCALibre_oneAPIExt = "oneAPI"
 
 [compat]
-oneAPI = "2"
 AMDGPU = "1"
 Accessors = "0.1.30"
 Adapt = "4.0"
@@ -51,6 +50,7 @@ Krylov = "0.9"
 KrylovPreconditioners = "0.3"
 LDLFactorizations = "0.10"
 LinearOperators = "2.8"
+oneAPI = "2"
 PrecompileTools = "1.2"
 ProgressMeter = "1.10"
 Reexport = "1.2"
