### 2025-02-01 ba8d50b
update documentation and readme
