### 2025-02-01 #7441 @mberto79
Hi @simone-silvestri

Thank you for your kind comments and the time devoted to reviewing our work.

Your final question was very helpful. It made us realised that we inadvertedly dropped support for INTEL hardware. This has now been reinstated.

Your comment also made us realise that the wording on the documentation was ambigous regarding support for Apple hardware. Essentially, it is the only backend not currently supported due to lack of support for sparse matrices in the upstream `Metal.jl` implementation. We have updated the readme to explicitly mention that Metal is not currently supported. We also updated the text in the user guide to remind users that `Float64` may not be supported (this should be most relevant to Intel hardward). This has been added to the landing page for the documentation (not the readme information to keep it light) here: https://mberto79.github.io/XCALibre.jl/dev/ (under "mulitple backend")

Changes are now up the main branch and I have merged into the branch with the paper. The revelant PR can be found here: https://github.com/mberto79/XCALibre.jl/pull/32 - we also improved the performance of preconditioners on NVIDIA and AMD gpus :-)

Thanks!
