### 2025-02-02 7b53961
specialise call to `update_preconditioner` for GPU backends
