### 2025-02-10 45e53d1
update OF buildings case
