### 2025-02-11 8d4c6ed
solver configuration uses kwdef struct instead of tuple
