### 2025-02-11 0ce3c41
simplify construction of preconditioners
