+++ src/Solve/Preconditioners/preconditioners_2_functions.jl
@@ -6,22 +6,22 @@ set_preconditioner(PT::T, eqn, BCs, config
 begin
     phi = get_phi(eqn)
     mesh = phi.mesh
-    TF = _get_float(mesh)
-    time = zero(TF) # assumes simulation starts at time = 0 (might need generalising)
+    # TF = _get_float(mesh)
+    # time = zero(TF) # assumes simulation starts at time = 0 (might need generalising)
 
-    if typeof(phi) <: AbstractVectorField
+    # if typeof(phi) <: AbstractVectorField
 
-        discretise!(
-            eqn, get_phi(eqn), config) # should this be float?
+    #     discretise!(
+    #         eqn, get_phi(eqn), config) # should this be float?
 
-        time = zero(TF)
-        apply_boundary_conditions!(eqn, phi.x.BCs, XDir(1), time, config)
+    #     time = zero(TF)
+    #     apply_boundary_conditions!(eqn, phi.x.BCs, XDir(1), time, config)
 
-    elseif typeof(phi) <: AbstractScalarField
+    # elseif typeof(phi) <: AbstractScalarField
 
-        discretise!(eqn, get_phi(eqn), config) # should this be float?
-        apply_boundary_conditions!(eqn, BCs, nothing, time, config)
-    end
+    #     discretise!(eqn, get_phi(eqn), config) # should this be float?
+    #     apply_boundary_conditions!(eqn, BCs, nothing, time, config)
+    # end
 
     P = Preconditioner{T}(eqn.equation.A)
     update_preconditioner!(P, mesh, config)
