### 2025-02-13 2bba9c2
calculate residuals within equation solution
