### 2025-02-13 5c64f50
update BFS case with new convergence tolerance
