### 2025-02-15 8d8aa1e
add previous residual definition
