### 2025-02-16 429c67b
complete functionality for showing residuals and convergence based stopping (needs to be cascaded to all solvers/models)
