+++ src/ModelPhysics/Turbulence/RANS_laminar.jl
@@ -10,7 +10,9 @@ struct Laminar <: AbstractRANSModel end
 Adapt.@adapt_structure Laminar
 
 # Model type definition (hold equation definitions and internal data)
-struct LaminarModel end 
+struct LaminarModel{S1}
+    state::S1
+end 
 Adapt.@adapt_structure LaminarModel
 
 # Model API constructor (pass user input as keyword arguments and process if needed)
@@ -47,7 +49,7 @@ Initialisation of turbulent transport equations.
 function initialise(
     turbulence::Laminar, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
     ) where {T,F,M,Tu,E,D,BI}
-    return LaminarModel()
+    return LaminarModel(TurbulenceState((), true))
 end
 
 # Model solver call (implementation)
