### 2025-02-16 5b3888f
consistent calculation of residual and stopping simulations - minor fixes to `LKE` model and `wall_dist`
