### 2025-02-16 56c70ed
add new residual framework to energymodels, define new `ModelState` type to hold information and update changelog
