### 2025-02-16 8ac8b71
add AcceleratedKernels to project.toml
