### 2025-02-16 b69f039
change BFS workgroup for CPU to 1024
