### 2025-02-16 dbec0c8
set 2mm mesh for testing of BFS case
