### 2025-02-17 504010b
change KOmega to use AK
