### 2025-02-17 #7441 @mberto79
> [@mberto79](https://github.com/mberto79) Can you fix the missing DOI? Is v0.3.3 the latest version? What is the DOI (for example, 10.5281/zenodo.xxxxxx)?

Hi @AnjaliSandip,

The code will be version v0.4.0. I am preparing the release now. I will also submit the code (joss-paper branch) to Zenodo. I will post the DOI when I have it.

I had checked the suggested DOI, but I found that it is **not** linked to Basilisk (the CFD solver), can you confirm it would be ok not to add?

Kind regards,
Humberto
