### 2025-02-17 #7441 @mberto79
Hi @AnjaliSandip

The DOI is 10.5281/zenodo.14884519

Also, I slightly changed the tittle from "XCALibre.jl: an XPU unstructured finite volume Computational Fluid Dynamics library" to "XCALibre.jl: A Julia XPU unstructured finite volume Computational Fluid Dynamics library". I hope this is ok, otherwise please let me know and I will change back.

Many thanks.

Kind regards,
Humberto
