### 2025-02-19 #7441 @AnjaliSandip
> Hi [@AnjaliSandip](https://github.com/AnjaliSandip)
>
> The DOI is 10.5281/zenodo.14884519
>
> Also, I slightly changed the tittle from "XCALibre.jl: an XPU unstructured finite volume Computational Fluid Dynamics library" to "XCALibre.jl: A Julia XPU unstructured finite volume Computational Fluid Dynamics library". I hope this is ok, otherwise please let me know and I will change back.
>
> Many thanks.
>
> Kind regards, Humberto

Thanks!  The title sounds good.
