### 2025-02-19 2709343
update Basilisk reference and include DOI
