+++ joss-paper/paper.bib
@@ -26,10 +26,11 @@
 }
 
 @article{Basilisk,
-    author = {Stephane Popinet et al.},
+    author = {Antoon van Hooft et al.},
     title = {Basilisk},
-    year = {2024},
-    journal = {Basilisk},
+    year = {2016},
+    publisher = {Zenodo},
+    doi  = {10.5281/zenodo.1203631}, 
     url = {http://basilisk.fr/}
 }
 
