### 2025-02-19 #7441 @mberto79
> > > [@mberto79](https://github.com/mberto79) Can you fix the missing DOI? Is v0.3.3 the latest version? What is the DOI (for example, 10.5281/zenodo.xxxxxx)?
> >
> >
> > Hi [@AnjaliSandip](https://github.com/AnjaliSandip),
> > The code will be version v0.4.0. I am preparing the release now. I will also submit the code (joss-paper branch) to Zenodo. I will post the DOI when I have it.
> > I had checked the suggested DOI, but I found that it is **not** linked to Basilisk (the CFD solver), can you confirm it would be ok not to add?
> > Kind regards, Humberto
>
> [@mberto79](https://github.com/mberto79) Is there DOI? If not can you find another suitable reference with the DOI?

Hi @AnjaliSandip

I had a dig around, and found out that Basilisk has been deposited in Zenodo so I updated the citation in the paper to match this resource.

Kind regards,
Humberto
