### 2025-02-24 bf9f4e2
BFS KOmega benchmarking case
