+++ examples/3D_BFS_KOmega.jl
@@ -1,0 +1,176 @@
+using Plots 
+# using ThreadPinning
+using XCALibre
+using CUDA
+
+# pinthreads(:cores)
+
+grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
+# grid = "bfs_unv_tet_4mm.unv"
+# grid = "bfs_unv_tet_5mm.unv"
+grid = "bfs_unv_tet_10mm.unv"
+mesh_file = joinpath(grids_dir, grid)
+
+mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm.unv"
+# mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_5mm_with_BL.unv"
+# mesh_file = "/home/humberto/foamCases/jCFD_benchmarks/3D_BFS/bfs_unv_tet_4mm.unv"
+# mesh_file = "bfs_unv_tet_5mm.unv"
+
+# mesh_file = "/Users/hmedi/Desktop/BFS_GRIDS/bfs_unv_tet_4mm.unv"
+# mesh_file = "/home/humberto/Desktop/BFS_GRIDS/bfs_unv_tet_5mm.unv"
+mesh = UNV3D_mesh(mesh_file, scale=0.001)
+
+# workgroup = cld(length(mesh.cells), Threads.nthreads())
+# backend = CPU(); activate_multithread(backend)
+# backend = CPU()
+# activate_multithread1()
+workgroup = 32
+backend = CUDABackend()
+
+mesh_dev = adapt(backend, mesh)
+
+# Inlet conditions
+nu = 1e-3
+velocity = [0.5, 0.0, 0.0]
+noSlip = [0.0, 0.0, 0.0]
+kInlet = velocity[1]*0.01
+omegaInlet = kInlet/(100*nu)
+Re = (0.2*velocity[1])/nu
+
+model = Physics(
+    time = Steady(),
+    fluid = Fluid{Incompressible}(nu = nu),
+    turbulence = RANS{KOmega}(),
+    energy = Energy{Isothermal}(),
+    domain = mesh_dev
+    )
+
+@assign! model momentum U ( 
+    Dirichlet(:inlet, velocity),
+    Neumann(:outlet, 0.0),
+    Wall(:wall, noSlip),
+    Neumann(:sides, 0.0),
+    Symmetry(:top)
+)
+
+@assign! model momentum p (
+    Neumann(:inlet, 0.0),
+    Dirichlet(:outlet, 0.0),
+    Neumann(:wall, 0.0),
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
+)
+
+@assign! model turbulence k (
+    Dirichlet(:inlet, kInlet),
+    Neumann(:outlet, 0.0),
+    KWallFunction(:wall),
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
+)
+
+@assign! model turbulence omega (
+    Dirichlet(:inlet, omegaInlet),
+    Neumann(:outlet, 0.0),
+    OmegaWallFunction(:wall),
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
+)
+
+@assign! model turbulence nut (
+    Neumann(:inlet, 0.0),
+    Neumann(:outlet, 0.0),
+    NutWallFunction(:wall),
+    Neumann(:sides, 0.0),
+    Neumann(:top, 0.0)
+)
+
+sloops = 10; omega = 2/3
+rtol = 1e-1
+prtol = 1e-2
+solvers = (
+    U = set_solver(
+        model.momentum.U;
[…]
