### 2025-02-24 #7441 @mberto79
Hi @AnjaliSandip , is there anything I can do on my end to help with this latex compilation error? The paper compiles ok with the generic template. I can't see the logs so it's hard to debug. The only clue is the "capacity exceeded..." I have seen this when a visual asset is large, especially when processed using pgfplots. I checked the size of the images in the paper and the largest ones are only ~ 140 Kb, which should be ok. I don't have any other ideas at the minute, maybe if you could share the logs? (assuming you have access?).
