### 2025-02-25 3e124ab
update name to XCALibre on archive branch for periodic implicit implementation
