+++ src/Solve/Solve.jl
@@ -9,10 +9,10 @@ using KernelAbstractions
 using Atomix
 # using ILUZero
 
-using FVM_1D.Mesh
-using FVM_1D.Fields
-using FVM_1D.ModelFramework
-using FVM_1D.Discretise
+using XCALibre.Mesh
+using XCALibre.Fields
+using XCALibre.ModelFramework
+using XCALibre.Discretise
 
 using Reexport
 
