+++ src/ModelFramework/ModelFramework.jl
@@ -9,8 +9,8 @@ using Krylov
 using KernelAbstractions
 # using CUDA, AMDGPU
 
-using FVM_1D.Mesh
-using FVM_1D.Fields
+using XCALibre.Mesh
+using XCALibre.Fields
 
 include("ModelFramework_0_types.jl")
 include("ModelFramework_1_operations.jl")
