+++ CASE_3D_BFS_periodic.jl
@@ -1,4 +1,4 @@
-using FVM_1D
+using XCALibre
 
 mesh_file = "unv_sample_meshes/BFS_UNV_3D_hex_5mm.unv"
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
