+++ src/Solvers/Solvers.jl
@@ -14,16 +14,16 @@ using KernelAbstractions
 using Atomix
 using Adapt
 
-using FVM_1D.Mesh
-using FVM_1D.Fields
-using FVM_1D.ModelFramework
-using FVM_1D.Discretise
-using FVM_1D.Solve
-using FVM_1D.Calculate
-using FVM_1D.ModelPhysics
-using FVM_1D.Turbulence
-using FVM_1D.Energy
-using FVM_1D.VTK
+using XCALibre.Mesh
+using XCALibre.Fields
+using XCALibre.ModelFramework
+using XCALibre.Discretise
+using XCALibre.Solve
+using XCALibre.Calculate
+using XCALibre.ModelPhysics
+using XCALibre.Turbulence
+using XCALibre.Energy
+using XCALibre.VTK
 
 include("Solvers_0_functions.jl")
 include("Solvers_1_SIMPLE.jl")
