### 2025-02-25 #7441 @kyleniemeyer
@mberto79 can you please delete the output `paper.pdf` from the paper's source directory? That usually conflicts with our build process and could be the source of this error.
