### 2025-02-25 dbec5d2
update case to run on cpu
