### 2025-02-25 b729f75
switch CI for paper generation off
