+++ .github/workflows/draft-paper.yml
@@ -1,34 +1,34 @@
-name: Draft PDF
-on:
-    push:
-      paths:
-        - joss-paper/**
-        - .github/workflows/draft-paper.yml
+# name: Draft PDF
+# on:
+#     push:
+#       paths:
+#         - joss-paper/**
+#         - .github/workflows/draft-paper.yml
 
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: joss-paper/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v4
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: joss-paper/paper.pdf
-      - name: Commit PDF to repository
-        uses: EndBug/add-and-commit@v9
-        with:
-            message: '(auto) Paper PDF Draft'
-            # This should be the path to the paper within your repo.
-            add: 'joss-paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
+# jobs:
+#   paper:
+#     runs-on: ubuntu-latest
+#     name: Paper Draft
+#     steps:
+#       - name: Checkout
+#         uses: actions/checkout@v4
+#       - name: Build draft PDF
+#         uses: openjournals/openjournals-draft-action@master
+#         with:
+#           journal: joss
+#           # This should be the path to the paper within your repo.
+#           paper-path: joss-paper/paper.md
+#       - name: Upload
+#         uses: actions/upload-artifact@v4
+#         with:
+#           name: paper
+#           # This is the output path where Pandoc will write the compiled
+#           # PDF. Note, this should be the same directory as the input
+#           # paper.md
+#           path: joss-paper/paper.pdf
+#       - name: Commit PDF to repository
+#         uses: EndBug/add-and-commit@v9
+#         with:
+#             message: '(auto) Paper PDF Draft'
+#             # This should be the path to the paper within your repo.
+#             add: 'joss-paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
