### 2025-02-26 d539f79
Removed unnecesary overconstraint on turbulence type. Turbulence already defined by model passed.
