+++ src/ModelPhysics/Turbulence/RANS_kOmega.jl
@@ -147,7 +147,7 @@ Run turbulence model transport equations.
 """
 function turbulence!(
     rans::KOmegaModel{E1,E2,S1}, model::Physics{T,F,M,Tu,E,D,BI}, S, prev, time, config
-    ) where {T,F,M,Tu<:KOmega,E,D,BI,E1,E2,S1}
+    ) where {T,F,M,Tu,E,D,BI,E1,E2,S1}
 
     mesh = model.domain
     
