### 2025-02-26 702639d
added notes about this PR to the changelog
