### 2025-02-27 e4ac15e
play with coeff defs
