### 2025-02-28 f533e73
good step forward - do not accumulate when adding neighbour coeffs during boundary update
