+++ CASE_3D_cascade_periodic.jl
@@ -79,7 +79,7 @@ solvers = (
 )
 
 runtime = set_runtime(
-    iterations=200, time_step=1, write_interval=100)
+    iterations=500, time_step=1, write_interval=100)
 
 hardware = set_hardware(backend=backend, workgroup=workgroup)
 
