### 2025-02-28 3d92991
update `H!` function to work on matrix coefficients direction - to  include periodic contributions (`faces_range` does not include boundary faces)
