+++ src/Solvers/Solvers_0_functions.jl
@@ -268,23 +268,39 @@ end
     sumz = zero(TF)
 
     @inbounds begin
-        (; faces_range) = cells[i]
+        # (; faces_range) = cells[i]
 
-        for ni ∈ faces_range
-            nID = cell_neighbours[ni]
-            zIndex = spindex(colptr, rowval, i, nID)
-            val = nzval[zIndex]
+        # for ni ∈ faces_range
+        #     nID = cell_neighbours[ni]
+        #     zIndex = spindex(colptr, rowval, i, nID)
+        #     val = nzval[zIndex]
+        #     sumx += val * Ux[nID]
+        #     sumy += val * Uy[nID]
+        #     sumz += val * Uz[nID]
+        # end
+
+        start_index = rowptr[i]
+        end_index = rowptr[i+1] - 1
+        for nzi ∈ start_index:end_index
+            nID = colval[nzi]
+            val = nzval[nzi]
             sumx += val * Ux[nID]
             sumy += val * Uy[nID]
             sumz += val * Uz[nID]
         end
 
-        DIndex = spindex(colptr, rowval, i, i)
+        DIndex = spindex(rowptr, colval, i, i)
+
+        # remove diagonal contribution
         D = nzval[DIndex]
+        sumx -= D*Ux[i]
+        sumy -= D*Uy[i]
+        sumz -= D*Uz[i]
+
         rD = 1/D
-        Hx[i] = (bx[i] - sumx) * rD
-        Hy[i] = (by[i] - sumy) * rD
-        Hz[i] = (bz[i] - sumz) * rD
+        Hx[i] = (bx[i] - sumx)*rD
+        Hy[i] = (by[i] - sumy)*rD
+        Hz[i] = (bz[i] - sumz)*rD
     end
 end
 
