+++ src/ModelFramework/ModelFramework_0_types.jl
@@ -115,11 +115,14 @@ struct ScalarEquation{VTf<:AbstractVector, ASA<:AbstractSparseArray, OP} <: Abst
     Fx::VTf
 end
 Adapt.@adapt_structure ScalarEquation
-ScalarEquation(mesh::AbstractMesh) = begin
+ScalarEquation(mesh::AbstractMesh, periodicConnectivity) = begin
     nCells = length(mesh.cells)
     Tf = _get_float(mesh)
     mesh_temp = adapt(CPU(), mesh) # WARNING: Temp solution 
     i, j, v = sparse_matrix_connectivity(mesh_temp) # This needs to be a kernel
+    i = [i; periodicConnectivity.i]
+    j = [j; periodicConnectivity.j]
+    v = zeros(Tf, length(j))
     backend = _get_backend(mesh)
     # A = _convert_array!(sparse(i, j, v), backend)
     A = _build_A(backend, i, j, v, nCells)
@@ -147,11 +150,14 @@ struct VectorEquation{VTf<:AbstractVector, ASA<:AbstractSparseArray, OP} <: Abst
     Fx::VTf
 end
 Adapt.@adapt_structure VectorEquation
-VectorEquation(mesh::AbstractMesh) = begin
+VectorEquation(mesh::AbstractMesh, periodicConnectivity) = begin
     nCells = length(mesh.cells)
     Tf = _get_float(mesh)
     mesh_temp = adapt(CPU(), mesh) # WARNING: Temp solution 
     i, j, v = sparse_matrix_connectivity(mesh_temp) # This needs to be a kernel
+    i = [i; periodicConnectivity.i]
+    j = [j; periodicConnectivity.j]
+    v = zeros(Tf, length(j))
     backend = _get_backend(mesh)
     # A = _convert_array!(sparse(i, j, v), backend) 
     # A0 = _convert_array!(sparse(i, j, v), backend)
