### 2025-02-28 a11f8f6
extend argument of boundary definition macro/function - note a macro for generating a call for all operators was temporarily added - permanent solution needed
