+++ src/Discretise/Discretise_3_boundary_conditions.jl
@@ -1,10 +1,14 @@
 
 # TRANSIENT TERM 
-@inline (bc::AbstractBoundary)( # Used for all schemes (using "T")
-    term::Operator{F,P,I,Time{T}}, cellID, zcellID, cell, face, fID, i, component, time
-    ) where {F,P,I,T} = begin
-    # nothing
-    0.0, 0.0 # need to add consistent return types
+# @inline (bc::AbstractBoundary)( # Used for all schemes (using "T")
+#     term::Operator{F,P,I,Time{T}}, cellID, zcellID, cell, face, fID, i, component, time
+#     ) where {F,P,I,T} = begin
+#     # nothing
+#     0.0, 0.0 # need to add consistent return types
+# end
+
+@define_boundary Time{All} begin
+    0.0, 0.0
 end
 
 # KWallFunction
