+++ examples/3D_cascade_periodic.jl
@@ -51,7 +51,7 @@ model = Physics(
 
 schemes = (
     # U = set_schemes(divergence=Upwind, gradient=Midpoint),
-    U = set_schemes(divergence=Linear, gradient=Midpoint),
+    U = set_schemes(divergence=Linear, gradient=Orthogonal),
     p = set_schemes(gradient=Midpoint)
     # p = set_schemes()
 )
