### 2025-03-02 909c93b
comment out changes to work without periodic BC helper functions - needs to generalise interface
