+++ examples/3D_cascade_periodic.jl
@@ -3,8 +3,8 @@ using XCALibre
 # using CUDA
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
-# grid = "cascade_3D_periodic_2p5mm.unv"
-grid = "cascade_3D_periodic_4mm.unv"
+grid = "cascade_3D_periodic_2p5mm.unv"
+# grid = "cascade_3D_periodic_4mm.unv"
 
 mesh_file = joinpath(grids_dir, grid)
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
@@ -50,8 +50,8 @@ model = Physics(
 )
 
 schemes = (
-    # U = set_schemes(divergence=Upwind, gradient=Midpoint),
-    U = set_schemes(divergence=Linear, gradient=Orthogonal),
+    U = set_schemes(divergence=Linear, gradient=Midpoint),
+    # U = set_schemes(divergence=Linear, gradient=Orthogonal),
     p = set_schemes(gradient=Midpoint)
     # p = set_schemes()
 )
@@ -64,8 +64,7 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.8,
-        rtol = 1e-1,
-        atol = 1e-10
+        rtol = 1e-1
     ),
     p = set_solver(
         model.momentum.p;
@@ -73,13 +72,12 @@ solvers = (
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.2,
-        rtol = 1e-2,
-        atol = 1e-10
+        rtol = 1e-2
     )
 )
 
 runtime = set_runtime(
-    iterations=500, time_step=1, write_interval=100)
+    iterations=1000, time_step=1, write_interval=100)
 
 hardware = set_hardware(backend=backend, workgroup=workgroup)
 
@@ -94,27 +92,9 @@ initialise!(model.momentum.p, 0.0)
 
 residuals = run!(model, config)
 
-# plot(; xlims=(0,1000))
-# plot!(1:length(Rx), Rx, yscale=:log10, label="Ux")
-# plot!(1:length(Ry), Ry, yscale=:log10, label="Uy")
-# plot!(1:length(Rp), Rp, yscale=:log10, label="p")
-
-# # # PROFILING CODE
-
-# using Profile, PProf
-
-# GC.gc()
-# initialise!(model.momentum.U, velocity)
-# initialise!(model.momentum.p, 0.0)
-
-# Profile.Allocs.clear()
-# Profile.Allocs.@profile sample_rate=1 begin 
-#     residuals = run!(model, config)
-# end
-
-# PProf.Allocs.pprof()
-
-# test(::Nothing, a) = print("nothing")
-# test(b, a) = print(a*a)
-
-# test(nothing, 1)
+using Plots
+fig = plot(; xlims=(0,runtime.iterations), ylims=(1e-10, 1e-4))
+plot!(fig, 1:runtime.iterations, residuals.Ux, yscale=:log10, label="Ux")
+plot!(fig, 1:runtime.iterations, residuals.Uy, yscale=:log10, label="Uy")
+plot!(fig, 1:runtime.iterations, residuals.p, yscale=:log10, label="p")
+fig
