### 2025-03-02 6d66d73
draft implementation to integrate new semi-implicit periodic implementation
