+++ src/Discretise/boundary_conditions/periodic.jl
@@ -1,6 +1,9 @@
-export Periodic, PeriodicConnectivity
+export Periodic, PeriodicParent, PeriodicConnectivity
 export construct_periodic
 export adjust_boundary!
+import XCALibre.ModelFramework._extend_matrix
+
+abstract type AbstractPeriodic <: AbstractPhysicalConstraint end
 
 
 """
@@ -15,11 +18,17 @@ Periodic boundary condition model.
 - 'ID' -- Boundary ID
 - `value` -- tuple containing information needed to apply this boundary
 """
-struct Periodic{I,V} <: AbstractPhysicalConstraint
+struct Periodic{I,V} <: AbstractPeriodic
     ID::I
     value::V
 end
 Adapt.@adapt_structure Periodic
+
+struct PeriodicParent{I,V} <: AbstractPeriodic
+    ID::I
+    value::V
+end
+Adapt.@adapt_structure PeriodicParent
 
 struct PeriodicConnectivity{I}
     i::I
@@ -27,13 +36,13 @@ struct PeriodicConnectivity{I}
 end
 Adapt.@adapt_structure PeriodicConnectivity
 
-function fixedValue(BC::Periodic, ID::I, value::V) where {I<:Integer,V}
+function fixedValue(BC::T, ID::I, value::V) where {I<:Integer,V, T<:AbstractPeriodic}
     # Exception 1: Value is scalar
     if V <: Number
-        return Periodic{I,typeof(value)}(ID, value)
+        return T.name.wrapper(ID, value)
         # Exception 2: value is a tupple
     elseif V <: NamedTuple
-        return Periodic{I,V}(ID, value)
+        return T.name.wrapper(ID, value)
     # Error if value is not scalar or tuple
     else
         throw("The value provided should be a scalar or a tuple")
@@ -100,10 +109,10 @@ function construct_periodic(mesh, backend, patch1::Symbol, patch2::Symbol)
         faceAddress2[i] = boundaries[idx1].IDs_range[idx]
     end
 
-    values1 = (index=idx1, distance=distance, face_map=faceAddress1, ismaster=true)
-    values2 = (index=idx2, distance=distance, face_map=faceAddress2, ismaster=false)
+    values1 = (patchID=idx2, distance=distance, face_map=faceAddress1, isparent=true)
+    values2 = (patchID=idx1, distance=distance, face_map=faceAddress2, isparent=false)
 
-    p1 = Periodic(patch1, values1)
+    p1 = PeriodicParent(patch1, values1)
     p2 = Periodic(patch2, values2)
 
     periodic1 = adapt(backend, p1)
@@ -112,24 +121,21 @@ function construct_periodic(mesh, backend, patch1::Symbol, patch2::Symbol)
     return (periodic1, periodic2)
 end
 
-function periodic_matrix_connectivity(mesh, patch1, patch2)
-    (; faces, cells, boundaries, boundary_cellsID) = mesh
-    boundary_information = boundary_map(mesh)
-    idx1 = boundary_index(boundary_information, patch1.ID)
-    idx2 = boundary_index(boundary_information, patch2.ID)
+_extend_matrix(BC::PeriodicParent, mesh,  i, j) = begin
+    i_ext, j_ext = periodic_matrix_connectivity(BC, mesh)
+    return [i; i_ext], [j; j_ext]
+end
 
-    BC1 = boundaries[idx1].IDs_range
-    BC2 = boundaries[idx2].IDs_range
+function periodic_matrix_connectivity(BC::PeriodicParent, mesh)
+    (; faces, boundaries) = mesh
[…]
