+++ src/Discretise/Discretise_3_boundary_conditions.jl
@@ -1,19 +1,15 @@
 
 # TRANSIENT TERM 
-# @inline (bc::AbstractBoundary)( # Used for all schemes (using "T")
-#     term::Operator{F,P,I,Time{T}}, cellID, zcellID, cell, face, fID, i, component, time
-#     ) where {F,P,I,T} = begin
-#     # nothing
-#     0.0, 0.0 # need to add consistent return types
-# end
-
-@define_boundary Time{All} begin
-    0.0, 0.0
+@inline (bc::AbstractBoundary)( # Used for all schemes (using "T")
+    term::Operator{F,P,I,Time{T}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time
+    ) where {F,P,I,T} = begin
+    # nothing
+    0.0, 0.0 # need to add consistent return types
 end
 
 # KWallFunction
 @inline (bc::KWallFunction)(
-    term::Operator{F,P,I,Laplacian{T}}, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I,T}  = begin
+    term::Operator{F,P,I,Laplacian{T}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I,T}  = begin
     phi = term.phi 
     values = get_values(phi, component)
     J = term.flux[fID]
@@ -26,7 +22,7 @@ end
 
 # OmegaWallFunction
 @inline (bc::OmegaWallFunction)(
-    term::Operator{F,P,I,Laplacian{T}}, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I,T} = begin
+    term::Operator{F,P,I,Laplacian{T}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I,T} = begin
     phi = term.phi 
     values = get_values(phi, component)
     J = term.flux[fID]
@@ -39,7 +35,7 @@ end
 
 # KWallFunction 
 @inline (bc::KWallFunction)(
-    term::Operator{F,P,I,Divergence{Linear}}, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I} = begin
+    term::Operator{F,P,I,Divergence{Linear}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I} = begin
     flux = term.flux[fID]
     ap = term.sign*(flux) 
     ap, 0.0 # original
@@ -47,7 +43,7 @@ end
 
 # OmegaWallFunction 
 @inline (bc::OmegaWallFunction)(
-    term::Operator{F,P,I,Divergence{Linear}}, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I}  = begin
+    term::Operator{F,P,I,Divergence{Linear}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I}  = begin
     flux = term.flux[fID]
     ap = term.sign*(flux) 
     ap, 0.0 # original
@@ -55,21 +51,21 @@ end
 
 # KWallFunction
 @inline (bc::KWallFunction)(
-    term::Operator{F,P,I,Divergence{LUST}}, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I} = begin
+    term::Operator{F,P,I,Divergence{LUST}}, colval, rowptr, nzval, cellID, zcellID, cell, face, fID, i, component, time) where {F,P,I} = begin
     flux = term.flux[fID]
     ap = term.sign*(flux) 
     ap, 0.0 # original
 end
 
 @inline (bc::KWallFunction)(
[…]

[…]
