### 2025-03-04 22bc7e4
test with multiple periodic boundaries and add `inbounds` to BC macro definition
