+++ examples/3D_cascade_periodic.jl
@@ -3,15 +3,16 @@ using XCALibre
 # using CUDA
 
 grids_dir = pkgdir(XCALibre, "examples/0_GRIDS")
-# grid = "cascade_3D_periodic_2p5mm.unv"
-grid = "cascade_3D_periodic_4mm.unv"
+grid = "cascade_3D_periodic_2p5mm.unv"
+# grid = "cascade_3D_periodic_4mm.unv"
 
 mesh_file = joinpath(grids_dir, grid)
 mesh = UNV3D_mesh(mesh_file, scale=0.001)
 
 # backend = CUDABackend(); workgroup=32
 backend = CPU(); workgroup = cld(length(mesh.cells), Threads.nthreads())
-periodic = construct_periodic(mesh, backend, :top, :bottom)
+periodic1 = construct_periodic(mesh, backend, :top, :bottom)
+periodic2 = construct_periodic(mesh, backend, :side1, :side2)
 
 mesh_dev = adapt(backend, mesh)
 
@@ -35,24 +36,28 @@ model = Physics(
     Wall(:plate, [0.0, 0.0, 0.0]),
     # Symmetry(:side1, 0.0),
     # Symmetry(:side2, 0.0),
-    Neumann(:side1, 0.0),
-    Neumann(:side2, 0.0),
-    periodic...
+    # Neumann(:side1, 0.0),
+    # Neumann(:side2, 0.0),
+    periodic1...,
+    periodic2...
 )
 
 @assign! model momentum p (
     Neumann(:inlet, 0.0),
     Dirichlet(:outlet, 0.0),
     Neumann(:plate, 0.0),
-    Neumann(:side1, 0.0),
-    Neumann(:side2, 0.0),
-    periodic...
+    # Neumann(:side1, 0.0),
+    # Neumann(:side2, 0.0),
+    periodic1...,
+    periodic2...
 )
 
 schemes = (
-    U = set_schemes(divergence=Linear, gradient=Midpoint),
-    # U = set_schemes(divergence=Linear, gradient=Orthogonal),
-    p = set_schemes(gradient=Midpoint)
+    # U = set_schemes(divergence=Linear, gradient=Midpoint),
+    # U = set_schemes(divergence=Upwind, gradient=Midpoint),
+    U = set_schemes(divergence=Linear, gradient=Orthogonal),
+    # p = set_schemes(gradient=Midpoint)
+    p = set_schemes(gradient=Orthogonal)
     # p = set_schemes()
 )
 
@@ -68,7 +73,7 @@ solvers = (
     ),
     p = set_solver(
         model.momentum.p;
-        solver      = CgSolver, #GmresSolver, #CgSolver, # BicgstabSolver, GmresSolver
+        solver      = GmresSolver, #GmresSolver, #CgSolver, # BicgstabSolver, GmresSolver
         preconditioner = Jacobi(),
         convergence = 1e-7,
         relax       = 0.2,
@@ -90,7 +95,7 @@ initialise!(model.momentum.U, velocity)
 # initialise!(model.momentum.U, [0.0, 0.0, 0.0 ])
 initialise!(model.momentum.p, 0.0)
 
-residuals = run!(model, config)
+residuals = run!(model, config)#, ncorrectors=2)
 
 using Plots
 fig = plot(; xlims=(0,runtime.iterations), ylims=(1e-10, 1e-4))
