+++ src/UNV3/UNV3_2_builder.jl
@@ -629,7 +629,8 @@ calculate_face_properties!(mesh) = begin
         c1_c2 = cell2.centre - cell1.centre
         fc_c1 = cell1.centre - face.centre
         fc_c2 = cell2.centre - face.centre
-        delta = norm(c1_c2)
+        # delta = norm(c1_c2)
+        delta = norm(fc_c1) + norm(fc_c2)
         e = c1_c2/delta
         weight = norm(fc_c2)/norm(c1_c2)
         @reset face.delta = delta
