### 2025-03-06 c26d060
temporary fix for periodic implementation to work on GPU
