### 2025-03-06 7152ba2
update compressible solvers to use new equation constructor implemented for periodic BC support
