### 2025-03-06 0929cbb
remove precompile statements that can cause errors
