### 2025-03-07 fd5f134
Added copy of kOmega for prototyping
