+++ src/ModelPhysics/Turbulence/RANS_kOmega copy.jl
@@ -1,243 +1,0 @@
-export KOmega
-
-# Reference:
-# Wilcox, D. C., Turbulence Modeling for CFD, 2nd edition, DCW Industries, Inc., La Canada CA, 1998
-
-# Model type definition
-"""
-    KOmega <: AbstractTurbulenceModel
-
-kOmega model containing all kOmega field parameters.
-
-### Fields
-- 'k' -- Turbulent kinetic energy ScalarField.
-- 'omega' -- Specific dissipation rate ScalarField.
-- 'nut' -- Eddy viscosity ScalarField.
-- 'kf' -- Turbulent kinetic energy FaceScalarField.
-- 'omegaf' -- Specific dissipation rate FaceScalarField.
-- 'nutf' -- Eddy viscosity FaceScalarField.
-- 'coeffs' -- Model coefficients.
-
-"""
-struct KOmega{S1,S2,S3,F1,F2,F3,C} <: AbstractRANSModel
-    k::S1
-    omega::S2
-    nut::S3
-    kf::F1
-    omegaf::F2
-    nutf::F3
-    coeffs::C
-end
-Adapt.@adapt_structure KOmega
-
-struct KOmegaModel{E1,E2,S1}
-    k_eqn::E1 
-    ω_eqn::E2
-    state::S1
-end
-Adapt.@adapt_structure KOmegaModel
-
-# Model API constructor (pass user input as keyword arguments and process as needed)
-RANS{KOmega}(; β⁺=0.09, α1=0.52, β1=0.072, σk=0.5, σω=0.5) = begin 
-    coeffs = (β⁺=β⁺, α1=α1, β1=β1, σk=σk, σω=σω)
-    ARG = typeof(coeffs)
-    RANS{KOmega,ARG}(coeffs)
-end
-
-# Functor as constructor (internally called by Physics API): Returns fields and user data
-(rans::RANS{KOmega, ARG})(mesh) where ARG = begin
-    k = ScalarField(mesh)
-    omega = ScalarField(mesh)
-    nut = ScalarField(mesh)
-    kf = FaceScalarField(mesh)
-    omegaf = FaceScalarField(mesh)
-    nutf = FaceScalarField(mesh)
-    coeffs = rans.args
-    KOmega(k, omega, nut, kf, omegaf, nutf, coeffs)
-end
-
-# Model initialisation
-"""
-    initialise(turbulence::KOmega, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
-    ) where {T,F,M,Tu,E,D,BI}
-
-Initialisation of turbulent transport equations.
-
-### Input
-- `turbulence` -- turbulence model.
-- `model`  -- Physics model defined by user.
-- `mdtof`  -- Face mass flow.
-- `peqn`   -- Pressure equation.
-- `config` -- Configuration structure defined by user with solvers, schemes, runtime and 
-          hardware structures set.
-
-### Output
-- `KOmegaModel(k_eqn, ω_eqn)`  -- Turbulence model structure.
-
-"""
-function initialise(
-    turbulence::KOmega, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
-    ) where {T,F,M,Tu,E,D,BI}
-
-    (; k, omega, nut) = turbulence
-    (; rho) = model.fluid
-    (; solvers, schemes, runtime) = config
-    mesh = mdotf.mesh
-    eqn = peqn.equation
-
-    # define fluxes and sources
-    mueffk = FaceScalarField(mesh)
-    mueffω = FaceScalarField(mesh)
-    Dkf = ScalarField(mesh)
-    Dωf = ScalarField(mesh)
[…]
