### 2025-03-07 f78aae0
add files to implement `Slip` boundary and play with `Wall` definition
