+++ src/Discretise/boundary_conditions/slip_interpolation.jl
@@ -1,16 +1,15 @@
-function adjust_boundary!(b_cpu, BC::Wall, phif::FaceScalarField, phi, boundaries, boundary_cellsID, time,  backend, workgroup)
+function adjust_boundary!(b_cpu, BC::Slip, phif::FaceScalarField, phi, boundaries, boundary_cellsID, time,  backend, workgroup)
     phif_values = phif.values
     phi_values = phi.values
 
     kernel_range = length(b_cpu[BC.ID].IDs_range)
 
-    kernel! = adjust_boundary_neumann_scalar!(backend, workgroup)
+    kernel! = adjust_boundary_dirichlet_scalar!(backend, workgroup)
     kernel!(BC, phif, phi, boundaries, boundary_cellsID, time, phif_values, phi_values, ndrange = kernel_range)
     # # KernelAbstractions.synchronize(backend)
 end
 
-
-function adjust_boundary!(b_cpu, BC::Wall, psif::FaceVectorField, psi::VectorField, boundaries, boundary_cellsID, time, backend, workgroup)
+function adjust_boundary!(b_cpu, BC::Slip, psif::FaceVectorField, psi::VectorField, boundaries, boundary_cellsID, time, backend, workgroup)
     (; x, y, z) = psif
 
     kernel_range = length(b_cpu[BC.ID].IDs_range)
