### 2025-03-08 6f1333c
add `BlockMesher2D` module bringing old mesh generator code
