### 2025-03-08 17334a8
update changelog and update version to `v0.4.2-DEV`
