+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 The format used for this `changelog` is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). Notice that until the package reaches version `v1.0.0` minor releases are likely to be `breaking`. Starting from version `v0.3.1` breaking changes will be recorded here. 
 
+## Version [v0.4.2] - 2025-xx-xx
+
+### Added
+* A very simple 2D block mesh generator has been added (not ready for general use as it needs to be documented)[#41](@ref)
+
+### Fixed
+* No fixes have been added
+
+### Changed
+* No changes have been made
+
+### Breaking
+* No breaking changes
+
+### Deprecated
+* No functions deprecated
+
+### Removed
+* No functionality has been removed
+
 ## Version [v0.4.1] - 2025-03-06
 
 ### Added
