### 2025-03-11 777bfa1
Update paper.bib
Fixes some references in JOSS paper
