+++ joss-paper/paper.bib
@@ -1,35 +1,35 @@
-@article{OpenFOAM,
-    author = {OpenFOAM Foundation},
+@software{OpenFOAM,
+    author = {{OpenFOAM Foundation}},
     title = {Open Field Operation and Manipulation},
     year = {2024},
     publisher = {Github},
-    journal = {GitHub repository},
+    howpublished = {GitHub repository},
     url = {https://github.com/OpenFOAM}
 }
 
-@article{SU2,
-    author = {Standford University},
+@software{SU2,
+    author = {{Stanford University}},
     title = {SU2},
     year = {2024},
     publisher = {GitHub},
-    journal = {GitHub repository},
+    howpublished = {GitHub repository},
     url = {https://github.com/su2code/SU2}
 }
 
-@article{CODE_SATURN,
-    author = {EDF},
+@software{CODE_SATURN,
+    author = {{EDF}},
     title = {CODE_SATURN},
     year = {2024},
     publisher = {GitHub},
-    journal = {GitHub repository},
+    howpublished = {GitHub repository},
     url = {https://github.com/code-saturne/code_saturne}
 }
 
-@article{Basilisk,
+@software{Basilisk,
     author = {Antoon van Hooft et al.},
     title = {Basilisk},
     year = {2016},
-    publisher = {Zenodo},
+    howpublished = {Zenodo},
     doi  = {10.5281/zenodo.1203631}, 
     url = {http://basilisk.fr/}
 }
@@ -47,27 +47,26 @@
 
 @article{innes:2018,
   author    = {Mike Innes},
-  title     = {Flux: Elegant Machine Learning with Julia},
+  title     = {Flux: Elegant Machine Learning with {Julia}},
   journal   = {Journal of Open Source Software},
   year      = {2018},
   doi       = {10.21105/joss.00602},
 }
 
-@article{KNET,
-    author = {Deniz Yuret and collaborators},
+@software{KNET,
+    author = {Deniz Yuret and {collaborators}},
     title = {Knet},
     year = {2024},
-    publisher = {Giuthub},
-    journal = {Giuthub repository},
+    publisher = {GitHub},
+    journal = {GitHub repository},
     url = {hhttps://github.com/denizyuret/Knet.jl}
 }
 
 @software{pal2023lux,
   author    = {Pal, Avik},
-  title     = {Lux: Explicit Parameterization of Deep Neural Networks in Julia},
+  title     = {Lux: Explicit Parameterization of Deep Neural Networks in {Julia}},
   month     = apr,
   year      = 2023,
-  note      = {If you use this software, please cite it as below.},
   publisher = {Zenodo},
   version   = {v0.5.0},
   doi       = {10.5281/zenodo.7808904},
@@ -127,13 +126,14 @@
 }
 
 @misc{WeymouthFont2024,
-    title         = {WaterLily.jl: A differentiable and backend-agnostic Julia solver to simulate incompressible viscous flow and dynamic bodies},
+    title         = {WaterLily.jl: A differentiable and backend-agnostic {Julia} solver to simulate incompressible viscous flow and dynamic bodies},
     author        = {Gabriel D. Weymouth and Bernat Font},
     url           = {https://arxiv.org/abs/2407.16032},
     eprint        = {2407.16032},
     archivePrefix = {arXiv},
     year          = {2024},
-    primaryClass  = {physics.flu-dyn}
+    primaryClass  = {physics.flu-dyn},
+    doi = {10.48550/arXiv.2407.16032}
 }
 
 
@@ -160,7 +160,6 @@
   year = {1988},
   doi = {10.2514/3.10041},
[…]
