### 2025-03-11 966923d
Update paper.bib
fixes DifferentialEquations.jl reference
