+++ joss-paper/paper.bib
@@ -35,14 +35,15 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title={Differentialequations.jl--a performant and feature-rich ecosystem for solving differential equations in julia},
+  title={DifferentialEquations.jl -- a performant and feature-rich ecosystem for solving differential equations in Julia},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of Open Research Software},
   volume={5},
   number={1},
   pages={15},
   year={2017},
-  publisher={Ubiquity Press}
+  publisher={Ubiquity Press},
+  doi = {10.5334/jors.151}
 }
 
 @article{innes:2018,
