### 2025-03-11 #7441 @kyleniemeyer
Hello @mberto79, I made some fixes to the references in https://github.com/mberto79/XCALibre.jl/pull/42, if you could review and merge.

However, I think a few references to software repositories on GitHub should be replaced with article citations, where applicable. For example, rather than citing the GitHub repository for the SU2 code, I saw in the README that they recommended this citation:
> "SU2: An open-source suite for multiphysics simulation and design," AIAA Journal, 54(3):828-846, 2016. http://arc.aiaa.org/doi/10.2514/1.J053813

Please take a look at the other repository citations and see if there is an alternate citation requested. Since your paper is mainly citing those to acknowledge related work, rather than citing a specific software used (for reproducibility), I think the papers would be more appropriate, if they exist.
