### 2025-03-11 a05249d
make `Neumann` )zero gradient fully explicit for testing
