+++ src/Discretise/boundary_conditions/neumann.jl
@@ -44,8 +44,8 @@ end
     (; area, delta) = face 
     flux = -J*area/delta
     ap = term.sign*(flux)
-    ap, ap*values[cellID] # original
-    # 0.0, 0.0 # go for this!
+    # ap, ap*values[cellID] # original
+    0.0, 0.0 # try this
     # 0.0, -flux*delta*bc.value # draft implementation to test!
 end
 
@@ -58,7 +58,10 @@ end
 @define_boundary Neumann Divergence{Upwind} begin
     flux = term.flux[fID]
     ap = term.sign*(flux) 
-    ap, 0.0 # original
+    # ap, 0.0 # original
+    phi = term.phi 
+    values = get_values(phi, component)
+    0.0, -ap*values[cellID] # try this
 end
 
 @define_boundary Neumann Divergence{LUST} begin
