+++ joss-paper/paper.bib
@@ -1,28 +1,31 @@
-@software{OpenFOAM,
-    author = {{OpenFOAM Foundation}},
-    title = {Open Field Operation and Manipulation},
-    year = {2024},
-    publisher = {Github},
-    howpublished = {GitHub repository},
-    url = {https://github.com/OpenFOAM}
+@book{OpenFOAM,
+  title     = "OpenFOAM v12 User Guide",
+  author    = "Greenshields, Christopher",
+  year      = 2024,
+  url       = "https://doc.cfd.direct/openfoam/user-guide-v12",
+  publisher = "The OpenFOAM Foundation",
+  address   = "London, UK"
 }
 
-@software{SU2,
-    author = {{Stanford University}},
-    title = {SU2},
-    year = {2024},
-    publisher = {GitHub},
-    howpublished = {GitHub repository},
-    url = {https://github.com/su2code/SU2}
+@article{SU2,
+    author = {Economon, Thomas D. and Palacios, Francisco and Copeland, Sean R. and Lukaczyk, Trent W. and Alonso, Juan J.},
+    title = {{SU2}: an open-source suite for multiphysics simulation and design},
+    journal = {{AIAA} Journal},
+    volume = {54},
+    number = {3},
+    pages = {828--846},
+    year = {2016},
+    doi = {10.2514/1.J053813},
+    URL = {https://doi.org/10.2514/1.J053813}
 }
 
-@software{CODE_SATURN,
-    author = {{EDF}},
-    title = {CODE_SATURN},
-    year = {2024},
-    publisher = {GitHub},
-    howpublished = {GitHub repository},
-    url = {https://github.com/code-saturne/code_saturne}
+@article{CODE_SATURN,
+    author = {Fr{\'e}d{\'e}ric Archambeau and Namane M{\'e}chitoua and Marc Sakiz},
+    journal = {International Journal on Finite Volumes},
+    title = {Code\_{S}aturne: a finite volume code for the computation of turbulent incompressible flows},
+    url = {https://hal.archives-ouvertes.fr/hal-01115371/document},
+    volume = {1},
+    year = {2004}
 }
 
 @software{Basilisk,
@@ -54,13 +57,15 @@
   doi       = {10.21105/joss.00602},
 }
 
-@software{KNET,
-    author = {Deniz Yuret and {collaborators}},
-    title = {Knet},
-    year = {2024},
-    publisher = {GitHub},
-    journal = {GitHub repository},
-    url = {hhttps://github.com/denizyuret/Knet.jl}
+@inproceedings{KNET,
+  author =        {Deniz Yuret},
+  title =         {Knet: beginning deep learning with 100 lines of {Julia}},
+  year  =         {2016},
+  eventtitle =    {{NIPS} 2016},
+  year =          {2016},
+  isbn =          {9781510838819},
+  publisher =     {Curran Associates Inc.},
+  location =      {Barcelona, Spain}
 }
 
 @software{pal2023lux,
@@ -79,7 +84,7 @@
     title = {{JuMP} 1.0: {R}ecent improvements to a modeling language for mathematical optimization},
     journal = {Mathematical Programming Computation},
     volume = {15},
-    pages = {581–589},
+    pages = {581--589},
     year = {2023},
     doi = {10.1007/s12532-023-00239-3}
 }
@@ -142,7 +147,7 @@
     title = {A novel laminar kinetic energy model for the prediction of pretransitional velocity fluctuations and boundary layer transition},
     journal = {International Journal of Heat and Fluid Flow},
     volume = {69},
-    pages = {150-163},
+    pages = {150--163},
     year = {2018},
     issn = {0142-727X},
[…]
