### 2025-03-11 #7441 @mberto79
@kyleniemeyer thank you for the fixes to the references, I have merged these.

I also replaced the software/repository citations for the following references:

- OpenFOAM: I'm citing the user guide as indicated [here](https://doc.cfd.direct/openfoam/user-guide-v12/index)
- SU2: as per your comment. Thank you
- CODE_SATURM: as recommended in their  [FAQ](https://www.code-saturne.org/cms/web/documentation/faq)
- Knet: changed reference to a paper listed on the GitHub [repository](https://github.com/denizyuret/Knet.jl)

I didn't change the following software references:
- Basilisk: I wasn't able to find an paper presenting/introducting the software on their site.
- Lux: the software GitHub [repository](https://github.com/LuxDL/Lux.jl?tab=readme-ov-file#-citation) explicitly request the use of the existing reference

I generated the pdf and double checked that these references were reproduced correctly, I didn't spot any issues, hopefully I didn't miss any. For the last entry in the list of references (Yuret), I wan't able to find a `DOI` but found the `ISNB` for the proceedings. I hope this is ok.

Thanks!
