### 2025-03-12 e263fab
New file for _nut_access() and changes so it's included
