### 2025-03-12 7101764
Cleanup where the Menterf1 Dummy struct is stored to match the in progress HP/DESMenter branch
