### 2025-03-12 10e7daa
Revert "Cleanup where the Menterf1 Dummy struct is stored to match the in progress HP/DESMenter branch"
This reverts commit 7101764145886d906f141607a50b9f3342a14c8f.
