+++ src/ModelPhysics/Turbulence/DES_Menter.jl
@@ -1,4 +1,0 @@
-export MenterF1
-
-struct MenterF1 #This is a dummy struct to allow the νₜ access function to be written ready for DES implementation
-end
