### 2025-03-12 1137a1e
Cleanup where the Menterf1 Dummy struct is stored to match the in progress HP/DESMenter branch
