### 2025-03-12 19bb630
Update JOSS references
