+++ joss-paper/paper.bib
@@ -29,7 +29,7 @@
 }
 
 @software{Basilisk,
-    author = {Antoon van Hooft et al.},
+    author = {Antoon van Hooft and {others}},
     title = {Basilisk},
     year = {2016},
     howpublished = {Zenodo},
@@ -38,7 +38,7 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title={DifferentialEquations.jl -- a performant and feature-rich ecosystem for solving differential equations in Julia},
+  title={DifferentialEquations.jl -- a performant and feature-rich ecosystem for solving differential equations in {Julia}},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of Open Research Software},
   volume={5},
