### 2025-03-12 #7441 @kyleniemeyer
@mberto79 thanks for that - I did make a few final minor changes in https://github.com/mberto79/XCALibre.jl/pull/44

(Regarding the Basilisk submission, I also could not find any guidance on citing—but I corrected the way "et al." was showing up as part of the author's name.)
