### 2025-03-12 #7441 @mberto79
> Congratulations [@mberto79](https://github.com/mberto79) on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.
>
> Many thanks to [@vlc1](https://github.com/vlc1) and [@simone-silvestri](https://github.com/simone-silvestri) for reviewing this submission, and [@AnjaliSandip](https://github.com/AnjaliSandip) for editing it.

This is great news! Thank you @AnjaliSandip and @kyleniemeyer for editing and managing the submission. Thank you @vlc1 and @simone-silvestri for reviewing the paper/code.
